### 2021-09-01 #46 @hodgestar
Presubmission inquiry: QuTiP
Submitting Author: Simon Cross (@hodgestar)
Package Name: qutip
One-Line Description of Package: Simulation of open quantum systems.
Repository Link (if existing):   https://github.com/qutip/qutip
Website: https://qutip.org/

---

"QuTiP is open-source software for simulating the dynamics of open quantum systems." -- our official tagline. :)

QuTiP falls under the category of "quantum physics" or perhaps more broadly "physics". It's intended for use by young and old physicists alike.

The target audience is quite broad -- the package is accessible both to undergraduates learning quantum mechanics for the first time (for who it provides a sort of experimental lab in which to try and out examine aspects of quantum systems and to build intution), all the way to postgraduates and researchers (for who it provides a platform to use when doing real research).

There are many Python packages for quantum *computing* but few others than focus specifically on quantum *physics* and that support the simulation of *open* quantum systems.

Some of QuTiP processes are already somewhat mature (e.g. we have a repository, comprehensive CI runs, okay packaging, and fairly automated building and deploying to PyPI) but I'm sure that there is lots of room to do things better too and I imagine it would be very healthy for QuTiP to go through the pyOpenSci process.
