### 2021-09-01 #46 @lwasser
hey there @hodgestar thank you for this presubmission! We will have a look and will get back to you about fit for pyOpenSci.
