### 2021-09-03 4c2d055
Remove operator in data-layers
