+++ qutip/solve/countstat.py
@@ -205,7 +205,7 @@ def countstat_current_noise(L, c_ops, wlist=None, rhoss=None, J_ops=None,
 
             Pop = _data.kron(rhoss_vec.data, tr_op_vec.data.transpose())
             Iop = _data.identity(N*N)
-            Q = Iop - Pop
+            Q = _data.sub(Iop, Pop)
 
             for k, w in enumerate(wlist):
 
@@ -259,8 +259,8 @@ def countstat_current_noise(L, c_ops, wlist=None, rhoss=None, J_ops=None,
 
                         X_rho_vec_i = _data.dense.fast_from_numpy(X_rho_vec_i)
                         S[j, i, k] -= (
-                            _data.expect_super(Jj @ Q, X_rho_vec_i)
-                            + _data.expect_super(Ji @ Q, X_rho_vec_j)
+                            _data.expect_super(_data.matmul(Jj, Q), X_rho_vec_i)
+                            + _data.expect_super(_data.matmul(Jj, Q), X_rho_vec_j)
                         ).real
 
         else:
