+++ qutip/core/blochredfield.py
@@ -1,0 +1,235 @@
+# This file is part of QuTiP: Quantum Toolbox in Python.
+#
+#    Copyright (c) 2011 and later, QuSTaR.
+#    All rights reserved.
+#
+#    Redistribution and use in source and binary forms, with or without
+#    modification, are permitted provided that the following conditions are
+#    met:
+#
+#    1. Redistributions of source code must retain the above copyright notice,
+#       this list of conditions and the following disclaimer.
+#
+#    2. Redistributions in binary form must reproduce the above copyright
+#       notice, this list of conditions and the following disclaimer in the
+#       documentation and/or other materials provided with the distribution.
+#
+#    3. Neither the name of the QuTiP: Quantum Toolbox in Python nor the names
+#       of its contributors may be used to endorse or promote products derived
+#       from this software without specific prior written permission.
+#
+#    THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+#    "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+#    LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
+#    PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+#    HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+#    SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+#    LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+#    DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+#    THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+#    (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+#    OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+###############################################################################
+import os
+import numpy as np
+from qutip.settings import settings as qset
+
+from . import Qobj, QobjEvo, liouvillian, Cubic_Spline, coefficient, sprepost
+from ._brtools import SpectraCoefficient, _EigenBasisTransform
+from ._brtensor import _BlochRedfieldElement
+
+
+__all__ = ['bloch_redfield_tensor', 'brterm']
+
+
+def _read_legacy_a_ops(a_op):
+    constant = coefficient("1")
+    parsed_a_op
+    op, spec = a_op
+    if isinstance(op, Qobj):
+        if isinstance(spec, str):
+            parsed_a_op = (op, coefficient(spec, args={'w':0}))
+        elif callable(spec):
+            parsed_a_op =(op, coefficient(lambda t, args: spec(args['w'])))
+        elif isinstance(spec, tuple):
+            if isinstance(spec[0], str):
+                freq_responce = coefficient(spec[0], args={'w':0})
+            elif isinstance(spec[0], Cubic_Spline):
+                freq_responce = SpectraCoefficient(coefficient(spec[0]))
+            else:
+                raise Exception('Invalid bath-coupling specification.')
+            if isinstance(spec[1], str):
+                time_responce = coefficient(spec[1], args={'w':0})
+            elif isinstance(spec[1], Cubic_Spline):
+                time_responce = coefficient(spec[1])
+            else:
[…]
