### 2021-09-07 f1251c8
Merge pull request #1652 from Erotemic/patch-1
Add docs for num_cbits in QubitCircuit docstring.
