### 2021-09-08 f994be6
measurement operators can be integers, tests fixed
