### 2021-09-08 202248a
Merge pull request #1654 from qutip/dependabot/pip/doc/pillow-8.3.2
Bump pillow from 8.2.0 to 8.3.2 in /doc.
