+++ doc/requirements.txt
@@ -23,7 +23,7 @@ packaging==20.9
 parso==0.8.2
 pexpect==4.8.0
 pickleshare==0.7.5
-Pillow==8.2.0
+Pillow==8.3.2
 prompt-toolkit==3.0.18
 ptyprocess==0.7.0
 Pygments==2.8.1
