### 2021-09-08 01f989f
Add solver class
