+++ qutip/tests/solver/test_sesolve.py
@@ -1,0 +1,238 @@
+# This file is part of QuTiP: Quantum Toolbox in Python.
+#
+#    Copyright (c) 2011 and later, Paul D. Nation and Robert J. Johansson.
+#    All rights reserved.
+#
+#    Redistribution and use in source and binary forms, with or without
+#    modification, are permitted provided that the following conditions are
+#    met:
+#
+#    1. Redistributions of source code must retain the above copyright notice,
+#       this list of conditions and the following disclaimer.
+#
+#    2. Redistributions in binary form must reproduce the above copyright
+#       notice, this list of conditions and the following disclaimer in the
+#       documentation and/or other materials provided with the distribution.
+#
+#    3. Neither the name of the QuTiP: Quantum Toolbox in Python nor the names
+#       of its contributors may be used to endorse or promote products derived
+#       from this software without specific prior written permission.
+#
+#    THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+#    "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+#    LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
+#    PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+#    HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+#    SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+#    LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+#    DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+#    THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+#    (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+#    OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+###############################################################################
+
+import pytest
+import pickle
+import numpy as np
+from numpy.testing import assert_, run_module_suite
+from qutip import (
+    sigmax, sigmay, sigmaz, qeye, basis, expect, num, destroy, create,
+    Cubic_Spline, QobjEvo, Qobj
+)
+from qutip.solver import SolverOptions, sesolve, SeSolver
+from qutip.solver.integrator import *
+
+all_ode_method = integrator_collection.list_keys('methods', time_dependent=True)
+
+
+class TestSeSolve():
+    H0 = 0.2 * np.pi * sigmaz()
+    H1 = np.pi * sigmax()
+    tlist = np.linspace(0, 20, 200)
+    S = Cubic_Spline(0, 20, np.exp(-0.5 * tlist))
+    args = {'alpha': 0.5}
+    w_a = 0.35
+    a = 0.5
+
+    @pytest.mark.parametrize(['unitary_op'],
+        [pytest.param(None, id="state"),
+         pytest.param(qeye(2), id="unitary"),
+    ])
+    @pytest.mark.parametrize(['H', 'analytical'],
+        [pytest.param(H1,
+                      lambda t, args: t,
+                      id='const_H'),
+         pytest.param(lambda t, args: np.pi * sigmax() * np.exp(-args['alpha'] * t),
+                      lambda t, args: ((1 - np.exp(-args['alpha'] * t))
+                                       / args['alpha']),
[…]
