+++ qutip/tests/solver/test_mesolve.py
@@ -1,0 +1,653 @@
+# This file is part of QuTiP: Quantum Toolbox in Python.
+#
+#    Copyright (c) 2011 and later, Paul D. Nation and Robert J. Johansson.
+#    All rights reserved.
+#
+#    Redistribution and use in source and binary forms, with or without
+#    modification, are permitted provided that the following conditions are
+#    met:
+#
+#    1. Redistributions of source code must retain the above copyright notice,
+#       this list of conditions and the following disclaimer.
+#
+#    2. Redistributions in binary form must reproduce the above copyright
+#       notice, this list of conditions and the following disclaimer in the
+#       documentation and/or other materials provided with the distribution.
+#
+#    3. Neither the name of the QuTiP: Quantum Toolbox in Python nor the names
+#       of its contributors may be used to endorse or promote products derived
+#       from this software without specific prior written permission.
+#
+#    THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+#    "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+#    LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
+#    PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+#    HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+#    SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+#    LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+#    DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+#    THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+#    (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+#    OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+###############################################################################
+
+import numpy as np
+from numpy.testing import assert_allclose
+from types import FunctionType
+
+from qutip import *
+from qutip.solver import *
+from qutip.random_objects import rand_ket
+import pickle
+import pytest
+all_ode_method = integrator_collection.list_keys('methods', time_dependent=True)
+
+
+def fidelitycheck(out1, out2, rho0vec):
+    fid = np.zeros(len(out1.states))
+    for i, E in enumerate(out2.states):
+        rhot = vector_to_operator(E*rho0vec)
+        fid[i] = fidelity(out1.states[i], rhot)
+    return fid
+
+
+class TestMESolveDecay:
+    N = 10
+    a = destroy(N)
+    kappa = 0.2
+    tlist = np.linspace(0, 10, 201)
+    ada = a.dag() * a
+
+    @pytest.fixture(params=[
+        pytest.param([ada, lambda t, args: 1], id='Hlist_func'),
+        pytest.param([ada, '1'], id='Hlist_str'),
+        pytest.param([ada, Cubic_Spline(0, 10, np.ones_like(tlist))],
+                      id='Hlist_cubic_spline'),
+        pytest.param([ada, np.ones_like(tlist)], id='Hlist_array'),
+        pytest.param(QobjEvo([ada, '1']), id='HQobjEvo'),
+        pytest.param(lambda t, args: create(10) * destroy(10),
[…]

[…]
