### 2021-09-09 56d954c
documenting list of callbacks for measurements
