+++ qutip/sesolve.py
@@ -54,8 +54,10 @@ def sesolve(H, psi0, tlist, e_ops=None, args=None, options=None,
     tlist : array_like of float
         List of times for :math:`t`.
 
-    e_ops : list of :class:`~Qobj` or callback function, optional
-        Single operator or list of operators for which to evaluate expectation
+    e_ops : list of :class:`~Qobj` or list of callback functions or
+    a single callback function, optional
+        Single operator, list of operators or list of callable functions
+        for which to evaluate expectation
         values.  For operator evolution, the overlap is computed: ::
 
             (e_ops[i].dag() * op(t)).tr()
