+++ qutip/mesolve.py
@@ -127,8 +127,10 @@ def mesolve(H, rho0, tlist, c_ops=None, e_ops=None, args=None, options=None,
         single collapse operator, or list of collapse operators, or a list
         of Liouvillian superoperators.
 
-    e_ops : None / list of :class:`qutip.Qobj` / callback function single
-        single operator or list of operators for which to evaluate
+    e_ops : None / list of :class:`qutip.Qobj` / list of callback functions
+    / callback function
+        single operator, list of operators or list of callable functions
+        for which to evaluate
         expectation values.
 
     args : None / *dictionary*
