+++ qutip/solver/ode/scipy_integrator.py
@@ -11,6 +11,7 @@ from qutip.core.data.reshape import column_unstack_dense, column_stack_dense
 from ..integrator import (IntegratorException, Integrator, sesolve_integrators,
                           mesolve_integrators, mcsolve_integrators,
                           add_integrator)
+from ..solver_base import Solver
 import warnings
 
 
@@ -311,8 +312,6 @@ class IntegratorScipylsoda(IntegratorScipyDop853):
         raise IntegratorException(messages[self._ode_solver._integrator.istate])
 
 
-sets = [sesolve_integrators, mesolve_integrators, mcsolve_integrators]
-for integrator_set in sets:
-    add_integrator(IntegratorScipyZvode, ['adams', 'bdf'], integrator_set, SolverOdeOptions)
-    add_integrator(IntegratorScipyDop853, ['dop853'], integrator_set, SolverOdeOptions)
-    add_integrator(IntegratorScipylsoda, ['lsoda'], integrator_set, SolverOdeOptions)
+Solver.add_integrator(IntegratorScipyZvode, ['adams', 'bdf'])
+Solver.add_integrator(IntegratorScipyDop853, ['dop853'])
+Solver.add_integrator(IntegratorScipylsoda, ['lsoda'])
