+++ qutip/solver/solver_base.py
@@ -38,9 +38,9 @@ __all__ = ['Solver']
 # from ..core import data as _data
 
 from .. import Qobj, QobjEvo
-from .options import SolverOptions
+from .options import SolverOptions, SolverOdeOptions
 from .result import Result
-from .integrator import integrator_collection
+from .integrator import Integrator
 from ..ui.progressbar import get_progess_bar
 from ..core.data import to
 from time import time
@@ -53,20 +53,6 @@ class Solver:
     evolver which deal in data and the Result which use Qobj.
     It's children (SeSolver, McSolver) are responsible with building the system
     (-1j*H).
-
-    methods
-    -------
-    run(state0, tlist, args={}):
-        Do an evolution starting with `state0` at `tlist[0]` and obtain a
-        result for each time in `tlist`.
-        The system's arguments can be changed with `args`.
-
-    start(state0, t0):
-        Set the initial values for an evolution by steps
-
-    step(t, args={}):
-        Do a step to `t`. The system arguments for this step can be updated
-        with `args`
 
     attributes
     ----------
@@ -89,90 +75,116 @@ class Solver:
     _t = 0
     _state = None
     _integrator = False
+    _avail_integrators = {}
 
     # Class of option used by the solver
     optionsclass = SolverOptions
+    odeoptionsclass = SolverOdeOptions
 
-    def __init__(self, e_ops, args, feedback_args):
+    def __init__(self):
         raise NotImplementedError
 
     def _prepare_state(self, state):
+        """ Extract the data and metadata of the Qobj state """
         # Do the dims checks
-        # prepare the data from the Qobj (reshape, update type)
-        # return state.data, info
+        # prepare the data from the Qobj (reshape, update type, ...)
+        # metadata pass dims, type, etc., from _prepare_state to _restore_state
+        # return state.data, metadata
         raise NotImplementedError
 
-    def _restore_state(self, state, info, copy=True):
-        # rebuild the Qobj from the state's data
-        # info pass dims, type, etc., from _prepare_state to _restore_state
+    def _restore_state(self, state, metadata, copy=True):
+        """ Retore the Qobj state from the data and metadata """
         raise NotImplementedError
 
     def run(self, state0, tlist, args={}):
-        _data0, info = self._prepare_state(state0)
+        """
+        Do the evolution of the Quantum system.
+
+        Parameters
+        ----------
+        state0 : :class:`Qobj`
+            Initial state of the evolution.
+
+        tlist : list of double
+            Time for which to save the results (state and/or expect) of the
+            evolution. The first element of the list is the initial time of the
+            evolution. Each times of the list must be increasing, but does not
+            need to be uniformy distributed.
+
+        args : dict, optional {None}
+            Set the ``args`` of the system for the evolution.
+
+        Return
+        ------
+        results : :class:`qutip.solver.Result`
+            Results of the evolution. States and/or expect will be saved. You
+            can control the saved data in the options.
+        """
+        _data0, state_metadata = self._prepare_state(state0)
         _integrator = self._get_integrator()
         if args:
             _integrator.update_args(args)
         _time_start = time()
         _integrator.set_state(tlist[0], _data0)
         self.stats["preparation time"] += time() - _time_start
-        res = Result(self.e_ops, self.options.results, state0)
-        res.add(tlist[0], state0)
+        results = Result(self.e_ops, self.options.results, state0)
+        results.add(tlist[0], state0)
 
         progress_bar = get_progess_bar(self.options['progress_bar'])
         progress_bar.start(len(tlist)-1, **self.options['progress_kwargs'])
         for t, state in _integrator.run(tlist):
             progress_bar.update()
-            res.add(t, self._restore_state(state, info, False))
+            results.add(t, self._restore_state(state, state_metadata, False))
         progress_bar.finished()
 
[…]
