+++ qutip/tests/solver/test_mesolve.py
@@ -35,25 +35,26 @@ import numpy as np
 from numpy.testing import assert_allclose
 from types import FunctionType
 
-from qutip import *
-from qutip.solver import *
-from qutip.random_objects import rand_ket
+import qutip
+from qutip.solver.mesolve import mesolve, MeSolver
+from qutip.solver.solver_base import Solver
+from qutip.solver.options import SolverOptions
 import pickle
 import pytest
-all_ode_method = integrator_collection.list_keys('methods', time_dependent=True)
+all_ode_method = Solver._avail_integrators.keys()
 
 
 def fidelitycheck(out1, out2, rho0vec):
     fid = np.zeros(len(out1.states))
     for i, E in enumerate(out2.states):
-        rhot = vector_to_operator(E*rho0vec)
-        fid[i] = fidelity(out1.states[i], rhot)
+        rhot = qutip.vector_to_operator(E*rho0vec)
+        fid[i] = qutip.fidelity(out1.states[i], rhot)
     return fid
 
 
 class TestMESolveDecay:
     N = 10
-    a = destroy(N)
+    a = qutip.destroy(N)
     kappa = 0.2
     tlist = np.linspace(0, 10, 201)
     ada = a.dag() * a
@@ -61,11 +62,11 @@ class TestMESolveDecay:
     @pytest.fixture(params=[
         pytest.param([ada, lambda t, args: 1], id='Hlist_func'),
         pytest.param([ada, '1'], id='Hlist_str'),
-        pytest.param([ada, Cubic_Spline(0, 10, np.ones_like(tlist))],
+        pytest.param([ada, qutip.Cubic_Spline(0, 10, np.ones_like(tlist))],
                       id='Hlist_cubic_spline'),
         pytest.param([ada, np.ones_like(tlist)], id='Hlist_array'),
-        pytest.param(QobjEvo([ada, '1']), id='HQobjEvo'),
-        pytest.param(lambda t, args: create(10) * destroy(10),
+        pytest.param(qutip.QobjEvo([ada, '1']), id='HQobjEvo'),
+        pytest.param(lambda t, args: qutip.create(10) * qutip.destroy(10),
                      id='func'),
     ])
     def H(self, request):
@@ -80,12 +81,12 @@ class TestMESolveDecay:
                   id='list_func'),
         pytest.param([a, 'sqrt(kappa)'],
                   id='list_str'),
-        pytest.param([a, Cubic_Spline(0, 10, np.sqrt(kappa) *
+        pytest.param([a, qutip.Cubic_Spline(0, 10, np.sqrt(kappa) *
                                       np.ones_like(tlist))],
                   id='list_cubic_spline'),
         pytest.param([a, np.sqrt(kappa) * np.ones_like(tlist)],
                   id='list_array'),
-        pytest.param(QobjEvo([a, 'sqrt(kappa)'], args={'kappa': kappa}),
+        pytest.param(qutip.QobjEvo([a, 'sqrt(kappa)'], args={'kappa': kappa}),
                   id='QobjEvo'),
     ])
     def cte_c_ops(self, request):
@@ -96,11 +97,11 @@ class TestMESolveDecay:
                   id='list_func'),
         pytest.param([a, 'sqrt(kappa * exp(-t))'],
                   id='list_str'),
-        pytest.param([a, Cubic_Spline(0, 10, np.sqrt(kappa * np.exp(-tlist)))],
+        pytest.param([a, qutip.Cubic_Spline(0, 10, np.sqrt(kappa * np.exp(-tlist)))],
                   id='list_cubic_spline'),
[…]

[…]
