### 2021-09-10 42aff8e
implemented changes requested during code review
