+++ qutip/mesolve.py
@@ -509,7 +509,8 @@ def _generic_ode_solve(func, ode_args, rho0, tlist, e_ops, opt,
             output.expect.append(e_ops(t, rho_t))
 
         for m in range(n_expt_op):
-            if not isinstance(e_ops[m], (Qobj, QobjEvo)) and callable(e_ops[m]):
+            if not isinstance(e_ops[m], (Qobj, QobjEvo)) \
+               and callable(e_ops[m]):
                 output.expect[m][t_idx] = e_ops[m](t, rho_t)
                 continue
             output.expect[m][t_idx] = expect_rho_vec(e_ops_data[m], r.y,
