### 2021-09-10 dd4adc8
Improve trace norm docstring.
