### 2021-09-10 b97a62c
fix stochastic tests seeds
