### 2021-09-10 66b3eff
add Classmethod
