### 2021-09-10 c606186
fix get_state output
