### 2021-09-10 5a8cb35
Added steadystate_floquet function
