+++ qutip/steadystate.py
@@ -4,8 +4,8 @@ open quantum systems defined by a Liouvillian or Hamiltonian and a list of
 collapse operators.
 """
 
-__all__ = ['steadystate', 'steady', 'steadystate_floquet', 'build_preconditioner',
-           'pseudo_inverse']
+__all__ = ['steadystate', 'steady', 'steadystate_floquet',
+           'build_preconditioner', 'pseudo_inverse']
 
 import functools
 import time
@@ -912,16 +912,18 @@ def _steadystate_power(L, ss_args):
         return rhoss, ss_args['info']
     else:
         return rhoss
-        
 
-def steadystate_floquet(H_0, c_ops, Op_t, w_d = 1.0, nmax = 3, sparse = False):
+
+def steadystate_floquet(H_0, c_ops, Op_t, w_d=1.0, nmax=3, sparse=False):
     """
-    Calculates the effective steady state for a driven system with a time-dependent cosinusoidal term:
-    
+    Calculates the effective steady state for a driven
+     system with a time-dependent cosinusoidal term:
+
     .. math::
 
-        \\mathcal{\\hat{H}}(t) = \\hat{H}_0 + \\mathcal{\\hat{O}} \\cos(\\omega_d t)
-    
+        \\mathcal{\\hat{H}}(t) = \\hat{H}_0 +
+         \\mathcal{\\hat{O}} \\cos(\\omega_d t)
+
     Parameters
     ----------
     H_0 : :obj:`~Qobj`
@@ -929,19 +931,20 @@ def steadystate_floquet(H_0, c_ops, Op_t, w_d = 1.0, nmax = 3, sparse = False):
 
     c_ops : list
         A list of collapse operators.
-        
+
     Op_t : :obj:`~Qobj`
         The the interaction operator which is multiplied by the cosine
-        
+
     w_d : float, default 1.0
         The frequency of the drive
-        
+
     nmax : int, default 3
         The maximum number of iteration for the solver
-        
+
     sparse : bool, default False
-        Solve for the steady state using sparse algorithms. Actually, dense seems to be faster.
-        
+        Solve for the steady state using sparse algorithms.
+        Actually, dense seems to be faster.
+
     Returns
     -------
     dm : qobj
@@ -953,22 +956,22 @@ def steadystate_floquet(H_0, c_ops, Op_t, w_d = 1.0, nmax = 3, sparse = False):
         L_0 = liouvillian(H_0, c_ops).data.tocsc()
         L_t = liouvillian(Op_t)
         L_p = (0.5 * L_t).data.tocsc()
-        L_m = L_p #(0.5 * A_l * L_t).data
+        L_m = L_p  # (0.5 * A_l * L_t).data
         L_p_array = L_p.todense()
         L_m_array = L_m.todense()
 
-        Id = sp.eye(N ** 2, format = "csc", dtype = np.complex128)
-        S, T = sp.csc_matrix((N ** 2, N ** 2), dtype = np.complex128), sp.csc_matrix((N ** 2, N ** 2), dtype = np.complex128)
+        Id = sp.eye(N ** 2, format="csc", dtype=np.complex128)
+        S = T = sp.csc_matrix((N ** 2, N ** 2), dtype=np.complex128)
 
         for n_i in np.arange(nmax, 0, -1):
-            L = sp.csc_matrix( L_0 - 1j * n_i * w_d * Id + L_m.dot(S) )
+            L = sp.csc_matrix(L_0 - 1j * n_i * w_d * Id + L_m.dot(S))
             L.sort_indices()
[…]
