### 2021-09-11 11a52f9
Reduce precision in spin Q and wigner tests for a speed-up.
