### 2021-09-12 bdf5e1a
Added the test function in the qutip directory
