+++ qutip/tests/test_correlation.py
@@ -271,3 +271,42 @@ def test_correlation_2op_1t_known_cases(solver,
     cmp = qutip.correlation_2op_1t(H, state, times, c_ops, a_op, b_op, solver=solver)
 
     np.testing.assert_allclose(base, cmp, atol=0.25 if solver == 'mc' else 2e-5)
+    
+    
+@pytest.mark.parametrize("sparse", [False, True])
+def test_steadystate_floquet(sparse):
+    """
+    Test the steadystate solution for a periodically
+    driven system.
+    """
+    N_c = _equivalence_dimension
+    
+    a = qutip.destroy(N_c)
+    a_d = a.dag()
+    X_c = a + a_d
+
+    w_c = 1
+
+    A_l = 0.001
+    w_l = w_c
+    gam = 0.01
+
+    H = w_c * a_d * a
+
+    H_t = [H, [X_c, lambda t, args: args["A_l"] * np.cos(args["w_l"] * t)]]
+
+    psi0 = qutip.fock(N_c, 0);
+
+    args = {"A_l": A_l, "w_l": w_l}
+
+    c_ops = []
+    c_ops.append(np.sqrt(gam) * a)
+
+    t_l = np.linspace(0, 20 / gam, 2000)
+
+    expect_me = qutip.mesolve(H_t, psi0, t_l, c_ops, [a_d * a], args = args).expect[0]
+
+    rho_ss = qutip.steadystate_floquet(H, c_ops, A_l * X_c, w_l, nmax = 3, sparse = sparse)
+    expect_ss = qutip.expect(a_d * a, rho_ss)
+    
+    np.testing.assert_allclose( expect_me[-20:], expect_ss, atol=1e-3 )
