### 2021-09-12 675374f
Added the test function in the test_steadystate file
