+++ qutip/tests/solver/test_mesolve.py
@@ -106,7 +106,7 @@ class TestMESolveDecay:
                           args={'kappa': kappa}),
                   id='QobjEvo'),
         pytest.param(lambda t, args: (np.sqrt(args['kappa'] * np.exp(-t)) *
-                                      destroy(10)),
+                                      qutip.destroy(10)),
                      id='func'),
     ])
     def c_ops(self, request):
@@ -175,7 +175,7 @@ class TestMESolveDecay:
     def testME_TDH_longTDDecay(self, H, c_ops):
         "mesolve: time-dependence as function list"
         me_error = 2e-5
-        psi0 = basis(self.N, 9)  # initial state
+        psi0 = qutip.basis(self.N, 9)  # initial state
         if isinstance(c_ops, FunctionType):
             return
         if isinstance(c_ops, qutip.QobjEvo):
@@ -198,7 +198,7 @@ class TestMESolveDecay:
         H = self.ada
         psi0 = qutip.basis(self.N, 9)  # initial state
         rho0vec = qutip.operator_to_vector(psi0*psi0.dag())
-        E0 = qutip.sprepost(qeye(self.N), qeye(self.N))
+        E0 = qutip.sprepost(qutip.qeye(self.N), qutip.qeye(self.N))
         options = SolverOptions(progress_bar=None)
         c_op_list = [c_ops]
         out1 = mesolve(H, psi0, self.tlist, c_op_list, [],
@@ -219,7 +219,7 @@ class TestMESolveDecay:
         L = qutip.liouvillian(H)
         psi0 = qutip.basis(self.N, 9)  # initial state
         rho0vec = qutip.operator_to_vector(psi0*psi0.dag())
-        E0 = qutip.sprepost(qeye(self.N), qeye(self.N))
+        E0 = qutip.sprepost(qutip.qeye(self.N), qutip.qeye(self.N))
         options = SolverOptions(progress_bar=None)
         c_op_list = [c_ops]
         out1 = mesolve(L, psi0, self.tlist, c_op_list, [],
@@ -252,7 +252,7 @@ class TestMESolveDecay:
                               args={"kappa":1.}, options=options)
         solver_obj.start(qutip.basis(self.N, 9), 0)
         state1 = solver_obj.step(1)
-        assert qutip.expect(self.ada, state1) != qutip.expect(self.ada, basis(self.N, 9))
+        assert qutip.expect(self.ada, state1) != qutip.expect(self.ada, qutip.basis(self.N, 9))
         state2 = solver_obj.step(2, {"kappa":0.})
         assert_allclose(qutip.expect(self.ada, state1),
                         qutip.expect(self.ada, state2), 1e-6)
@@ -263,10 +263,10 @@ class TestMESolveDecay:
 def testME_SesolveFallback(super_):
     "mesolve: final_state has correct dims"
     N = 5
-    a = tensor(qutip.destroy(N+1), qutip.qeye(N+1), qutip.qeye(N+1))
-    b = tensor(qutip.qeye(N+1), qutip.destroy(N+1), qutip.qeye(N+1))
-    c = tensor(qutip.qeye(N+1), qutip.qeye(N+1), qutip.destroy(N+1))
-    psi0 = tensor(qutip.basis(N+1,0), qutip.basis(N+1,0), qutip.basis(N+1,N))
+    a = qutip.tensor(qutip.destroy(N+1), qutip.qeye(N+1), qutip.qeye(N+1))
+    b = qutip.tensor(qutip.qeye(N+1), qutip.destroy(N+1), qutip.qeye(N+1))
+    c = qutip.tensor(qutip.qeye(N+1), qutip.qeye(N+1), qutip.destroy(N+1))
[…]
