### 2021-09-13 d33cbbf
Allow reset before set_state
