### 2021-09-14 f9a8ad9
Update qutip/solver/integrator.py
Co-authored-by: Asier Galicia <57414022+AGaliciaMartinez@users.noreply.github.com>
