### 2021-09-15 065fdde
fixed incorrect test of a callable
