### 2021-09-15 381ac39
Add settings option for function coefficient signature.
