+++ qutip/core/options.py
@@ -29,6 +29,22 @@ class CoreOptions:
         The absolute tolerance used in automatic tidyup (see the ``auto_tidyup``
         parameter above) and the default value of ``atol`` used in
         :method:`Qobj.tidyup`.
+
+    function_coefficient_signature : str {"auto"}
+        The signature expected by function coefficients. The options are:
+
+        - "pythonic": the signature should be ``f(t, ...)`` where ``t``
+          is the time and the ``...`` are the remaining arguments passed
+          directly into the function. E.g. ``f(t, w, b=5)``.
+
+        - "dict": the signature shoule be ``f(t, args)`` where ``t`` is
+          the time and ``args`` is a dict containing the remaining arguments.
+          E.g. ``f(t, {"w": w, "b": 5})``.
+
+        - "auto": select automatically between the two options above based
+          on signature of the supplied function. If the function signature
+          is exactly ``f(t, args)`` then ``dict`` is used. Otherwise
+          ``pythonic`` is used.
     """
     options = {
         # use auto tidyup
@@ -42,5 +58,7 @@ class CoreOptions:
         # general relative tolerance
         "rtol": 1e-12,
         # use auto tidyup absolute tolerance
-        "auto_tidyup_atol": 1e-14
+        "auto_tidyup_atol": 1e-14,
+        # signature style expected by function coefficients
+        "function_coefficient_signature": "auto",
     }
