### 2021-09-15 282c770
Fix up docstrings.
