### 2021-09-15 97df3a6
Add function_style argument to coefficient and QobjEvo.
