### 2021-09-15 7464201
Document function_style for coefficient.
