### 2021-09-15 8d25c48
Fix up references to function style setting in docstrings.
