### 2021-09-15 6199d93
Merge pull request #1632 from maij/fix/spin_q_function_transcription
Fix bugs in spin_q_function and spin_wigner and add test cases.
