### 2021-09-15 c8c962d
Fix typo in coefficient docstring.
