### 2021-09-16 3d9fbe1
Set types of _f_pythonic and _f_parameters and add additional check on how they are passed in.
