### 2021-09-16 f76b505
Avoid copying function coefficient args unnecessarily.
