### 2021-09-16 f76005d
Copy function coefficient args if they're supplied by the user so that we don't rely on external dictionaries not being modified.
