+++ qutip/core/cy/coefficient.pyx
@@ -200,6 +200,8 @@ cdef class FunctionCoefficient(Coefficient):
                 func, style=style)
             if _f_parameters is not None:
                 args = {k: args[k] for k in _f_parameters & args.keys()}
+            else:
+                args = args.copy()
         self.func = func
         self.args = args
         self._f_pythonic = _f_pythonic
