### 2021-09-17 d3220f9
add element_wise_multiply
