+++ qutip/tests/core/data/test_mathematics.py
@@ -731,6 +731,18 @@ class TestMatmul(BinaryOpMixin):
     ]
 
 
+class TestElement_wise_multiply(BinaryOpMixin):
+    def op_numpy(self, left, right):
+        return left * right
+
+    shapes = shapes_binary_identical()
+    bad_shapes = shapes_binary_bad_identical()
+    specialisations = [
+        pytest.param(data.element_wise_multiply_csr, CSR, CSR, CSR),
+        pytest.param(data.element_wise_multiply_dense, Dense, Dense, Dense),
+    ]
+
+
 class TestMul(UnaryScalarOpMixin):
     def op_numpy(self, matrix, scalar):
         return scalar * matrix
