### 2021-09-17 fad659e
Refactor _process_input into helper functions and move all attribute assignment into __init__.
