+++ qutip/nonmarkov/bofin.py
@@ -88,51 +88,158 @@ def _heom_state_dictionaries(dims, excitations):
     return nstates, state2idx, idx2state
 
 
-def _check_Hsys(H_sys):
-    # Check if Hamiltonians are one of the allowed types
-    if not isinstance(H_sys, (Qobj, QobjEvo, list)):
-        msg = "Hamiltonian format is incorrect."
-        raise RuntimeError(msg)
-
-    # Check if Hamiltonians supplied in the list are correct
-    if isinstance(H_sys, list):
-        for H in H_sys:
-            # If not a list of time dependent Hamiltonians
-            if not isinstance(H, list):
-                # Just check if it is a Qobj
-                _check_Hsys(H)
-
-            # Check if time dependent Hamiltonian terms are correct
-            # in the list fomat if it is a tuple of [H, callable]
-            elif isinstance(H[0], Qobj):
-                if not callable(H[1]):
-                    msg = "Incorrect time dependent function for Hamiltonian."
-                    raise RuntimeError(msg)
-
-            else:
-                _check_Hsys(H[0])
-
-        return True
-
+def _convert_h_sys(H_sys):
+    """ Process input system Hamiltonian, converting and raising as needed.
+    """
+    if isinstance(H_sys, (Qobj, QobjEvo)):
+        pass
+    elif isinstance(H_sys, list):
+        try:
+            H_sys = QobjEvo(H_sys)
+        except Exception as err:
+            raise ValueError(
+                "Hamiltonian (H_sys) of type list cannot be converted to"
+                " QObjEvo"
+            ) from err
     else:
-        return True
+        raise TypeError(
+            f"Hamiltonian (H_sys) has unsupported type: {type(H_sys)!r}")
+    return H_sys
 
 
-def _check_coup_ops(coup_op, length):
-
-    if (type(coup_op) != Qobj) and (
-        type(coup_op) == list and type(coup_op[0]) != Qobj
-    ):
-        raise RuntimeError(
-            "Coupling operator must be a QObj or list " + " of QObjs."
+def _convert_coup_op(coup_op, coup_op_len):
+    """ Convert coup_op to a list of the appropriate length. """
+    if isinstance(coup_op, Qobj):
+        coup_op = [coup_op] * coup_op_len
+    elif (isinstance(coup_op, list)
+            and all(isinstance(x, Qobj) for x in coup_op)):
+        if len(coup_op) != coup_op_len:
+            raise ValueError(
+                f"Expected {coup_op_len} coupling operators.")
+    else:
+        raise TypeError(
+            "Coupling operator (coup_op) must be a Qobj or a list of Qobjs"
         )
+    return coup_op
 
-    if type(coup_op) == list:
-        if len(coup_op) != (length):
-            raise RuntimeError(
-                "Expected " + str(length)
-                + " coupling operators."
-            )
+
+def _convert_bath_exponents_bosonic(ckAI, ckAR, vkAI, vkAR):
+    all_k = (ckAI, ckAR, vkAI, vkAR)
+    if any(not isinstance(k, list) for k in all_k):
+        raise TypeError(
+            "The bath exponents ckAI, ckAR, vkAI and vkAR must all be lists")
+    if len(ckAI) != len(vkAI) or len(ckAR) != len(vkAR):
+        raise ValueError(
+            "The bath exponent lists ckAI and vkAI, and ckAR and vkAR must"
+            " be the same length"
+        )
+    if any(isinstance(x, list) for k in all_k for x in k):
+        raise ValueError(
+            "The bath exponent lists ckAI, ckAR, vkAI and vkAR should not"
+            " themselves contain lists"
+        )
+    # warn if any of the vkAR's are close
+    for i in range(len(vkAR)):
[…]
