### 2021-09-20 5b670ae
Simplify construction of solver parameters in time-dependent case. Add RHS matrix to solver parameters in Fermionic case.
