### 2021-09-20 f65f48a
Add BathStates class.
