+++ qutip/nonmarkov/bofin.py
@@ -8,7 +8,6 @@ hierarchy equations of motion (HEOM).
 
 import warnings
 from copy import deepcopy
-from math import factorial
 
 import numpy as np
 import scipy.sparse as sp
@@ -27,65 +26,97 @@ from qutip.cy.spconvert import dense2D_to_fastcsr_fmode
 from qutip.ui.progressbar import BaseProgressBar
 
 
-def add_at_idx(seq, k, val):
+class BathStates:
     """
-    Add (subtract) a value in the tuple at position k
-    """
-    list_seq = list(seq)
-    list_seq[k] += val
-    return tuple(list_seq)
+    A description of bath states coupled to quantum system in the hierarchical
+    equations of motion formulation.
 
-
-def prevhe(current_he, k, ncut):
-    """
-    Calculate the previous heirarchy index
-    for the current index `n`.
-    """
-    nprev = add_at_idx(current_he, k, -1)
-    if nprev[k] < 0:
-        return False
-    return nprev
-
-
-def nexthe(current_he, k, ncut):
-    """
-    Calculate the next heirarchy index
-    for the current index `n`.
-    """
-    nnext = add_at_idx(current_he, k, 1)
-    if sum(nnext) > ncut:
-        return False
-    return nnext
-
-
-def _heom_state_dictionaries(dims, excitations):
-    """
-    Return the number of states, and lookup-dictionaries for translating
-    a state tuple to a state index, and vice versa, for a system with a given
-    number of components and maximum number of excitations.
     Parameters
     ----------
-    dims: list
-        A list with the number of states in each sub-system.
-    excitations : integer
-        The maximum numbers of dimension
-    Returns
-    -------
-    nstates, state2idx, idx2state: integer, dict, dict
-        The number of states `nstates`, a dictionary for looking up state
-        indices from a state tuple, and a dictionary for looking up state
-        state tuples from state indices.
+    cutoff : int
+        The maximum number of excitations.
+    dims : list of int
+        The dimensions of each compontent system within the bath.
+
+    Attributes
+    ----------
+    cutoff : int
+        The maximum number of excitations.
+    dims : list of int
+        The dimensions of each compontent system within the bath.
+    states: list of tuples
+        A list of the state vectors within the bath.
+    n_states: int
+        Total number of states. Equivalent to ``len(bath.states)``.
     """
-    nstates = 0
-    state2idx = {}
-    idx2state = {}
+    def __init__(self, dims, cutoff):
+        self.cutoff = cutoff
+        self.dims = dims
+        self.states = list(state_number_enumerate(dims, cutoff))
+        self.n_states = len(self.states)
+        self._state_idx = {s: i for i, s in enumerate(self.states)}
 
-    for state in state_number_enumerate(dims, excitations):
-        state2idx[state] = nstates
-        idx2state[nstates] = state
-        nstates += 1
+    def idx(self, state):
+        """
+        Return the index of the state within the list of bath states,
+        i.e. within ``self.states``.
 
-    return nstates, state2idx, idx2state
+        Parameters
+        ----------
+        state : tuple
+            The state to look up.
+
+        Returns
+        -------
+        int
+            The index of the state within the list of bath states.
+        """
+        return self._state_idx[state]
+
+    def next(self, state, k):
+        """
+        Return the state with one more excitation in the k'th bath dimension
[…]
