### 2021-09-20 b3c6be7
Factor out more initialization into __init__.
