+++ qutip/nonmarkov/bofin.py
@@ -261,8 +261,21 @@ def _h_sys_is_hamiltonian(H_sys):
     return H_sys.type == "oper"
 
 
+def _h_sys_dim(H_sys, isHamiltonian):
+    """ Return the system dimension for the given system.
+    """
+    if type(H_sys) is QobjEvo:
+        H_sys_list = H_sys.to_list()
+        H_shape = H_sys_list[0].shape
+    else:
+        H_shape = H_sys.shape
+
+    H_sys_dim = H_shape[0] if isHamiltonian else int(np.sqrt(H_shape[0]))
+    return H_sys_dim
+
+
 def _h_sup_dim(H_sys, isHamiltonian):
-    """ Return the super operator dimension for the given hamiltonian.
+    """ Return the super operator dimension for the given system.
     """
     if type(H_sys) is QobjEvo:
         H_sys_list = H_sys.to_list()
@@ -271,6 +284,16 @@ def _h_sup_dim(H_sys, isHamiltonian):
         H_shape = H_sys.shape
     H_sup_dim = H_shape[0] ** 2 if isHamiltonian else H_shape[0]
     return H_sup_dim
+
+
+def _h_sys_liouvillian(H_sys, isHamiltonian):
+    """ Return the Liouvillian dimension for the given system.
+    """
+    if type(H_sys) is QobjEvo:
+        H_sys_list = H_sys.to_list()
+        H_sys = H_sys_list[0]
+    L = liouvillian(H_sys) if isHamiltonian else H_sys
+    return L.data
 
 
 class BosonicHEOMSolver(object):
@@ -345,6 +368,9 @@ class BosonicHEOMSolver(object):
         self.isTimeDep = isinstance(self.H_sys, QobjEvo)
         self.isHamiltonian = _h_sys_is_hamiltonian(self.H_sys)
         self._sup_dim = _h_sup_dim(self.H_sys, self.isHamiltonian)
+        self._sys_dim = _h_sys_dim(self.H_sys, self.isHamiltonian)
+        self._h_sys_dims = self.H_sys.dims if not self.isTimeDep else self.H_sys.to_list()[0].dims
+        self.L = _h_sys_liouvillian(self.H_sys, self.isHamiltonian)
         self.N_cut = N_cut
         coup_op = _convert_coup_op(coup_op, len(ckAR) + len(ckAI))
         ckAR, ckAI, vkAR, vkAI = _convert_bath_exponents_bosonic(
@@ -354,6 +380,9 @@ class BosonicHEOMSolver(object):
         )
 
         self.progress_bar = BaseProgressBar()
+
+        self.spreQ = [spre(op).data for op in self.coup_op]
+        self.spostQ = [spost(op).data for op in self.coup_op]
 
         self.kcut = len(self.coup_op)
         self.bath = BathStates([self.N_cut + 1] * self.kcut, self.N_cut)
@@ -469,46 +498,9 @@ class BosonicHEOMSolver(object):
 
         return RHS
 
-    def _boson_solver(self):
-        """
-        Utility function for bosonic solver.
-        """
-        # Separate cases for Hamiltonian and Liouvillian
-        if self.isHamiltonian:
-
-            if self.isTimeDep:
-                H_sys_list = self.H_sys.to_list()
-                self.N = H_sys_list[0].shape[0]
-                self.L = liouvillian(H_sys_list[0], []).data
-
-            else:
-                self.N = self.H_sys.shape[0]
-                self.L = liouvillian(self.H_sys, []).data
-
-        else:
-            if self.isTimeDep:
-                H_sys_list = self.H_sys.to_list()
-                self.N = int(np.sqrt(H_sys_list[0].shape[0]))
-                self.L = H_sys_list[0].data
-
-            else:
-                self.N = int(np.sqrt(self.H_sys.shape[0]))
[…]
