### 2021-09-20 3946a3b
Make mcsolve more memory efficient mcsolve would always save the final state of all trajectories, needed or not, Now only save them when needed.
