### 2021-09-21 8cb0f07
Small syntax clean ups.
