### 2021-09-21 4d86f23
Remove unneeded copy
