### 2021-09-21 00e96b6
Simplify sums of vk's in gradient sums.
