### 2021-09-21 af80846
self.states_out set to None when unused
