### 2021-09-22 870bb55
Minor changes about notation
