### 2021-09-22 d5e7c59
Added the reference of the algorithm
