### 2021-09-23 a24450b
first set of corrections
