### 2021-09-23 bfe3631
Apply suggestions from code review
Co-authored-by: Eric Giguère <eric.giguere@calculquebec.ca>
