+++ qutip/sesolve.py
@@ -54,10 +54,13 @@ def sesolve(H, psi0, tlist, e_ops=None, args=None, options=None,
     tlist : array_like of float
         List of times for :math:`t`.
 
-    e_ops : None / list of :class:`qutip.Qobj` and callback functions, optional
-        single operator, list of operators and callable functions
-        for which to evaluate expectation values. For operator evolution,
-        the overlap is computed: ::
+    e_ops : None / list / callback function, optional
+        A list of operators as `Qobj` and/or callable functions (can be mixed)
+        or a single callable function. For callable functions, they are called as ``f(t, state)`` 
+        and return the expectation value. A single callback's expectation value can be any type, 
+        but a callback part of a list must return a number as the expectation value. For operators, 
+        the result's expect will be  computed by :func:`qutip.expect` when the state is a ``ket``.
+        For operator evolution, the overlap is computed by: ::
 
             (e_ops[i].dag() * op(t)).tr()
 
@@ -276,7 +279,7 @@ def _generic_ode_solve(func, ode_args, psi0, tlist, e_ops, opt,
             opt.store_states = True
         else:
             for op in e_ops:
-                if not isinstance(op, (Qobj, QobjEvo)) and callable(op):
+                if not isinstance(op, Qobj) and callable(op):
                     output.expect.append(np.zeros(n_tsteps, dtype=complex))
                     continue
                 if op.isherm:
@@ -285,13 +288,13 @@ def _generic_ode_solve(func, ode_args, psi0, tlist, e_ops, opt,
                     output.expect.append(np.zeros(n_tsteps, dtype=complex))
         if oper_evo:
             for e in e_ops:
-                if isinstance(e, (Qobj, QobjEvo)):
+                if isinstance(e, Qobj):
                     e_ops_data.append(e.dag().data)
                     continue
                 e_ops_data.append(e)
         else:
             for e in e_ops:
-                if isinstance(e, (Qobj, QobjEvo)):
+                if isinstance(e, Qobj):
                     e_ops_data.append(e.data)
                     continue
                 e_ops_data.append(e)
