### 2021-09-23 30724cd
Update qutip/solver/integrator.py
Co-authored-by: Simon Cross <hodgestar+github@gmail.com>
