+++ qutip/solver/ode/scipy_integrator.py
@@ -56,6 +56,7 @@ class IntegratorScipyZvode(Integrator):
     def get_state(self, copy=True):
         if not self._is_set:
             raise RuntimeError("The state is not initialted")
+        self._check_failed_integration()
         t = self._ode_solver.t
         if self._mat_state:
             state = _data.column_unstack_dense(
@@ -66,14 +67,22 @@ class IntegratorScipyZvode(Integrator):
             state = _data.dense.Dense(self._ode_solver._y, copy=copy)
         return t, state
 
-    def integrate(self, t, step=False, copy=True):
-        if step:
-            self._one_step(t)
-        elif self._ode_solver.t < t:
+    def integrate(self, t, copy=True):
+        if t != self._ode_solver.t:
             self._ode_solver.integrate(t)
-        elif self._ode_solver.t > t:
+        return self.get_state(copy)
+
+    def mcstep(self, t, copy=True):
+        if self._ode_solver.t < t:
+            self._back = self.get_state(copy=True)
+            self._one_step(t)
+        elif self._back[0] <= t:
             self._backstep(t)
-        self._check_failed_integration()
+        else:
+            raise IntegratorException(
+                "`t` is outside the integration range: "
+                f"{self._back[0]}..{self._ode_solver.t}."
+            )
         return self.get_state(copy)
 
     def _one_step(self, t):
@@ -82,10 +91,8 @@ class IntegratorScipyZvode(Integrator):
         Should be able to retreive the state at any time between the present
         time and the resulting time using `backstep`.
         """
-        # integrate(t, step=True) ignore the time and advance one step.
         # Here we want to advance up to t doing maximum one step.
-        # So we check if a new step is really needed.
-        self._back = self.get_state(copy=True)
+        # We check if a new step is needed.
         t_front = self._ode_solver._integrator.rwork[12]
         t_ode = self._ode_solver.t
         if t > t_front and t_ode >= t_front:
@@ -111,11 +118,6 @@ class IntegratorScipyZvode(Integrator):
         # but not over all the step interval.
         # About 90% of the last step interval?
         # Scipy raise a warning, not an Error.
-        t_prev, _ = self._back
-        if t < t_prev:
-            raise IntegratorException(
-                "Cannot integrate to previous times without using integrate "
-                "with step=True first.")
         with warnings.catch_warnings():
             warnings.filterwarnings("ignore")
             self._ode_solver.integrate(t)
@@ -137,7 +139,9 @@ class IntegratorScipyZvode(Integrator):
             -6: 'Error weight became zero during problem. (Solution'
                 ' component i vanished, and ATOL or ATOL(i) = 0.)'
         }
-        raise IntegratorException(messages[self._ode_solver._integrator.istate])
+        raise IntegratorException(
+            messages[self._ode_solver._integrator.istate]
+        )
 
 
 class IntegratorScipyDop853(Integrator):
@@ -171,29 +175,31 @@ class IntegratorScipyDop853(Integrator):
         column_stack_dense(out, inplace=True)
         return out.as_ndarray().ravel().view(np.float64)
 
-    def integrate(self, t, step=False, copy=True):
-        if step:
+    def integrate(self, t, copy=True):
+        if t != self._ode_solver.t:
+            self._ode_solver.integrate(t)
+        return self.get_state(copy)
+
+    def mcstep(self, t, copy=True):
+        if self._ode_solver.t <= t:
             # Scipy's DOP853 does not have a step function.
             # It has a safe step lengh, but can be 0 if unknown.
             dt = self._ode_solver._integrator.work[6]
             if dt:
                 t = min(self._ode_solver.t + dt, t)
-
-        if self._ode_solver.t > t:
+            self._ode_solver.integrate(t)
+        else:
[…]
