### 2021-09-23 501440e
Compute the tensor directly in sparse format
