### 2021-09-23 a6e5036
tests use np.testing.assert_allclose
