### 2021-09-24 e436b31
Simplify sign calculations in fermionic case.
