### 2021-09-24 16fdb23
Implement Bath and BathMode.
