+++ qutip/nonmarkov/bofin.py
@@ -6,6 +6,8 @@ hierarchy equations of motion (HEOM).
 # Authors: Neill Lambert, Tarun Raheja, Shahnawaz Ahmed
 # Contact: nwlambert@gmail.com
 
+import enum
+import itertools
 import warnings
 from copy import deepcopy
 
@@ -24,6 +26,79 @@ from qutip.cy.spmatfuncs import cy_ode_rhs
 from qutip.solver import Options, Result
 from qutip.cy.spconvert import dense2D_to_fastcsr_fmode
 from qutip.ui.progressbar import BaseProgressBar
+
+
+class BathMode:
+    """
+    Represents a single excitation mode within a bath.
+    """
+
+    types = enum.Enum("BathModeTypes", ["R", "I", "RI", "+", "-"])
+
+    def __init__(self, type, dim, Q, ck, vk, ck2=None):
+        if not isinstance(type, self.types):
+            type = self.types[type]
+        self.type = type
+        self.dim = dim
+        self.Q = Q
+        self.ck = ck
+        self.vk = vk
+        self.ck2 = ck2
+
+
+class Bath:
+    """
+    Represents a bath (i.e. a list of modes).
+    """
+    def __init__(self, modes):
+        self.modes = modes
+
+
+class BosonicBath(Bath):
+    def __init__(self, Q, ckAR, vkAR, ckAI, vkAI, combine=True):
+        Q = _convert_coup_op(Q, len(ckAR) + len(ckAI))
+        ckAR, ckAI, vkAR, vkAI = _convert_bath_exponents_bosonic(
+            ckAR, ckAI, vkAR, vkAI)
+
+        if combine:
+            Q, ck, vk, NR, NI = (
+                _mangle_bath_exponents_bosonic(Q, ckAR, ckAI, vkAR, vkAI)
+            )
+        else:
+            ck = ckAR + ckAI
+            vk = vkAR + vkAI
+            NR = len(ckAR)
+            NI = len(ckAI)
+
+        modes = []
+        modes.extend(
+            BathMode("R", None, Q[i], ck[i], vk[i])
+            for i in range(0, NR))
+        modes.extend(
+            BathMode("I", None, Q[i], ck[i], vk[i])
+            for i in range(NR, NR + NI))
+        modes.extend(
+            BathMode("RI", None, Q[i], ck[i], vk[i], ck2=ck[i+1])
+            for i in range(NR + NI, len(ck), 2))
+        super().__init__(modes)
+
+
+class FermionicBath(Bath):
+    def __init__(self, Q, ck, vk):
+        ck, vk = _convert_bath_exponents_fermionic(ck, vk)
+        Q = _convert_coup_op(Q, len(ck))
+
+        modes = []
+        for i in range(len(ck)):
+            # currently "-" modes are generated by adding extra
+            # baths outside with Q == Q.dag() when calling
+            # FermionicHEOMSolver
+            modes.extend(
+                BathMode("+", 2, Q[i], ck[i][j], vk[i][j])
+                for j in range(len(ck[i]))
+            )
+
+        super().__init__(modes)
 
 
 class BathStates:
@@ -49,10 +124,16 @@ class BathStates:
     n_states: int
         Total number of states. Equivalent to ``len(bath.states)``.
     """
-    def __init__(self, dims, cutoff):
+    def __init__(self, modes, cutoff):
+        self.modes = modes
         self.cutoff = cutoff
-        self.dims = dims
-        self.states = list(state_number_enumerate(dims, cutoff))
+
+        self.dims = [mode.dim or (cutoff + 1) for mode in self.modes]
+        self.vk = [mode.vk for mode in self.modes]
+        self.ck = [mode.ck for mode in self.modes]
[…]
