+++ qutip/nonmarkov/bofin.py
@@ -7,7 +7,6 @@ hierarchy equations of motion (HEOM).
 # Contact: nwlambert@gmail.com
 
 import enum
-import itertools
 import warnings
 from copy import deepcopy
 
@@ -35,15 +34,24 @@ class BathMode:
 
     types = enum.Enum("BathModeTypes", ["R", "I", "RI", "+", "-"])
 
-    def __init__(self, type, dim, Q, ck, vk, ck2=None):
+    def __init__(
+            self, type, dim, Q, ck, vk, ck2=None, sigma_bar_k_offset=None):
         if not isinstance(type, self.types):
             type = self.types[type]
+        if type == self.types["RI"] and ck2 is None:
+            raise ValueError("RI bath modes require ck2")
+        if (
+                type in (self.types["+"], self.types["-"]) and
+                sigma_bar_k_offset is None
+        ):
+            raise ValueError("+ and - bath modes require sigma_bar_k")
         self.type = type
         self.dim = dim
         self.Q = Q
         self.ck = ck
         self.vk = vk
         self.ck2 = ck2
+        self.sigma_bar_k_offset = sigma_bar_k_offset
 
 
 class Bath:
@@ -93,8 +101,17 @@ class FermionicBath(Bath):
             # currently "-" modes are generated by adding extra
             # baths outside with Q == Q.dag() when calling
             # FermionicHEOMSolver
+            if i % 2 == 0:
+                type = "+"
+                sbk_offset = len(ck[i])
+            else:
+                type = "-"
+                sbk_offset = -len(ck[i - 1])
             modes.extend(
-                BathMode("+", 2, Q[i], ck[i][j], vk[i][j])
+                BathMode(
+                    type, 2, Q[i], ck[i][j], vk[i][j],
+                    sigma_bar_k_offset=sbk_offset
+                )
                 for j in range(len(ck[i]))
             )
 
@@ -132,6 +149,9 @@ class BathStates:
         self.vk = [mode.vk for mode in self.modes]
         self.ck = [mode.ck for mode in self.modes]
         self.ck2 = [mode.ck2 for mode in self.modes]
+        self.sigma_bar_k_offset = [
+            mode.sigma_bar_k_offset for mode in self.modes
+        ]
 
         self.states = list(state_number_enumerate(self.dims, cutoff))
         self.n_states = len(self.states)
@@ -458,6 +478,7 @@ class BosonicHEOMSolver(object):
         self.N_cut = N_cut
 
         bath = BosonicBath(coup_op, ckAR, vkAR, ckAI, vkAI)
+
         self.bath = BathStates(bath.modes, self.N_cut)
         self.coup_op = [mode.Q for mode in bath.modes]
         self.spreQ = [spre(op).data for op in self.coup_op]
@@ -896,13 +917,10 @@ class FermionicHEOMSolver(object):
         self.L = _h_sys_liouvillian(self.H_sys, self.isHamiltonian)
         self.N_cut = N_cut
 
-        self.offsets = list(
-            itertools.accumulate((len(elem) for elem in ck), initial=0)
-        )
-
         bath = FermionicBath(coup_op, ck, vk)
         self.bath = BathStates(bath.modes, self.N_cut)
-        self.coup_op = [bath.modes[i].Q for i in self.offsets[:-1]]
+
+        self.coup_op = [mode.Q for mode in self.bath.modes]
         self.spreQ = [spre(op).data for op in self.coup_op]
         self.spostQ = [spost(op).data for op in self.coup_op]
         self.spreQdag = [spre(op.dag()).data for op in self.coup_op]
@@ -932,47 +950,43 @@ class FermionicHEOMSolver(object):
[…]
