+++ qutip/nonmarkov/bofin.py
@@ -397,79 +397,19 @@ def _h_sys_liouvillian(H_sys, isHamiltonian):
     return L.data
 
 
-class BosonicHEOMSolver(object):
+class HEOMSolver:
     """
-    This is a class for solvers that use the HEOM method for
-    calculating the dynamics evolution.
-    The method can compute open system dynamics without using any Markovian
-    or rotating wave approximations (RWA) for systems where the bath
-    correlations can be approximated to a sum of complex exponentials.
-    The method builds a matrix of linked differential equations, which are
-    then solved used the same ODE solvers as other qutip solvers
-    (e.g. mesolve)
-
-    Attributes
-    ----------
-    H_sys : Qobj or list
-        System Hamiltonian
-        Or
-        Liouvillian
-        Or
-        QobjEvo
-        Or
-        list of Hamiltonians with time dependence
-
-        Format for input (if list):
-        [time_independent_part, [H1, time_dep_function1],
-        [H2, time_dep_function2]]
-
-    coup_op : Qobj or list
-        Operator describing the coupling between system and bath.
-        Could also be a list of operators, which needs to be the same length
-        as ck's and vk's.
-
-    ckAR, ckAI, vkAR, vkAI : lists
-        Lists containing coefficients for fitting spectral density correlation
-
-        # Expand bath correlation function as prefactors times exponentials.
-        # ck - prefactors
-        # vk - frequency of the exponential
-        #
-        # ck * exp (-vk * t)
-        #
-        # ckAR * exp (-vkR * t) + i * ckAI * exp (-vkI * t)
-
-    N_cut : int
-        # The number of auxiliary density operators (ADO) to retain within the
-        # hierarchy.
-        #
-        # For each exponential in the speci ADO
-        #
-        # Cutoff parameter for the bath
-        #
-        # Number of operators kept in the hierarhcy.
-        #
-        # For each exponential we have a certain number of terms in the
-        # hierarchy.
-        #
-        # Total number of exponentials kept in the hierarchy.
-        #
-        # A bit like having total number of shared excitations.
-
-    options : :class:`qutip.solver.Options`
-        Generic solver options.
-        If set to None the default options will be used
+    HEOM solver that supports a single bath which may be either bosonic or
+    fermionic.
     """
 
-    def __init__(
-        self, H_sys, coup_op, ckAR, ckAI, vkAR, vkAI, N_cut, options=None
-    ):
+    def __init__(self, H_sys, bath, N_cut, options=None):
         self.H_sys = _convert_h_sys(H_sys)
         self.options = Options() if options is None else options
         self.isTimeDep = isinstance(self.H_sys, QobjEvo)
         self.isHamiltonian = _h_sys_is_hamiltonian(self.H_sys)
-        self._sup_dim = _h_sup_dim(self.H_sys, self.isHamiltonian)
         self._sys_dim = _h_sys_dim(self.H_sys, self.isHamiltonian)
+        self._sup_dim = _h_sup_dim(self.H_sys, self.isHamiltonian)
         self._h_sys_dims = (
             self.H_sys.dims if not self.isTimeDep
             else self.H_sys.to_list()[0].dims
@@ -477,12 +417,13 @@ class BosonicHEOMSolver(object):
         self.L = _h_sys_liouvillian(self.H_sys, self.isHamiltonian)
         self.N_cut = N_cut
 
-        bath = BosonicBath(coup_op, ckAR, vkAR, ckAI, vkAI)
-
         self.bath = BathStates(bath.modes, self.N_cut)
[…]
