### 2021-09-25 eae1f6f
Add new color_style argument to hinton() (#1595).
