### 2021-09-25 bf133e2
Update qutip/steadystate.py
Co-authored-by: Eric Giguère <eric.giguere@calculquebec.ca>
