+++ qutip/steadystate.py
@@ -960,9 +960,11 @@ def steadystate_floquet(H_0, c_ops, Op_t, w_d=1.0, n_it=3, sparse=False):
         L_0 = liouvillian(H_0, c_ops).data.tocsc()
         L_t = liouvillian(Op_t)
         L_p = (0.5 * L_t).data.tocsc()
+        # L_p and L_m correspond to the positive and negative frequency terms respectively.
+        # They are independent in the model, so we keep both names.
         L_m = L_p
         L_p_array = L_p.todense()
-        L_m_array = L_m.todense()
+        L_m_array = L_p_array
 
         Id = sp.eye(N ** 2, format="csc", dtype=np.complex128)
         S = T = sp.csc_matrix((N ** 2, N ** 2), dtype=np.complex128)
