### 2021-09-25 cb99936
Update qutip/steadystate.py
Co-authored-by: Eric Giguère <eric.giguere@calculquebec.ca>
