### 2021-09-25 a82ec68
Address code climate issues.
