### 2021-09-26 0ab0386
solve some pep8 issues
