### 2021-09-27 c108f6d
Rename BathMode to BathExponent for clarity.
