### 2021-09-27 199b74d
Document BathExponent.
