### 2021-09-27 a7fa19b
Document Bath and BathStates.
