### 2021-09-27 ebfa7a4
Add optional inplace RHS creation.
