### 2021-09-27 beb5d26
proper nan support
