### 2021-09-28 c7b8539
Removed sparse keyword for trace_csr.
This also lead to merging _trace_csr into trace_csr.
