### 2021-09-28 7a90dbd
Changed tolerance for sparse norm.
