+++ qutip/core/expect.py
@@ -101,11 +101,12 @@ def _single_qobj_expect(oper, state):
         raise ValueError(msg)
     out = _data.expect(oper.data, state.data)
 
-    # This ensures that expect can return something that is not a number.
-    try:
-        return out.real if oper.isherm and (state.isket or state.isherm) else out
-    except AttributeError:
-        return out
+    # This ensures that expect can return something that is not a number such
+    # as a `tensorflow.Tensor` in qutip-tensorflow.
+    return out.real if (oper.isherm
+                        and (state.isket or state.isherm)
+                        and hasattr(out, "real")
+                        ) else out
 
 
 def variance(oper, state):
