+++ qutip/nonmarkov/bofin.py
@@ -26,6 +26,12 @@ from qutip.solver import Options, Result
 from qutip.cy.spconvert import dense2D_to_fastcsr_fmode
 from qutip.ui.progressbar import BaseProgressBar
 
+# Load MKL spsolve if avaiable
+if settings.has_mkl:
+    from qutip._mkl.spsolve import mkl_spsolve
+else:
+    mkl_spsolve = None
+
 
 class BathExponent:
     """
@@ -435,50 +441,6 @@ def _convert_bath_exponents_fermionic(ck, vk):
     return ck, vk
 
 
-def _h_sys_is_hamiltonian(H_sys):
-    """ Return True if H_sys is a Hamiltonian and False if it is a Liouvillian.
-    """
-    if type(H_sys) is QobjEvo:
-        H_sys_list = H_sys.to_list()
-        H_sys = H_sys_list[0]
-    return H_sys.type == "oper"
-
-
-def _h_sys_dim(H_sys, isHamiltonian):
-    """ Return the system dimension for the given system.
-    """
-    if type(H_sys) is QobjEvo:
-        H_sys_list = H_sys.to_list()
-        H_shape = H_sys_list[0].shape
-    else:
-        H_shape = H_sys.shape
-
-    H_sys_dim = H_shape[0] if isHamiltonian else int(np.sqrt(H_shape[0]))
-    return H_sys_dim
-
-
-def _h_sup_dim(H_sys, isHamiltonian):
-    """ Return the super operator dimension for the given system.
-    """
-    if type(H_sys) is QobjEvo:
-        H_sys_list = H_sys.to_list()
-        H_shape = H_sys_list[0].shape
-    else:
-        H_shape = H_sys.shape
-    H_sup_dim = H_shape[0] ** 2 if isHamiltonian else H_shape[0]
-    return H_sup_dim
-
-
-def _h_sys_liouvillian(H_sys, isHamiltonian):
-    """ Return the Liouvillian dimension for the given system.
-    """
-    if type(H_sys) is QobjEvo:
-        H_sys_list = H_sys.to_list()
-        H_sys = H_sys_list[0]
-    L = liouvillian(H_sys) if isHamiltonian else H_sys
-    return L.data
-
-
 class HEOMSolver:
     """
     HEOM solver that supports a single bath which may be either bosonic or
@@ -488,18 +450,18 @@ class HEOMSolver:
     def __init__(self, H_sys, bath, N_cut, options=None):
         self.H_sys = _convert_h_sys(H_sys)
         self.options = Options() if options is None else options
-        self.isTimeDep = isinstance(self.H_sys, QobjEvo)
-        self.isHamiltonian = _h_sys_is_hamiltonian(self.H_sys)
-        self._sys_dim = _h_sys_dim(self.H_sys, self.isHamiltonian)
-        self._sup_dim = _h_sup_dim(self.H_sys, self.isHamiltonian)
-        self._h_sys_dims = (
-            self.H_sys.dims if not self.isTimeDep
-            else self.H_sys.to_list()[0].dims
-        )
-        self.L = _h_sys_liouvillian(self.H_sys, self.isHamiltonian)
-        self.N_cut = N_cut
+        self.is_timedep = isinstance(self.H_sys, QobjEvo)
+        self.H0 = self.H_sys.to_list()[0] if self.is_timedep else self.H_sys
+        self.is_hamiltonian = self.H0.type == "oper"
+        self.L0 = liouvillian(self.H0) if self.is_hamiltonian else self.H0
 
-        self.bath = BathStates(bath.modes, self.N_cut)
+        self._sys_shape = (
+            self.H0.shape[0] if self.is_hamiltonian
+            else int(np.sqrt(self.H0.shape[0]))
+        )
[…]
