### 2021-09-29 2aa1d9b
Combine fermionic and bosonic tests.
