+++ qutip/tests/test_bofin.py
@@ -1,0 +1,435 @@
+"""
+Tests for the Bosonic HEOM solvers.
+"""
+
+import numpy as np
+import pytest
+from numpy.linalg import eigvalsh
+from scipy.integrate import quad
+
+from qutip import (
+    Qobj, QobjEvo, sigmaz, sigmax, basis, destroy, expect, Options
+)
+from qutip.nonmarkov.bofin import (
+    _convert_h_sys,
+    BathExponent,
+    BathStates,
+    BosonicHEOMSolver,
+    FermionicHEOMSolver,
+    HSolverDL,
+)
+from qutip.states import enr_state_dictionaries
+
+
+class TestBathStates:
+    def mk_modes(self, dims):
+        return [
+            BathExponent("I", dim, Q=None, ck=1.0, vk=2.0) for dim in dims
+        ]
+
+    def test_create(self):
+        bath = BathStates(self.mk_modes([2, 3]), cutoff=2)
+        assert bath.dims == [2, 3]
+        assert bath.cutoff == 2
+        assert bath.states == [
+            (0, 0), (0, 1), (0, 2), (1, 0), (1, 1),
+        ]
+        assert bath.n_states == 5
+
+    def test_state_idx(self):
+        bath = BathStates(self.mk_modes([2, 3]), cutoff=2)
+        assert bath.idx((0, 0)) == 0
+        assert bath.idx((0, 1)) == 1
+        assert bath.idx((0, 2)) == 2
+        assert bath.idx((1, 0)) == 3
+        assert bath.idx((1, 1)) == 4
+
+    def test_next(self):
+        bath = BathStates(self.mk_modes([2, 3]), cutoff=2)
+        assert bath.next((0, 0), 0) == (1, 0)
+        assert bath.next((0, 0), 1) == (0, 1)
+        assert bath.next((1, 0), 0) is None
+        assert bath.next((1, 0), 1) == (1, 1)
+        assert bath.next((1, 1), 1) is None
+
+    def test_prev(self):
+        bath = BathStates(self.mk_modes([2, 3]), cutoff=2)
+        assert bath.prev((0, 0), 0) is None
+        assert bath.prev((0, 0), 1) is None
+        assert bath.prev((1, 0), 0) == (0, 0)
+        assert bath.prev((0, 1), 1) == (0, 0)
+        assert bath.prev((1, 1), 1) == (1, 0)
+        assert bath.prev((0, 2), 1) == (0, 1)
+
+
+def test_convert_h_sys():
+    """Tests the function for checking system Hamiltonian"""
+    _convert_h_sys(sigmax())
+    _convert_h_sys([sigmax(), sigmaz()])
+    _convert_h_sys([[sigmax(), np.sin], [sigmaz(), np.cos]])
+    _convert_h_sys([[sigmax(), np.sin], [sigmaz(), np.cos]])
+    _convert_h_sys(QobjEvo([sigmaz(), sigmax(), sigmaz()]))
+
+    with pytest.raises(TypeError) as err:
+        _convert_h_sys(sigmax().full())
+    assert str(err.value) == (
+        "Hamiltonian (H_sys) has unsupported type: <class 'numpy.ndarray'>"
+    )
+
+    with pytest.raises(ValueError) as err:
+        _convert_h_sys([[1, 0], [0, 1]])
+    assert str(err.value) == (
+        "Hamiltonian (H_sys) of type list cannot be converted to QObjEvo"
+    )
+    assert isinstance(err.value.__cause__, TypeError)
+    assert str(err.value.__cause__) == "Incorrect Q_object specification"
+
+
+@pytest.mark.filterwarnings("ignore::scipy.integrate.IntegrationWarning")
+@pytest.mark.parametrize(['bnd_cut_approx', 'tol', 'fake_timedep'], [
[…]

[…]
