### 2021-09-29 46b883b
Add tests for BathExponent.
