### 2021-09-29 9600a27
Add test for Bath creation.
