### 2021-09-29 7850c40
Test more BathStates attributes.
