### 2021-09-29 b0a7034
Add tests for _convert_coup_op.
