+++ qutip/tests/test_bofin.py
@@ -12,6 +12,7 @@ from qutip import (
 )
 from qutip.nonmarkov.bofin import (
     _convert_h_sys,
+    _convert_coup_op,
     BathExponent,
     Bath,
     BathStates,
@@ -145,27 +146,50 @@ class TestBathStates:
         assert bath.prev((0, 2), 1) == (0, 1)
 
 
-def test_convert_h_sys():
-    """Tests the function for checking system Hamiltonian"""
-    _convert_h_sys(sigmax())
-    _convert_h_sys([sigmax(), sigmaz()])
-    _convert_h_sys([[sigmax(), np.sin], [sigmaz(), np.cos]])
-    _convert_h_sys([[sigmax(), np.sin], [sigmaz(), np.cos]])
-    _convert_h_sys(QobjEvo([sigmaz(), sigmax(), sigmaz()]))
+class TestParameterConversionUtilities:
+    def test_convert_h_sys(self):
+        """Tests the function for checking system Hamiltonian"""
+        _convert_h_sys(sigmax())
+        _convert_h_sys([sigmax(), sigmaz()])
+        _convert_h_sys([[sigmax(), np.sin], [sigmaz(), np.cos]])
+        _convert_h_sys([[sigmax(), np.sin], [sigmaz(), np.cos]])
+        _convert_h_sys(QobjEvo([sigmaz(), sigmax(), sigmaz()]))
 
-    with pytest.raises(TypeError) as err:
-        _convert_h_sys(sigmax().full())
-    assert str(err.value) == (
-        "Hamiltonian (H_sys) has unsupported type: <class 'numpy.ndarray'>"
-    )
+        with pytest.raises(TypeError) as err:
+            _convert_h_sys(sigmax().full())
+        assert str(err.value) == (
+            "Hamiltonian (H_sys) has unsupported type: <class 'numpy.ndarray'>"
+        )
 
-    with pytest.raises(ValueError) as err:
-        _convert_h_sys([[1, 0], [0, 1]])
-    assert str(err.value) == (
-        "Hamiltonian (H_sys) of type list cannot be converted to QObjEvo"
-    )
-    assert isinstance(err.value.__cause__, TypeError)
-    assert str(err.value.__cause__) == "Incorrect Q_object specification"
+        with pytest.raises(ValueError) as err:
+            _convert_h_sys([[1, 0], [0, 1]])
+        assert str(err.value) == (
+            "Hamiltonian (H_sys) of type list cannot be converted to QObjEvo"
+        )
+        assert isinstance(err.value.__cause__, TypeError)
+        assert str(err.value.__cause__) == "Incorrect Q_object specification"
+
+    def test_convert_coup_op(self):
+        sx = sigmax()
+        sz = sigmaz()
+        assert _convert_coup_op(sx, 2) == [sx, sx]
+        assert _convert_coup_op([sx, sz], 2) == [sx, sz]
+
+        with pytest.raises(TypeError) as err:
+            _convert_coup_op(3, 1)
+        assert str(err.value) == (
+            "Coupling operator (coup_op) must be a Qobj or a list of Qobjs"
+        )
+
+        with pytest.raises(TypeError) as err:
+            _convert_coup_op([sx, 1], 2)
+        assert str(err.value) == (
+            "Coupling operator (coup_op) must be a Qobj or a list of Qobjs"
+        )
+
+        with pytest.raises(ValueError) as err:
+            _convert_coup_op([sx, sz], 3)
+        assert str(err.value) == "Expected 3 coupling operators"
 
 
 @pytest.mark.filterwarnings("ignore::scipy.integrate.IntegrationWarning")
