### 2021-09-29 d41cd6d
Move HEOMSolver tests into classes.
