+++ qutip/tests/test_bofin.py
@@ -192,351 +192,358 @@ class TestParameterConversionUtilities:
         assert str(err.value) == "Expected 3 coupling operators"
 
 
-@pytest.mark.filterwarnings("ignore::scipy.integrate.IntegrationWarning")
-@pytest.mark.parametrize(['bnd_cut_approx', 'tol', 'fake_timedep'], [
-    pytest.param(True, 1e-4, False, id="bnd_cut_approx_static"),
-    pytest.param(False,  1e-3, False, id="no_bnd_cut_approx_static"),
-    pytest.param(True, 1e-4, True, id="bnd_cut_approx_timedep"),
-    pytest.param(False,  1e-3, True, id="no_bnd_cut_approx_timedep"),
-])
-def test_pure_dephasing_model_HSolverDL(bnd_cut_approx, tol, fake_timedep):
-    """
-    HSolverDL: Compare with pure-dephasing analytical assert that the
-    analytical result and HEOM produce the same time dephasing evoltion.
-    """
-    cut_frequency = 0.05
-    coupling_strength = 0.025
-    lam_c = coupling_strength / np.pi
-    temperature = 1 / 0.95
-    times = np.linspace(0, 10, 21)
-
-    def _integrand(omega, t):
-        J = 2*lam_c * omega * cut_frequency / (omega**2 + cut_frequency**2)
-        return (-4 * J * (1 - np.cos(omega*t))
-                / (np.tanh(0.5*omega / temperature) * omega**2))
-
-    # Calculate the analytical results by numerical integration
-    expected = [0.5*np.exp(quad(_integrand, 0, np.inf, args=(t,))[0])
-                for t in times]
-
-    H_sys = Qobj(np.zeros((2, 2)))
-    if fake_timedep:
-        H_sys = [H_sys]
-    Q = sigmaz()
-    initial_state = 0.5*Qobj(np.ones((2, 2)))
-    projector = basis(2, 0) * basis(2, 1).dag()
-    options = Options(nsteps=15_000, store_states=True)
-
-    with pytest.warns(
-        UserWarning,
-        match=(
-            "Two similar real and imag exponents have been collated"
-            " automatically"
-        ),
-    ):
-        hsolver = HSolverDL(H_sys, Q, coupling_strength, temperature,
-                            14, 2, cut_frequency,
-                            bnd_cut_approx=bnd_cut_approx,
-                            options=options)
-
-    test = expect(hsolver.run(initial_state, times).states, projector)
-
-    np.testing.assert_allclose(test, expected, atol=tol)
-
-
-@pytest.mark.filterwarnings("ignore::scipy.integrate.IntegrationWarning")
-@pytest.mark.parametrize(['fake_timedep'], [
-    pytest.param(False, id="static"),
-    pytest.param(True, id="timedep"),
-])
-def test_pure_dephasing_model_BosonicHEOMSolver(fake_timedep):
-    """
-    BosonicHEOMSolver: Compare with pure-dephasing analytical assert that the
-    analytical result and HEOM produce the same time dephasing evoltion.
-    """
-    tol = 1e-3
-    gamma = 0.05
-    lam = 0.025
-    lam_c = lam / np.pi
-    T = 1 / 0.95
-    times = np.linspace(0, 10, 21)
-
-    def _integrand(omega, t):
-        J = 2*lam_c * omega * gamma / (omega**2 + gamma**2)
-        return (-4 * J * (1 - np.cos(omega*t))
-                / (np.tanh(0.5*omega / T) * omega**2))
-
-    # Calculate the analytical results by numerical integration
-    expected = [0.5*np.exp(quad(_integrand, 0, np.inf, args=(t,))[0])
[…]
