### 2021-09-29 df25c77
Factor out construction of steady-state L.
