### 2021-09-30 652e0dc
Add gather ops method for constructing the RHS.
