+++ qutip/nonmarkov/bofin.py
@@ -488,7 +488,7 @@ class HEOMSolver:
 
         self.progress_bar = BaseProgressBar()
 
-        self._configure_solver()
+        self._mode = "gather-ops"  # pad-op, inplace-add-op, gather-ops
 
     def _pad_op(self, op, row_he, col_he):
         """
@@ -510,6 +510,61 @@ class HEOMSolver:
         colpos = self.bath.idx(col_he) * block
         L[rowpos: rowpos + block, colpos: colpos + block] += op
 
+    def _gather_ops(self, ops, shape):
+        """ Create the HEOM liouvillian from a list of smaller CSR
+        """
+        # csr_matrix((data, indices, indptr), [shape=(M, N)])
+        #
+        # is the standard CSR representation where the column indices for row i
+        # are stored in indices[indptr[i]:indptr[i+1]] and their corresponding
+        # values are stored in data[indptr[i]:indptr[i+1]].
+        if not ops:
+            return sp.csr_matrix(shape, dtype=np.complex128)
+        ops = sorted(
+            (self.bath.idx(row_he), self.bath.idx(col_he), op)
+            for op, row_he, col_he in ops
+        )
+        block = self._sup_shape
+        nnz = sum(op.nnz for _, _, op in ops)
+        indptr = np.zeros(shape[0] + 1, dtype=np.int32)
+        indices = np.zeros(nnz, dtype=np.int32)
+        data = np.zeros(nnz, dtype=np.complex128)
+        end = 0
+        op_idx = 0
+        op_len = len(ops)
+
+        for row_idx in range(self.bath.n_states):
+            prev_op_idx = op_idx
+            while op_idx < op_len:
+                if ops[op_idx][0] != row_idx:
+                    break
+                op_idx += 1
+
+            row_ops = ops[prev_op_idx: op_idx]
+            rowpos = row_idx * block
+            for op_row in range(block):
+                for _, col_idx, op in row_ops:
+                    colpos = col_idx * block
+                    op_row_start = op.indptr[op_row]
+                    op_row_end = op.indptr[op_row + 1]
+                    op_row_len = op_row_end - op_row_start
+                    assert op_row_len == op[op_row, :].nnz
+                    if op_row_len == 0:
+                        continue
+                    indices[end: end + op_row_len] = (
+                        op.indices[op_row_start: op_row_end] + colpos
+                    )
+                    assert all(op.data[op_row_start: op_row_end] != 0)
+                    data[end: end + op_row_len] = (
+                        op.data[op_row_start: op_row_end]
+                    )
+                    assert all(data[end: end + op_row_len] != 0)
+                    end += op_row_len
+                indptr[rowpos + op_row + 1] = end
+        return sp.csr_matrix(
+            (data, indices, indptr), shape=shape, dtype=np.complex128,
+        )
+
     def _dsuper_list_td(self, t, y, L_list):
         """ Auxiliary function for the integration. Called every time step. """
         L = L_list[0][0]
@@ -524,22 +579,22 @@ class HEOMSolver:
         op = L - vk_sum * sp.eye(L.shape[0], dtype=complex, format="csr")
         return op
 
-    def _grad_prev(self, he_n, k, prev_he):
+    def _grad_prev(self, he_n, k):
         """ Get the previous gradient. """
         if self.bath.modes[k].type in (
                 BathExponent.types.R, BathExponent.types.I,
                 BathExponent.types.RI
         ):
-            return self._grad_prev_bosonic(he_n, k, prev_he)
+            return self._grad_prev_bosonic(he_n, k)
         elif self.bath.modes[k].type in (
[…]
