+++ qutip/utilities.py
@@ -36,8 +36,7 @@ This module contains utility functions that are commonly needed in other
 qutip modules.
 """
 
-__all__ = ['n_thermal', 'linspace_with', 'clebsch', 'convert_unit',
-           'view_methods']
+__all__ = ['n_thermal', 'clebsch', 'convert_unit']
 
 import numpy as np
 
@@ -76,39 +75,6 @@ def n_thermal(w, w_th):
             return 1.0 / (np.exp(w / w_th) - 1.0)
         else:
             return 0.0
-
-
-def linspace_with(start, stop, num=50, elems=[]):
-    """
-    Return an array of numbers sampled over specified interval
-    with additional elements added.
-
-    Returns `num` spaced array with elements from `elems` inserted
-    if not already included in set.
-
-    Returned sample array is not evenly spaced if addtional elements
-    are added.
-
-    Parameters
-    ----------
-    start : int
-        The starting value of the sequence.
-    stop : int
-        The stoping values of the sequence.
-    num : int, optional
-        Number of samples to generate.
-    elems : list/ndarray, optional
-        Requested elements to include in array
-
-    Returns
-    -------
-    samples : ndadrray
-        Original equally spaced sample array with additional
-        elements added.
-    """
-    elems = np.array(elems)
-    lspace = np.linspace(start, stop, num)
-    return np.union1d(lspace, elems)
 
 
 def _factorial_prod(N, arr):
@@ -389,32 +355,6 @@ def convert_mK_to_GHz(w):
     """
     w_GHz = w * 1.0e-12 * (_kB / _h)
     return w_GHz
-
-
-def view_methods(Q):
-    """
-    View the methods and corresponding doc strings
-    for a Qobj class.
-
-    Parameters
-    ----------
-    Q : Qobj
-        Input Quantum object.
-
-    """
-    meth = dir(Q)
-    qobj_props = ['data', 'dims', 'isherm', 'shape', 'type']
-    pub_meth = [x for x in meth if x.find('_') and x not in qobj_props]
-    ml = max([len(x) for x in pub_meth])
-    nl = len(Q.__class__.__name__ + 'Class Methods:')
-    print(Q.__class__.__name__ + ' Class Methods:')
-    print('-' * nl)
-    for ii in range(len(pub_meth)):
-        m = getattr(Q, pub_meth[ii])
-        meth_str = m.__doc__
-        ind = meth_str.find('\n')
-        pub_len = len(pub_meth[ii] + ': ')
-        print(pub_meth[ii] + ':' + ' ' * (ml+3-pub_len) + meth_str[:ind])
 
 
 def _version2int(version_string):
