+++ qutip/tests/test_utilities.py
@@ -32,106 +32,109 @@
 ###############################################################################
 
 import numpy as np
-from numpy.testing import assert_, run_module_suite
+from qutip import convert_unit, clebsch, n_thermal
+import qutip.utilities as utils
+from functools import partial
+import pytest
 
-from qutip import convert_unit, clebsch
+
+@pytest.mark.parametrize(['w', 'w_th', 'expected'], [
+    pytest.param(np.log(2), 1, 1, id='log(2)'),
+    pytest.param(np.log(2)*5, 5, 1, id='5*log(2)'),
+    pytest.param(0, 1, 0, id="0_energy"),
+    pytest.param(1, -1, 0, id="neg_temp"),
+    pytest.param(np.array([np.log(2), np.log(3), np.log(4)]), 1,
+                 np.array([1, 1/2, 1/3]), id="array"),
+])
+def test_n_thermal(w, w_th, expected):
+    np.testing.assert_allclose(n_thermal(w, w_th), expected)
 
 
-def test_unit_conversions():
-    "utilities: energy unit conversions"
+def _get_converter(orig, target):
+    """get funtion 'convert_{}_to_{}' when available for coverage """
+    try:
+        func = getattr(utils, f'convert_{orig}_to_{target}')
+    except AttributeError:
+        func = partial(convert_unit, orig=orig, to=target)
+    return func
 
+
+@pytest.mark.parametrize('orig', ["J", "eV", "meV", "GHz", "mK"])
+@pytest.mark.parametrize('target', ["J", "eV", "meV", "GHz", "mK"])
+def test_unit_conversions(orig, target):
     T = np.random.rand() * 100.0
+    T_converted = convert_unit(T, orig=orig, to=target)
+    T_back = convert_unit(T_converted, orig=target, to=orig)
 
-    diff = convert_unit(convert_unit(T, orig="mK", to="GHz"),
-                        orig="GHz", to="mK") - T
-    assert_(abs(diff) < 1e-6)
-    diff = convert_unit(convert_unit(T, orig="mK", to="meV"),
-                        orig="meV", to="mK") - T
-    assert_(abs(diff) < 1e-6)
+    assert T == pytest.approx(T_back)
 
-    diff = convert_unit(convert_unit(convert_unit(T, orig="mK", to="GHz"),
-                                     orig="GHz", to="meV"),
-                        orig="meV", to="mK") - T
-    assert_(abs(diff) < 1e-6)
+    T_converted = _get_converter(orig=orig, target=target)(T)
+    T_back = _get_converter(orig=target, target=orig)(T_converted)
 
-    w = np.random.rand() * 100.0
+    assert T == pytest.approx(T_back)
 
-    diff = convert_unit(convert_unit(w, orig="GHz", to="meV"),
-                        orig="meV", to="GHz") - w
-    assert_(abs(diff) < 1e-6)
 
-    diff = convert_unit(convert_unit(w, orig="GHz", to="mK"),
-                        orig="mK", to="GHz") - w
-    assert_(abs(diff) < 1e-6)
+@pytest.mark.parametrize('orig', ["J", "eV", "meV", "GHz", "mK"])
+@pytest.mark.parametrize('middle', ["J", "eV", "meV", "GHz", "mK"])
+@pytest.mark.parametrize('target', ["J", "eV", "meV", "GHz", "mK"])
+def test_unit_conversions_loop(orig, middle, target):
[…]
