### 2021-10-01 1aee4b2
Changed tol to 3.7.
I also changed the function in op_numpy due to a bug in numpy's nuclear
norm function where nan is returned for certain inputs.
