### 2021-10-01 939b2a6
Remove asserts from _gather_ops to make it MUCH faster.
