### 2021-10-01 983253e
Put back _configure_solver.
