### 2021-10-01 e2557cd
add tests for simdiags
