### 2021-10-01 9271f4c
add test for degen
