### 2021-10-01 67ad058
Add a test for degen
