### 2021-10-04 a9e5c88
Update qutip/tests/test_simdiag.py
Co-authored-by: Simon Cross <hodgestar+github@gmail.com>
