### 2021-10-04 7bfff46
Apply suggestions from code review
Co-authored-by: Simon Cross <hodgestar+github@gmail.com>
