### 2021-10-04 30ef581
Added the code to construct arcs on the bloch sphere and line segment connecting two points on the Bloch sphere
