+++ qutip/bloch.py
@@ -15,8 +15,9 @@ try:
     import matplotlib
     import matplotlib.pyplot as plt
     from mpl_toolkits.mplot3d import Axes3D
-    from matplotlib.patches import FancyArrowPatch
+    from matplotlib.patches import FancyArrowPatch, Arc
     from mpl_toolkits.mplot3d import proj3d
+    from scipy.spatial.transform import Rotation as R
 
     # Define a custom _axes3D function based on the matplotlib version.
     # The auto_add_to_figure keyword is new for matplotlib>=3.4.
@@ -177,6 +178,11 @@ class Bloch:
         self.savenum = 0
         # Style of points, 'm' for multiple colors, 's' for single color
         self.point_style = []
+        # Data for arcs
+        self.arcs_x = []
+        self.arcs_y= []
+        self.arcs_z = []
+        self.line_segment = []
 
         # status of rendering
         self._rendered = False
@@ -295,6 +301,10 @@ class Bloch:
         self.vectors = []
         self.point_style = []
         self.annotations = []
+        self.arcs_x = []
+        self.arcs_y= []
+        self.arcs_z = []
+        self.line_segment = []
 
     def add_points(self, points, meth='s'):
         """Add a list of data points to bloch sphere.
@@ -401,6 +411,178 @@ class Bloch:
                                  'text': text,
                                  'opts': kwargs})
 
+    def add_latitude(self, init_pt, fin_pt): 
+        # connects two points with an arc on a same latitude. Inputs have to be in sphereical coordinates.
+        # Used in the function add_arc
+        
+        if abs(init_pt[2]-fin_pt[2])> pi:
+            if init_pt[2] <= fin_pt[2]:
+                init_pt[2] = 2*pi + init_pt[2]
+            else:
+                fin_pt[2] = 2*pi + fin_pt[2]
+        
+        phi = linspace(init_pt[2],fin_pt[2], 200)
+        
+        r = init_pt[0]
+        theta = linspace(init_pt[1],init_pt[1],200)
+        
+        arc_x = r*sin(theta)*cos(phi)
+        arc_y = r*sin(theta)*sin(phi)
+        arc_z = r*cos(theta)
+        
+        self.arcs_x.append(arc_x)
+        self.arcs_y.append(arc_y)
+        self.arcs_z.append(arc_z)
+    
+    def add_longitude(self, init_pt, fin_pt): 
+        # connects two points with an arc on a same longitude. Inputs have to be in sphereical coordinates.
+        # Used in the function add_arc
+        
+        r = init_pt[0]
+        phi = linspace(fin_pt[2],fin_pt[2],200)
+        theta = linspace(init_pt[1],fin_pt[1], 200)
+        
+        arc_x = r*sin(theta)*cos(phi)
+        arc_y = r*sin(theta)*sin(phi)
+        arc_z = r*cos(theta)
+       
+        self.arcs_x.append(arc_x)
+        self.arcs_y.append(arc_y)
+        self.arcs_z.append(arc_z)
+        
+    def cart_to_sph(self, cart_cord):
+        # converts cartesian coordinates to spherical
+        r = 0
+        for i in cart_cord:
+            r += i**2
+        r = sqrt(r)
+        
+        if r == 0:
+            print("Polar and azimuthal points not defined for origin")
+            return
+        
+        theta = arccos(cart_cord[2]/r)
+        
+        if theta < 1e-12 or abs(theta - pi) < 1e-12:
+            phi = 0
+        else:
+            x = cart_cord[0]
+            y = cart_cord[1]
+            if x == 0:
+                if y > 0:
+                    phi = pi/2
+                else:
+                    phi = pi + pi/2
[…]
