### 2021-10-04 01c376d
make qutip_zvode private
