+++ qutip/bloch.py
@@ -3,7 +3,7 @@ __all__ = ['Bloch']
 import os
 
 from numpy import (ndarray, array, linspace, pi, outer, cos, sin, ones, size,
-                   sqrt, real, mod, append, ceil, arange, arccos, arctan,arcsin)
+                   sqrt, real, mod, append, ceil, arange, arccos, arctan, arcsin)
 
 from packaging.version import parse as parse_version
 
@@ -180,7 +180,7 @@ class Bloch:
         self.point_style = []
         # Data for arcs
         self.arcs_x = []
-        self.arcs_y= []
+        self.arcs_y = []
         self.arcs_z = []
         self.line_segment = []
 
@@ -302,7 +302,7 @@ class Bloch:
         self.point_style = []
         self.annotations = []
         self.arcs_x = []
-        self.arcs_y= []
+        self.arcs_y = []
         self.arcs_z = []
         self.line_segment = []
 
@@ -411,58 +411,58 @@ class Bloch:
                                  'text': text,
                                  'opts': kwargs})
 
-    def add_latitude(self, init_pt, fin_pt): 
-        # connects two points with an arc on a same latitude. Inputs have to be in sphereical coordinates.
-        # Used in the function add_arc
-        
-        if abs(init_pt[2]-fin_pt[2])> pi:
+    def add_latitude(self, init_pt, fin_pt):
+        # connects two points with an arc on a same latitude.
+        # Inputs have to be in spherical coordinates.
+        # Used in the function add_arc        
+        if abs(init_pt[2] - fin_pt[2]) > pi:
             if init_pt[2] <= fin_pt[2]:
                 init_pt[2] = 2*pi + init_pt[2]
             else:
                 fin_pt[2] = 2*pi + fin_pt[2]
-        
-        phi = linspace(init_pt[2],fin_pt[2], 200)
-        
+          
+        phi = linspace(init_pt[2], fin_pt[2], 200)
+
         r = init_pt[0]
-        theta = linspace(init_pt[1],init_pt[1],200)
-        
+        theta = linspace(init_pt[1], init_pt[1], 200)
+
         arc_x = r*sin(theta)*cos(phi)
         arc_y = r*sin(theta)*sin(phi)
         arc_z = r*cos(theta)
-        
+
         self.arcs_x.append(arc_x)
         self.arcs_y.append(arc_y)
         self.arcs_z.append(arc_z)
-    
-    def add_longitude(self, init_pt, fin_pt): 
-        # connects two points with an arc on a same longitude. Inputs have to be in sphereical coordinates.
+
+    def add_longitude(self, init_pt, fin_pt):
+        # connects two points with an arc on a same longitude.
+        # Inputs have to be in spherical coordinates.
         # Used in the function add_arc
-        
         r = init_pt[0]
-        phi = linspace(fin_pt[2],fin_pt[2],200)
-        theta = linspace(init_pt[1],fin_pt[1], 200)
-        
+        phi = linspace(fin_pt[2], fin_pt[2], 200)
+        theta = linspace(init_pt[1], fin_pt[1], 200)
+
         arc_x = r*sin(theta)*cos(phi)
         arc_y = r*sin(theta)*sin(phi)
         arc_z = r*cos(theta)
-       
+
         self.arcs_x.append(arc_x)
         self.arcs_y.append(arc_y)
         self.arcs_z.append(arc_z)
-        
+
     def cart_to_sph(self, cart_cord):
         # converts cartesian coordinates to spherical
         r = 0
         for i in cart_cord:
             r += i**2
         r = sqrt(r)
-        
+
         if r == 0:
             print("Polar and azimuthal points not defined for origin")
             return
[…]
