### 2021-10-05 95d9086
Apply suggestions from code review
Co-authored-by: Simon Cross <hodgestar+github@gmail.com>
