+++ qutip/tests/core/test_brtools.py
@@ -1,10 +1,10 @@
 import pytest
+import qutip
 import numpy as np
 from qutip.core._brtools import matmul_var_data, _EigenBasisTransform
 from qutip.core._brtensor import (_br_term_dense, _br_term_sparse,
-                                  _BlochRedfieldElement)
-from qutip.core.blochredfield import brterm, bloch_redfield_tensor
-import qutip
+                                  _br_term_data, _BlochRedfieldElement)
+
 
 def _make_rand_data(shape):
     np.random.seed(11)
@@ -18,6 +18,7 @@ transform = {
     2: qutip.data.conj,
     3: qutip.data.adjoint
 }
+
 
 @pytest.mark.parametrize('datatype', qutip.data.to.dtypes)
 @pytest.mark.parametrize('transleft', [0, 1, 2, 3],
@@ -42,7 +43,7 @@ def test_matmul_var(datatype, transleft, transright):
 @pytest.mark.parametrize('sparse', [False, True], ids=['Dense', 'Sparse'])
 def test_eigen_transform(sparse):
     a = qutip.destroy(5)
-    f = lambda t, _: t
+    f = lambda t: t
     op = qutip.QobjEvo([a*a.dag(), [a+a.dag(), f]])
     eigenT = _EigenBasisTransform(op, sparse=sparse)
     evecs_qevo = eigenT.as_Qobj()
@@ -134,11 +135,12 @@ def test_eigen_transform_super_ops():
     np.testing.assert_allclose(computed, expected, rtol=1e-14, atol=1e-14)
 
 
-@pytest.mark.parametrize('func', [_br_term_dense, _br_term_sparse],
-                         ids=['dense', 'sparse'])
-def test_brterm_linbblad_comp(func):
+@pytest.mark.parametrize('func',
+                         [_br_term_dense, _br_term_sparse, _br_term_data],
+                         ids=['dense', 'sparse', 'data'])
+def test_br_term_linbblad_comp(func):
     N = 5
-    a = qutip.destroy(N) + qutip.destroy(N)**2/2
+    a = qutip.destroy(N) + qutip.destroy(N)**2 / 2
     A_op = a + a.dag()
     H = qutip.num(N)
     diag = H.eigenenergies()
@@ -155,7 +157,7 @@ def test_brterm_linbblad_comp(func):
     pytest.param(lambda skew: np.ones_like(skew), id='no_filter'),
     pytest.param(lambda skew: np.exp(skew)/10, id='smooth_filter'),
 ])
-def test_brterm(cutoff, spectra):
+def test_br_term(cutoff, spectra):
     N = 5
     a = qutip.destroy(N) @ qutip.coherent_dm(N, 0.5) * 0.5
     A_op = a + a.dag()
@@ -165,141 +167,6 @@ def test_brterm(cutoff, spectra):
     spectrum = spectra(skew)
     R_dense = _br_term_dense(A_op.data, spectrum, skew, cutoff).to_array()
     R_sparse = _br_term_sparse(A_op.data, spectrum, skew, cutoff).to_array()
+    R_data = _br_term_data(A_op.data, spectrum, skew, cutoff).to_array()
     np.testing.assert_allclose(R_dense, R_sparse, rtol=1e-14, atol=1e-14)
-
-
-@pytest.mark.parametrize('cutoff', [0, 0.1, 1, 3, np.inf])
-def test_brterm(cutoff):
-    N = 5
-    H = qutip.num(N)
-    a = qutip.destroy(N)
-    A_op = a + a.dag()
-    spectra = qutip.coefficient("(w>0)*0.5", args={'w':0})
[…]
