+++ qutip/core/_brtools.pyx
@@ -12,7 +12,6 @@ from qutip.core.cy.coefficient cimport Coefficient
 from qutip.core.data cimport Data, Dense, idxint
 import qutip.core.data as _data
 from qutip.core import Qobj
-from qutip.core.data.eigen import eigs
 from scipy.linalg cimport cython_blas as blas
 
 
@@ -22,7 +21,7 @@ __all__ = ['SpectraCoefficient']
 cdef class SpectraCoefficient(Coefficient):
     """
     Change a Coefficient with `t` dependence to one with `w` dependence to use
-    in bloch redfield tensor. Allow array based coefficients to be used as
+    in Bloch-Redfield tensor to allow array based coefficients to be used as
     spectrum function.
     If 2 coefficients are passed, the first one is the frequence response and
     the second is the time response.
@@ -41,13 +40,17 @@ cdef class SpectraCoefficient(Coefficient):
         """Return a copy of the :obj:`Coefficient`."""
         return SpectraCoefficient(self.coeff_t, self.coeff_w, self.w)
 
-    def replace(self, *, _args=None, w=None, **kwargs):
-        if w is not None:
-            return SpectraCoefficient(self.coeff_t, self.coeff_w, w)
+    def replace(self, *, w=None, _args=None, **kwargs):
         if _args:
             kwargs.update(_args)
-        if 'w' in kwargs:
-            return SpectraCoefficient(self.coeff_t, self.coeff_w, kwargs['w'])
+        if kwargs:
+            return SpectraCoefficient(
+                self.coeff_w.replace(**kwargs),
+                self.coeff_t.replace(**kwargs) if self.coeff_t else None,
+                kwargs.get('w', w or 0.)
+              )
+        if w is not None:
+            return SpectraCoefficient(self.coeff_w, self.coeff_t, w)
         return self
 
 
@@ -57,6 +60,7 @@ cdef size_t _mul_checked(size_t a, size_t b) except? -1:
 
 
 cdef Data _apply_trans(Data original, int trans):
+    """helper function for matmul_var_data, apply transform."""
     cdef Data out
     if trans == 0:
         out = original
@@ -70,6 +74,7 @@ cdef Data _apply_trans(Data original, int trans):
 
 
 cdef char _fetch_trans_code(int trans):
+    """helper function for matmul_var_Dense, fetch the blas flag byte"""
     if trans == 0:
         return b'N'
     elif trans == 1:
@@ -110,7 +115,18 @@ cpdef Data matmul_var_data(Data left, Data right,
 
 cdef Dense matmul_var_Dense(Dense left, Dense right,
                             int transleft, int transright):
-    # The present implementation only support square matrices.
+    """
+    matmul which input matrices can be transposed or adjoint.
+    out = transleft(left) @ transright(right)
+
+    with trans[left, right]:
+       0 : Normal
+       1 : Transpose
+       2 : Conjugate
+       3 : Adjoint
+    """
+    # blas support matmul for normal, transpose, adjoint for fortran ordered
+    # matrices.
     if not (
         left.shape[0] == left.shape[1]
         and left.shape[0] == right.shape[0]
@@ -127,11 +143,12 @@ cdef Dense matmul_var_Dense(Dense left, Dense right,
 
     # blas.zgemm does not support adjoint.
     if tleft + tright == 5:
-        # adjoint and conjugate, we can't use transpose to use zgemm
+        # adjoint and conjugate, we can't transpose the output to use zgemm
         out = matmul_var_data(left, right, transleft-2, transright-2)
         return out.conj()
     if tleft == 2 or tright == 2:
-        # A.dag @ B^op -> (B^T^op @ A.conj)^T
[…]
