+++ qutip/core/blochredfield.py
@@ -42,47 +42,9 @@ from ._brtensor import _BlochRedfieldElement
 __all__ = ['bloch_redfield_tensor', 'brterm']
 
 
-def _read_legacy_a_ops(a_op):
-    constant = coefficient("1")
-    parsed_a_op
-    op, spec = a_op
-    if isinstance(op, Qobj):
-        if isinstance(spec, str):
-            parsed_a_op = (op, coefficient(spec, args={'w':0}))
-        elif callable(spec):
-            parsed_a_op =(op, coefficient(lambda t, args: spec(args['w'])))
-        elif isinstance(spec, tuple):
-            if isinstance(spec[0], str):
-                freq_responce = coefficient(spec[0], args={'w':0})
-            elif isinstance(spec[0], Cubic_Spline):
-                freq_responce = SpectraCoefficient(coefficient(spec[0]))
-            else:
-                raise Exception('Invalid bath-coupling specification.')
-            if isinstance(spec[1], str):
-                time_responce = coefficient(spec[1], args={'w':0})
-            elif isinstance(spec[1], Cubic_Spline):
-                time_responce = coefficient(spec[1])
-            else:
-                raise Exception('Invalid bath-coupling specification.')
-            parsed_a_op = (op, freq_coeff * time_coeff)
-    elif isinstance(op, tuple):
-        qobj1, qobj2 = op
-        if isinstance(spec[0], str):
-            freq_responce = coefficient(spec[0], args={'w':0})
-        elif isinstance(spec[0], Cubic_Spline):
-            freq_responce = SpectraCoefficient(coefficient(spec[0]))
-        else:
-            raise Exception('Invalid bath-coupling specification.')
-        parsed_a_op =  (
-            QobjEvo([[qobj1, spec[1]], [qobj1, spec[2]]]),
-            freq_responce
-        )
-    return parsed_a_op
-
-
-def bloch_redfield_tensor(H, a_ops, c_ops=[],
-                          use_secular=True, sec_cutoff=0.1,
-                          fock_basis=False, legacy_a_ops=False, sparse=False):
+def bloch_redfield_tensor(H, a_ops, c_ops=[], use_secular=True, sec_cutoff=0.1,
+                          fock_basis=False, sparse_eigensolver=False,
+                          br_dtype='sparse'):
     """
     Calculates the Bloch-Redfield tensor for a system given
     a set of operators and corresponding spectral functions that describes the
@@ -105,9 +67,10 @@ def bloch_redfield_tensor(H, a_ops, c_ops=[],
             The corresponding bath spectra.
             Can be a `Coefficient` using an 'w' args or a function of the
             frequence. The `SpectraCoefficient` can be used to use array based
-            coefficient.
-            Example:
+            coefficient. They can also depend on ``t`` if the corresponding
+            ``a_op`` is a :cls:`QobjEvo`.
 
+        Example:
             a_ops = [
                 (a+a.dag(), coefficient('w>0', args={"w": 0})),
                 (QobjEvo([b+b.dag(), f(t)]), g(w)),
@@ -128,15 +91,15 @@ def bloch_redfield_tensor(H, a_ops, c_ops=[],
         Whether to return the tensor in the input basis or the diagonalized
         basis.
 
-    legacy_a_ops : bool {False}
-        Whether to use the v4's brmesolve's a_ops specification.
+    sparse_eigensolver : bool {False}
+        Whether to use the sparse eigensolver
 
-    sparse : bool {False}
-        Whether to use the sparse eigensolver if
[…]

[…]
