### 2021-10-07 9e6970c
Fix tests and optimize
