+++ qutip/core/blochredfield.py
@@ -1,35 +1,3 @@
-# This file is part of QuTiP: Quantum Toolbox in Python.
-#
-#    Copyright (c) 2011 and later, QuSTaR.
-#    All rights reserved.
-#
-#    Redistribution and use in source and binary forms, with or without
-#    modification, are permitted provided that the following conditions are
-#    met:
-#
-#    1. Redistributions of source code must retain the above copyright notice,
-#       this list of conditions and the following disclaimer.
-#
-#    2. Redistributions in binary form must reproduce the above copyright
-#       notice, this list of conditions and the following disclaimer in the
-#       documentation and/or other materials provided with the distribution.
-#
-#    3. Neither the name of the QuTiP: Quantum Toolbox in Python nor the names
-#       of its contributors may be used to endorse or promote products derived
-#       from this software without specific prior written permission.
-#
-#    THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-#    "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-#    LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
-#    PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
-#    HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
-#    SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
-#    LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
-#    DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
-#    THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-#    (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-#    OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-###############################################################################
 import os
 import numpy as np
 from qutip.settings import settings as qset
@@ -107,7 +75,7 @@ def bloch_redfield_tensor(H, a_ops, c_ops=[], use_secular=True, sec_cutoff=0.1,
         column.
     """
     R = liouvillian(H, c_ops)
-    H_transform = _EigenBasisTransform(QobjEvo(H), sparse)
+    H_transform = _EigenBasisTransform(QobjEvo(H), sparse_eigensolver)
 
     if fock_basis:
         for a_op in a_ops:
@@ -185,12 +153,16 @@ def brterm(H, a_op, spectra, use_secular=True, sec_cutoff=0.1,
     if isinstance(H, _EigenBasisTransform):
         Hdiag = H
     else:
-        Hdiag = _EigenBasisTransform(QobjEvo(H), sparse=sparse)
+        Hdiag = _EigenBasisTransform(QobjEvo(H), sparse=sparse_eigensolver)
 
     sec_cutoff = sec_cutoff if use_secular else np.inf
     R = QobjEvo(_BlochRedfieldElement(Hdiag, QobjEvo(a_op), spectra,
                 sec_cutoff, not fock_basis, dtype=br_dtype))
 
-    if isinstance(H, Qobj) and isinstance(a_op, Qobj):
+    if (
+        ((isinstance(H, _EigenBasisTransform) and H.isconstant)
[…]
