+++ qutip/core/_brtensor.pyx
@@ -26,14 +26,17 @@ cpdef enum TensorType:
 @cython.wraparound(False)
 cpdef Data _br_term_data(Data A, double[:, ::1] spectrum,
                          double[:, ::1] skew, double cutoff):
-    """ Compute the contribution of A to the Bloch Redfield tensor."""
+    """
+    Compute the contribution of A to the Bloch Redfield tensor.
+    Computation are done using dispatched function.
+    """
     cdef object cutoff_arr
-    cdef int nrow = A.shape[0], a, b, c, d
+    cdef int nrows = A.shape[0], a, b, c, d
     cdef Data S, I, AS, AST, out, C
     cdef type cls = type(A)
 
     S = _data.to(cls, _data.mul(_data.Dense(spectrum, copy=False), 0.5))
-    I = _data.identity[cls](nrow)
+    I = _data.identity[cls](nrows)
     AS = _data.multiply(A, S)
     AST = _data.multiply(A, _data.transpose(S))
 
@@ -48,13 +51,15 @@ cpdef Data _br_term_data(Data A, double[:, ::1] spectrum,
     # The cutoff_arr should be sparse most of the time, but it depend on the
     # cutoff and we cannot easily guess a nnz to make it efficiently...
     # But there is probably room from improvement.
-    cutoff_arr = np.zeros((nrow*nrow, nrow*nrow), dtype=np.complex128)
-    for a in range(nrow):
-        for b in range(nrow):
-            for c in range(nrow):
-                for d in range(nrow):
+    cutoff_arr = np.zeros((nrows*nrows, nrows*nrows), dtype=np.complex128)
+    # skew[a,b] = w[a] - w[b]
+    # cutoff_arr[a,b,c,d] = (w[a] - w[b] - w[c] + w[d]) < cutoff
+    for a in range(nrows):
+        for b in range(nrows):
+            for c in range(nrows):
+                for d in range(nrows):
                     if fabs(skew[a, b] - skew[c, d]) < cutoff:
-                        cutoff_arr[a * nrow + b, c * nrow + d] = 1.
+                        cutoff_arr[a * nrows + b, c * nrows + d] = 1.
     C = _data.to(cls, _data.Dense(cutoff_arr, copy=False))
     return _data.multiply(out, C)
 
@@ -63,6 +68,10 @@ cpdef Data _br_term_data(Data A, double[:, ::1] spectrum,
 @cython.wraparound(False)
 cpdef Dense _br_term_dense(Data A, double[:, ::1] spectrum,
                            double[:, ::1] skew, double cutoff):
+    """
+    Compute the contribution of A to the Bloch Redfield tensor.
+    Allocate a Dense array and fill it.
+    """
     cdef size_t nrows = A.shape[0]
     cdef size_t a, b, c, d, k # matrix indexing variables
     cdef double complex elem
@@ -85,28 +94,24 @@ cpdef Dense _br_term_dense(Data A, double[:, ::1] spectrum,
 
     # TODO: we could use openmp to speed up.
     for a in range(nrows):
-        for b in range(nrows):
+        for b in range(nrows-1, -1, -1):
             if fabs(skew[a, b]) < cutoff:
                 for k in range(nrows):
                     ac_term[a, b] += A_mat[a, k] * A_mat[k, b] * spectrum[a, k]
                     bd_term[a, b] += A_mat[a, k] * A_mat[k, b] * spectrum[b, k]
 
     for a in range(nrows):
-        for b in range(nrows):
-            for c in range(nrows):
+        for b in range(nrows-1, -1, -1):
+            for c in range(nrows-1, -1, -1):
                 for d in range(nrows):
-                    elem = 0.
[…]

[…]
