### 2021-10-12 2e8d9a6
Changed None to string in test.
None is a common object that could potentially be returned by other means.

Co-authored-by: Simon Cross <hodgestar+github@gmail.com>
