### 2021-10-12 8671d15
fix degen bug for not properly orthogonal vectors
