### 2021-10-12 5d3dbd4
Solver usable by itself
