### 2021-10-12 7d53917
Fix for instrument construction with manually specified outcome labels.
