### 2021-10-13 4456bea
More efficient creation of the hadamard matrix.
For loops are slow.
