### 2021-10-13 ede4a5b
Update qutip/qip/operations/gates.py
Co-authored-by: Simon Cross <hodgestar+github@gmail.com>
