### 2021-10-13 311ba24
Update qutip/qip/operations/gates.py
Co-authored-by: Simon Cross <hodgestar+github@gmail.com>
