### 2021-10-13 5642874
Merge pull request #1688 from AGaliciaMartinez/hadamard_transform_efficient
Improve Hadamard transform efficiency by up to ~70x.
