### 2021-10-13 22ab4ba
Apply suggestions from code review
Co-authored-by: Jake Lishman <jake@binhbar.com>
