### 2021-10-13 36ceb79
copy options before passing to solvers
