### 2021-10-14 96077d4
Consistently use qutip.fastsparse fast_csr_matrix (Qobjs already use them everywhere).
