### 2021-10-14 c6aca10
Use Nk exponents in bofin.HSolverDL to match heom.HSolverDL use of Nk.
