### 2021-10-14 6ba9885
Select gather_ops as the RHS construction strategy.
