+++ qutip/nonmarkov/bofin.py
@@ -20,7 +20,6 @@ from qutip import state_number_enumerate
 from qutip.qobj import Qobj
 from qutip.qobjevo import QobjEvo
 from qutip.superoperator import liouvillian, spre, spost, vec2mat
-from qutip.cy.heom import cy_pad_csr
 from qutip.cy.spmatfuncs import cy_ode_rhs
 from qutip.solver import Options, Result
 from qutip.cy.spconvert import dense2D_to_fastcsr_fmode
@@ -497,81 +496,7 @@ class HEOMSolver:
 
         self.progress_bar = BaseProgressBar()
 
-        self._mode = "gather-ops"  # pad-op, inplace-add-op, gather-ops
-
         self._configure_solver()
-
-    def _pad_op(self, op, row_he, col_he):
-        """
-        Pad op into its correct position within the larger HEOM liouvillian
-        for the given row and column bath states.
-        """
-        nhe = self.bath.n_states
-        rowidx = self.bath.idx(row_he)
-        colidx = self.bath.idx(col_he)
-        return cy_pad_csr(op, nhe, nhe, rowidx, colidx)
-
-    def _inplace_add_op(self, L, op, row_he, col_he):
-        """
-        Add the operation ``op`` to its correct position within the
-        larger HEOM liouvillian for the given row and column bath states.
-        """
-        block = self._sup_shape
-        rowpos = self.bath.idx(row_he) * block
-        colpos = self.bath.idx(col_he) * block
-        L[rowpos: rowpos + block, colpos: colpos + block] += op
-
-    def _gather_ops(self, ops, block, nhe):
-        """ Create the HEOM liouvillian from a list of smaller CSRs.
-
-            Parameters
-            ----------
-            ops : list of (row_idx, col_idx, op)
-                Operators to combine.
-            block : int
-                The size of a single Liovillian operator in the hierarchy.
-            nhe : int
-                The number of ADOs in the hierarchy.
-        """
-        shape = (block * nhe, block * nhe)
-        if not ops:
-            return sp.csr_matrix(shape, dtype=np.complex128)
-        nnz = sum(op.nnz for _, _, op in ops)
-        indptr = np.zeros(shape[0] + 1, dtype=np.int32)
-        indices = np.zeros(nnz, dtype=np.int32)
-        data = np.zeros(nnz, dtype=np.complex128)
-        end = 0
-        op_idx = 0
-        op_len = len(ops)
-
-        for row_idx in range(nhe):
-            prev_op_idx = op_idx
-            while op_idx < op_len:
-                if ops[op_idx][0] != row_idx:
-                    break
-                op_idx += 1
-
-            row_ops = ops[prev_op_idx: op_idx]
-            rowpos = row_idx * block
-            for op_row in range(block):
-                for _, col_idx, op in row_ops:
-                    colpos = col_idx * block
-                    op_row_start = op.indptr[op_row]
-                    op_row_end = op.indptr[op_row + 1]
-                    op_row_len = op_row_end - op_row_start
-                    if op_row_len == 0:
-                        continue
-                    indices[end: end + op_row_len] = (
-                        op.indices[op_row_start: op_row_end] + colpos
-                    )
-                    data[end: end + op_row_len] = (
-                        op.data[op_row_start: op_row_end]
-                    )
-                    end += op_row_len
-                indptr[rowpos + op_row + 1] = end
-        return fast_csr_matrix(
-            (data, indices, indptr), shape=shape, dtype=np.complex128,
-        )
 
     def _dsuper_list_td(self, t, y, L_list):
         """ Auxiliary function for the integration. Called every time step. """
@@ -675,62 +600,22 @@ class HEOMSolver:
     def _rhs(self, L):
[…]
