### 2021-10-14 b5f5186
Add a DrudeLorentz bath (which exposes the Matsubara parameters and terminator nicely while simultaneously making HSolverDL much simpler).
