### 2021-10-14 a8b8786
Initial draft of a DrudeLorentzPadeBath.
