### 2021-10-17 aa49adf
Added feature to add arcs and lines
