### 2021-10-18 4d7802a
Description of add_arc added. Fixed whitespces.
