### 2021-10-18 fe05f1b
Update qutip/bloch.py
Co-authored-by: Asier Galicia <57414022+AGaliciaMartinez@users.noreply.github.com>
