### 2021-10-18 a28b2cf
Update qutip/bloch.py
Co-authored-by: Asier Galicia <57414022+AGaliciaMartinez@users.noreply.github.com>
