### 2021-10-18 06082d3
Update qutip/bloch.py
Co-authored-by: Asier Galicia <57414022+AGaliciaMartinez@users.noreply.github.com>
