### 2021-10-18 3fb98af
Update qutip/bloch.py
Co-authored-by: Asier Galicia <57414022+AGaliciaMartinez@users.noreply.github.com>
