### 2021-10-20 403bc69
Merge pull request #1692 from AGaliciaMartinez/fix#1691_allow_pow_with_scalars
Add support for computing powers of scalar (1x1) Qobjs. Fixes #1691.
