### 2021-10-20 ceb4f12
Fix using openmp flag on linux
