### 2021-10-20 2660e1f
Remove adding flags twice on mac
