### 2021-10-21 e77cfcf
Avoid modifying global kwargs so that test_driven_cavity can be run multiple times.
