### 2021-10-21 9d2b95b
Remove mutable default for steadystate c_op_list parameter.
