### 2021-10-21 733ef3c
Fix reference to method in _pseudo_inverse_sparse.
