### 2021-10-21 bc48b54
Only set method in pseudo_inverse if one is explicitly defined.
