### 2021-10-21 4ecff27
Try drop back to pytest-cov 2.12.1.
