### 2021-10-21 8815740
Revert pytest-cov version pinning.
