+++ .github/workflows/tests.yml
@@ -88,7 +88,7 @@ jobs:
             conda install nomkl numpy "scipy${{ matrix.scipy-requirement }}"
           fi
           python -m pip install -e .[$QUTIP_TARGET]
-          python -m pip install pytest-cov==2.12.1 coveralls
+          python -m pip install pytest-cov coveralls
 
       - name: Package information
         run: |
