### 2021-10-21 166cda5
Remove use of not thread-safe catch_warnings.
