### 2021-10-21 85ba43f
Revert "Remove mutable default for steadystate c_op_list parameter."
This reverts commit 9d2b95bd8d88daa2cf33a0def9c2fc23d1b2d303.
