### 2021-10-21 5a9d203
Make np_pt explicitly a int32 pointer (since we are calling the padiso 32-bit API).
