### 2021-10-21 e3e9388
Revert "Only set method in pseudo_inverse if one is explicitly defined."
This reverts commit bc48b54476945505eaeeb70234834869dacb44b1.
