+++ .github/workflows/tests.yml
@@ -107,7 +107,7 @@ jobs:
             # truly being executed.
             export QUTIP_NUM_PROCESSES=2
           fi
-          pytest -Werror --strict-config --strict-markers --durations=0 --durations-min=1.0 --verbosity=1 --cov=qutip --cov-report= --color=yes qutip/tests/test_steadystate.py
+          pytest -Werror --strict-config --strict-markers --durations=0 --durations-min=1.0 --verbosity=1 --cov=qutip --cov-report= --color=yes qutip/tests/test_steadystate.py qutip/tests/test_states.py
           # Above flags are:
           #  -Werror
           #     treat warnings as errors
