### 2021-10-21 bfe2e7e
Scatter assertions to hunt for the nans in coherent_dm.
