### 2021-10-21 10a1a24
Litter asserts on nans in sp_expm.
