### 2021-10-21 5d94526
Fix asset for diagonal case.
