+++ qutip/sparse.py
@@ -400,7 +400,12 @@ def sp_expm(A, sparse=False):
         assert not np.isnan(E.data).any()
     else:
         E = spla.expm(A.toarray())
-        assert not np.isnan(E).any()
+        if np.isnan(E).any():
+            print("A:", A)
+            print("A data:", A.indices, A.indptr, A.shape)
+            print("A toarray:", A.toarray())
+            print("E:", E)
+            raise RuntimeError("NaNs generated by sp_expm.")
     X = sp.csr_matrix(E)
     assert not np.isnan(X.data).any()
     return X
