### 2021-10-21 cf44e0d
Merge pull request #1695 from qutip/dependabot/pip/doc/babel-2.9.1
Bump documentation babel requirement from 2.9.0 to 2.9.1.
