+++ doc/requirements.txt
@@ -1,6 +1,6 @@
 alabaster==0.7.12
 appnope==0.1.2
-Babel==2.9.0
+Babel==2.9.1
 backcall==0.2.0
 certifi==2020.12.5
 chardet==4.0.0
