### 2021-10-21 dba3a24
Print dtypes of expm matrices.
