+++ qutip/sparse.py
@@ -399,11 +399,14 @@ def sp_expm(A, sparse=False):
         E = spla.expm(A.tocsc())
         assert not np.isnan(E.data).any()
     else:
-        E = spla.expm(A.toarray())
+        At = A.toarray()
+        E = spla.expm(At)
         if np.isnan(E).any():
             print("A:", A)
+            print("A dtype:", A.dtype)
             print("A data:", A.indices, A.indptr, A.shape)
-            print("A toarray:", A.toarray())
+            print("A toarray:", At)
+            print("A toarray dtype:", At.dtype)
             print("E:", E)
             raise RuntimeError("NaNs generated by sp_expm.")
     X = sp.csr_matrix(E)
