### 2021-10-21 5f4ddf7
Add 3.9 tests with nomkl and old numpy
Only 3.9 seems to fails so migrate all test there.
