### 2021-10-21 4b027c3
Repeat coherent_dm test on various sizes.
Tests sometime passes and sometime don't,
1) If it is to fail, I want it to always fail.
2) Are tries independent?
3) Does size matter?
