### 2021-10-21 baa51c7
Print more information about the vm
Tests seems to pass or fail together.
Maybe there is a conflict with some VM configuration or cpus. So I am storing cpu, ram, and distribution info.
