### 2021-10-21 a874ee2
remove mac tests
Distribution info fail, not pytest.
