### 2021-10-22 a2f788c
Add support for setting the numpy version in test runs.
