### 2021-10-22 fa90e5d
Pin base numpy to 1.21.0 in test runs.
