+++ .github/workflows/tests.yml
@@ -27,6 +27,7 @@ jobs:
         # matrix size; make sure to test all supported versions in some form.
         python-version: [3.9]
         case-name: [defaults]
+        numpy-requirement: ["==1.21.0"]
         # Extra special cases.  In these, the new variable defined should always
         # be a truth-y value (hence 'nomkl: 1' rather than 'mkl: 0'), because
         # the lack of a variable is _always_ false-y, and the defaults lack all
