### 2021-10-22 166ffc8
Add host environment information.
