### 2021-10-22 6c28e82
Don't call hostnamectl or lscpu on OS X.
