### 2021-10-22 cd4a90c
Remove redundant setting of numpy-requirement in OpenMP tests run.
