### 2021-10-22 9922162
Use the selected version of numpy in all cases.
