### 2021-10-22 bc1b285
Pin numpy requirement for OpenMP test case on Python 3.9.
