### 2021-10-22 818214c
Remove brackets around numpy-requirement.
