### 2021-10-24 f951c7b
Clear up logic for when to recreate the figure and axes in .render().
