### 2021-10-24 e2b53c3
Simplify setting of _ext_fig and _ext_axes.
