### 2021-10-24 ff25943
Fix bare 'except:' statements.
