### 2021-10-24 b1d84b3
Remove brocken _show state tracking. Users should call .show() if they would like the sphere to be displayed and .render() if they'd only like it to be re-rendered.
