### 2021-10-25 0afa9e2
Merge pull request #1693 from Ericgig/omp_flag
openmp flag on linux
