### 2021-10-25 555e39b
Fix definition of site dims in plotting functions.
