+++ qutip/qutip/qobj.py
@@ -1,0 +1,2506 @@
+# This file is part of QuTiP: Quantum Toolbox in Python.
+#
+#    Copyright (c) 2011 and later, Paul D. Nation and Robert J. Johansson.
+#    All rights reserved.
+#
+#    Redistribution and use in source and binary forms, with or without
+#    modification, are permitted provided that the following conditions are
+#    met:
+#
+#    1. Redistributions of source code must retain the above copyright notice,
+#       this list of conditions and the following disclaimer.
+#
+#    2. Redistributions in binary form must reproduce the above copyright
+#       notice, this list of conditions and the following disclaimer in the
+#       documentation and/or other materials provided with the distribution.
+#
+#    3. Neither the name of the QuTiP: Quantum Toolbox in Python nor the names
+#       of its contributors may be used to endorse or promote products derived
+#       from this software without specific prior written permission.
+#
+#    THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+#    "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+#    LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
+#    PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+#    HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+#    SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+#    LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+#    DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+#    THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+#    (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+#    OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+###############################################################################
+"""The Quantum Object (Qobj) class, for representing quantum states and
+operators, and related functions.
+"""
+
+__all__ = ['Qobj', 'qobj_list_evaluate', 'ptrace', 'dag', 'isequal',
+           'issuper', 'isoper', 'isoperket', 'isoperbra', 'isket', 'isbra',
+           'isherm', 'shape', 'dims']
+
+import warnings
+import types
+
+try:
+    import builtins
+except:
+    import __builtin__ as builtins
+
+# import math functions from numpy.math: required for td string evaluation
+from numpy import (arccos, arccosh, arcsin, arcsinh, arctan, arctan2, arctanh,
+                   ceil, copysign, cos, cosh, degrees, e, exp, expm1, fabs,
+                   floor, fmod, frexp, hypot, isinf, isnan, ldexp, log, log10,
+                   log1p, modf, pi, radians, sin, sinh, sqrt, tan, tanh, trunc)
+
+import numpy as np
+import scipy.sparse as sp
+import scipy.linalg as la
+import qutip.settings as settings
+from qutip import __version__
+from qutip.fastsparse import fast_csr_matrix, fast_identity
+from qutip.cy.ptrace import _ptrace
+from qutip.permute import _permute
+from qutip.sparse import (sp_eigs, sp_expm, sp_fro_norm, sp_max_norm,
[…]

[…]
