+++ qutip/lattice.py
@@ -1052,7 +1052,9 @@ class Lattice1d():
                         Qin[i, j] = self._H_intra[
                                 i0*self._length_for_site+i,
                                 j0*self._length_for_site+j]
-                dim_site = list(np.delete(self.cell_tensor_config, [0], None))
+                if len(self.cell_tensor_config) > 1:
+                    dim_site = list(filter(lambda a: a != 1, self.cell_tensor_config))
+                dim_site = self.cell_tensor_config
                 dims_site = [dim_site, dim_site]
                 Hcell[i0][j0] = Qobj(Qin, dims=dims_site)
 
@@ -1152,7 +1154,9 @@ class Lattice1d():
                         Qin[i, j] = self._H_intra[
                                 i0*self._length_for_site+i,
                                 j0*self._length_for_site+j]
-                dim_site = list(np.delete(self.cell_tensor_config, [0], None))
+                if len(self.cell_tensor_config) > 1:
+                    dim_site = list(filter(lambda a: a != 1, self.cell_tensor_config))
+                dim_site = self.cell_tensor_config
                 dims_site = [dim_site, dim_site]
                 Hcell[i0][j0] = Qobj(Qin, dims=dims_site)
 
@@ -1242,6 +1246,7 @@ class Lattice1d():
         plt.axis('off')
         plt.show()
         plt.close()
-        dim_site = list(np.delete(self.cell_tensor_config, [0], None))
+        if len(self.cell_tensor_config) > 1:
+            dim_site = list(filter(lambda a: a != 1, self.cell_tensor_config))
         dims_site = [dim_site, dim_site]
         return Qobj(inter_T, dims=dims_site)
