### 2021-10-26 e00a8cd
Improve and extend bosonic bath mode combining.
