+++ qutip/tests/test_bofin.py
@@ -240,18 +240,10 @@ class TestBosonicHEOMSolver:
         projector = basis(2, 0) * basis(2, 1).dag()
         options = Options(nsteps=15000, store_states=True)
 
-        with pytest.warns(
-            UserWarning,
-            match=(
-                "Two similar real and imag exponents have been collated"
-                " automatically"
-            ),
-        ):
-            hsolver = BosonicHEOMSolver(
-                H_sys, Q2, ckAR, ckAI, vkAR, vkAI,
-                14, options=options,
-            )
-
+        hsolver = BosonicHEOMSolver(
+            H_sys, Q2, ckAR, ckAI, vkAR, vkAI,
+            14, options=options,
+        )
         test = expect(hsolver.run(initial_state, times).states, projector)
 
         np.testing.assert_allclose(test, expected, atol=tol)
@@ -290,17 +282,10 @@ class TestHSolverDL:
         projector = basis(2, 0) * basis(2, 1).dag()
         options = Options(nsteps=15_000, store_states=True)
 
-        with pytest.warns(
-            UserWarning,
-            match=(
-                "Two similar real and imag exponents have been collated"
-                " automatically"
-            ),
-        ):
-            hsolver = HSolverDL(H_sys, Q, coupling_strength, temperature,
-                                14, 2, cut_frequency,
-                                bnd_cut_approx=bnd_cut_approx,
-                                options=options)
+        hsolver = HSolverDL(H_sys, Q, coupling_strength, temperature,
+                            14, 2, cut_frequency,
+                            bnd_cut_approx=bnd_cut_approx,
+                            options=options)
 
         test = expect(hsolver.run(initial_state, times).states, projector)
 
