+++ qutip/nonmarkov/bofin.py
@@ -7,7 +7,6 @@ hierarchy equations of motion (HEOM).
 # Contact: nwlambert@gmail.com
 
 import enum
-import warnings
 from copy import deepcopy
 
 import numpy as np
@@ -148,31 +147,66 @@ class Bath:
 
 class BosonicBath(Bath):
     def __init__(self, Q, ckAR, vkAR, ckAI, vkAI, combine=True):
-        Q = _convert_coup_op(Q, len(ckAR) + len(ckAI))
-        ckAR, ckAI, vkAR, vkAI = _convert_bath_exponents_bosonic(
-            ckAR, ckAI, vkAR, vkAI)
-
-        if combine:
-            Q, ck, vk, NR, NI = (
-                _mangle_bath_exponents_bosonic(Q, ckAR, ckAI, vkAR, vkAI)
+        if len(ckAI) != len(vkAI) or len(ckAR) != len(vkAR):
+            raise ValueError(
+                "The bath exponent lists ckAI and vkAI, and ckAR and vkAR must"
+                " be the same length."
             )
-        else:
-            ck = ckAR + ckAI
-            vk = vkAR + vkAI
-            NR = len(ckAR)
-            NI = len(ckAI)
+        Q = _convert_coup_op(Q, len(ckAR) + len(ckAI))
 
         modes = []
         modes.extend(
-            BathExponent("R", None, Q[i], ck[i], vk[i])
-            for i in range(0, NR))
+            BathExponent("R", None, Qk, ck, vk)
+            for Qk, ck, vk in zip(Q[:len(ckAR)], ckAR, vkAR))
         modes.extend(
-            BathExponent("I", None, Q[i], ck[i], vk[i])
-            for i in range(NR, NR + NI))
-        modes.extend(
-            BathExponent("RI", None, Q[i], ck[i], vk[i], ck2=ck[i+1])
-            for i in range(NR + NI, len(ck), 2))
+            BathExponent("I", None, Qk, ck, vk)
+            for Qk, ck, vk in zip(Q[len(ckAR):], ckAI, vkAI))
+
+        if combine:
+            modes = self._combine(modes)
+
         super().__init__(modes)
+
+    def _combine(self, modes):
+        groups = []
+        remaining = modes[:]
+
+        while remaining:
+            m1 = remaining.pop(0)
+            group = [m1]
+            for m2 in remaining[:]:
+                if (
+                    np.isclose(m1.vk, m2.vk, rtol=1e-5, atol=1e-7) and
+                    np.allclose(m1.Q, m2.Q, rtol=1e-5, atol=1e-7)
+                ):
+                    group.append(m2)
+                    remaining.remove(m2)
+            groups.append(group)
+
+        assert len(modes) == sum(len(g) for g in groups)
+
+        new_modes = []
+        for combine in groups:
+            m1 = combine[0]
+            if len(combine) == 1:
+                new_modes.append(m1)
+            elif all(m2.type == m1.type for m2 in combine):
+                ck = sum(m2.ck for m2 in combine)
+                new_modes.append(BathExponent(m1.type, None, m1.Q, ck, m1.vk))
+            else:
+                ck_R = (
+                    sum(m.ck for m in combine if m.type == m.types.R) +
+                    sum(m.ck for m in combine if m.type == m.types.RI)
+                )
+                ck_I = (
+                    sum(m.ck for m in combine if m.type == m.types.I) +
+                    sum(m.ck2 for m in combine if m.type == m.types.RI)
+                )
+                new_modes.append(
+                    BathExponent("RI", None, m1.Q, ck_R, m1.vk, ck2=ck_I)
+                )
+
+        return new_modes
 
 
[…]
