### 2021-10-26 077bd2f
Update the bofin module docstring.
