### 2021-10-26 6988fa1
Document the bath classes and helpers and rename Bath.modes to Bath.exponents.
