+++ qutip/nonmarkov/bofin.py
@@ -135,9 +135,9 @@ class Bath:
     """
     Represents a list of bath expansion exponents.
 
-    Parameter
-    ---------
-    modes : list of BathExponent
+    Parameters
+    ----------
+    exponents : list of BathExponent
         The exponents of the correlation function describing the bath.
 
     Attributes
@@ -145,82 +145,145 @@ class Bath:
 
     All of the parameters are available as attributes.
     """
-    def __init__(self, modes):
-        self.modes = modes
+    def __init__(self, exponents):
+        self.exponents = exponents
 
 
 class BosonicBath(Bath):
-    def __init__(self, Q, ckAR, vkAR, ckAI, vkAI, combine=True):
+    """
+    A helper class for constructing a bosonic bath from the expansion
+    coefficients and frequencies for the real and imaginary parts of
+    the bath correlation function.
+
+    Parameters
+    ----------
+    Q : Qobj or list of Qobj
+        The coupling operator for the bath. If a list is provided, it
+        represents the coupling operators of the ckAR exponents, followed
+        by the operators for the ckAI exponents.
+
+        FIXME: Separate the real and imaginary coupling operators.
+
+    ckAR : list of complex
+        The coefficients of the expansion terms for the real part of the
+        correlation function. The corresponding frequencies are passed as
+        vkAR.
+
+    vkAR : list of complex
+        The frequencies (exponents) of the expansion terms for the real part of
+        the correlation function. The corresponding ceofficients are passed as
+        ckAR.
+
+    ckAI : list of complex
+        The coefficients of the expansion terms in the imaginary part of the
+        correlation function. The corresponding frequencies are passed as
+        vkAI.
+
+    vkAI : list of complex
+        The frequencies (exponents) of the expansion terms for the imaginary
+        part of the correlation function. The corresponding ceofficients are
+        passed as ckAI.
+
+    combine : bool, default True
+        Whether to combine exponents with the same frequency (and coupling
+        operator). See :meth:`combine` for details.
+    """
+    def _check_cks_and_vks(self, ckAR, vkAR, ckAI, vkAI):
         if len(ckAI) != len(vkAI) or len(ckAR) != len(vkAR):
             raise ValueError(
                 "The bath exponent lists ckAI and vkAI, and ckAR and vkAR must"
                 " be the same length."
             )
+
+    def __init__(self, Q, ckAR, vkAR, ckAI, vkAI, combine=True):
+        self._check_cks_and_vks(ckAR, vkAR, ckAI, vkAI)
         Q = _convert_coup_op(Q, len(ckAR) + len(ckAI))
 
-        modes = []
-        modes.extend(
+        exponents = []
+        exponents.extend(
             BathExponent("R", None, Qk, ck, vk)
             for Qk, ck, vk in zip(Q[:len(ckAR)], ckAR, vkAR))
-        modes.extend(
+        exponents.extend(
             BathExponent("I", None, Qk, ck, vk)
             for Qk, ck, vk in zip(Q[len(ckAR):], ckAI, vkAI))
 
         if combine:
-            modes = self._combine(modes)
+            exponents = self.combine(exponents)
 
-        super().__init__(modes)
+        super().__init__(exponents)
 
-    def _combine(self, modes):
+    @classmethod
+    def combine(cls, exponents, rtol=1e-5, atol=1e-7):
+        """
+        Group bosonic exponents with the same frequency and return a
+        single exponent for each frequency present.
+
+        Exponents with the same frequency are only combined if they share the
+        same coupling operator ``.Q``.
+
+        Parameters
+        ----------
+        exponents : list of BathExponent
[…]
