+++ qutip/nonmarkov/bofin.py
@@ -97,7 +97,7 @@ class BathExponent:
     def _check_ck2(self, type, ck2):
         if type == self.types["RI"]:
             if ck2 is None:
-                raise ValueError("RI bath modes require ck2")
+                raise ValueError("RI bath exponents require ck2")
         else:
             if ck2 is not None:
                 raise ValueError(
@@ -108,7 +108,7 @@ class BathExponent:
     def _check_sigma_bar_k_offset(self, type, offset):
         if type in (self.types["+"], self.types["-"]):
             if offset is None:
-                raise ValueError("+ and - bath modes require sigma_bar_k")
+                raise ValueError("+ and - bath exponents require sigma_bar_k")
         else:
             if offset is not None:
                 raise ValueError(
@@ -548,35 +548,54 @@ class FermionicBath(Bath):
         super().__init__(exponents)
 
 
-class BathStates:
+class HierarchyADOs:
     """
-    A description of bath states coupled to quantum system in the hierarchical
-    equations of motion formulation.
+    A description of ADOs (auxilliary density operators) with the
+    hierarchical equations of motion.
+
+    The list of ADOs is constructed from a list of bath exponents
+    (corresponding to one or more baths). Each ADO is referred to by a label
+    that lists the number of "excitations" of each bath exponent. The
+    level of a label within the hierarchy is the sum of the "excitations"
+    within the label.
+
+    For example the label ``(0, 0, ..., 0)`` represents the density matrix
+    of the system being solved and is the only 0th level label.
+
+    The labels with a single 1, i.e. ``(1, 0, ..., 0)``, ``(0, 1, 0, ... 0)``,
+    etc. are the 1st level labels.
+
+    The second level labels all have either two 1s or a single 2, and so on
+    for the third and higher levels of the hierarchy.
 
     Parameters
     ----------
-    modes : list of BathExponent
-        The exponents of the correlation function describing the bath.
+    exponents : list of BathExponent
+        The exponents of the correlation function describing the bath or
+        baths.
 
     cutoff : int
-        The maximum number of excitations.
+        The maximum depth of the hierarchy (i.e. the maximum sum of
+        "excitations" in the hierarchy ADO labels).
 
     Attributes
     ----------
-    modes : list of BathExponent
-        The exponents of the correlation function describing the bath.
+    exponents : list of BathExponent
+        The exponents of the correlation function describing the bath or
+        baths.
 
     cutoff : int
-        The maximum number of excitations.
+        The maximum depth of the hierarchy (i.e. the maximum sum of
+        "excitations" in the hierarchy ADO labels).
 
     dims : list of int
-        The dimensions of each compontent system within the bath.
+        The dimensions of each exponent within the bath(s).
 
     vk : list of complex
-        The frequency of each exponent within the bath.
+        The frequency of each exponent within the bath(s).
 
     ck : list of complex
-        The coefficient of each exponent within the bath.
+        The coefficient of each exponent within the bath(s).
 
     ck2: list of complex
         For exponents of type "RI", the coefficient of the exponent within
@@ -587,89 +606,86 @@ class BathStates:
         of the corresponding "-" or "+" exponent. For other exponent types,
         the entry is None.
 
-    states: list of tuples
-        A list of the state vectors within the bath.
-
-    n_states: int
[…]
