### 2021-10-26 b925c5c
Document the HEOMSolver and rename N_cut to max_depth for clarity.
