### 2021-10-26 3560a0a
Clean up docstrings for sub-classes of the HEOMSolver.
