### 2021-10-26 795df63
Remove indentation in docstring.
