### 2021-10-26 353d8b8
add multitraj solver
