+++ qutip/solver/mcsolve.py
@@ -1,0 +1,357 @@
+# This file is part of QuTiP: Quantum Toolbox in Python.
+#
+#    Copyright (c) 2011 and later, Paul D. Nation and Robert J. Johansson.
+#    All rights reserved.
+#
+#    Redistribution and use in source and binary forms, with or without
+#    modification, are permitted provided that the following conditions are
+#    met:
+#
+#    1. Redistributions of source code must retain the above copyright notice,
+#       this list of conditions and the following disclaimer.
+#
+#    2. Redistributions in binary form must reproduce the above copyright
+#       notice, this list of conditions and the following disclaimer in the
+#       documentation and/or other materials provided with the distribution.
+#
+#    3. Neither the name of the QuTiP: Quantum Toolbox in Python nor the names
+#       of its contributors may be used to endorse or promote products derived
+#       from this software without specific prior written permission.
+#
+#    THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+#    "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+#    LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
+#    PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+#    HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+#    SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+#    LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+#    DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+#    THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+#    (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+#    OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+###############################################################################
+
+__all__ = ['mcsolve', "McSolver", "MeMcSolver"]
+
+import warnings
+
+import numpy as np
+import numpy.random as np_rng
+from numpy.random import Generator, SeedSequence
+from scipy.integrate import ode
+from scipy.integrate._ode import zvode
+from ..core import (Qobj, QobjEvo, spre, spost, liouvillian, isket, ket2dm,
+                    stack_columns, unstack_columns)
+from ..core.data import to
+from ..core import data as _data
+from .options import SolverOptions
+from .result import Result, MultiTrajResult, MultiTrajResultAveraged
+from .solver_base import Solver, _to_qevo
+from .sesolve import sesolve
+from .mesolve import mesolve
+from .evolver import *
+from time import time
+
+
+def mcsolve(H, psi0, tlist, c_ops=None, e_ops=None, ntraj=1,
+            args=None, options=None, seeds=None):
+    r"""Monte Carlo evolution of a state vector :math:`|\psi \rangle` for a
+    given Hamiltonian and sets of collapse operators, and possibly, operators
+    for calculating expectation values. Options for the underlying ODE solver
+    are given by the Options class.
+
+    mcsolve supports time-dependent Hamiltonians and collapse operators using
+    either Python functions of strings to represent time-dependent
+    coefficients. Note that, the system Hamiltonian MUST have at least one
+    constant term.
[…]

[…]
