### 2021-10-27 89821bf
Rename ckAR, etc to ck_real, etc and numerous small clean-ups from pairing session with Neil.
