+++ qutip/nonmarkov/bofin.py
@@ -154,57 +154,62 @@ class BosonicBath(Bath):
     coefficients and frequencies for the real and imaginary parts of
     the bath correlation function.
 
+    FIXME: Add support for multiple baths of the same kind to HEOMSolver
+    FIXME: Check that exponents are the same kind.
+
     Parameters
     ----------
-    Q : Qobj or list of Qobj
-        The coupling operator for the bath. If a list is provided, it
-        represents the coupling operators of the ckAR exponents, followed
-        by the operators for the ckAI exponents.
+    Q : Qobj
+        The coupling operator for the bath.
 
-        FIXME: Separate the real and imaginary coupling operators.
-
-    ckAR : list of complex
+    ck_real : list of complex
         The coefficients of the expansion terms for the real part of the
         correlation function. The corresponding frequencies are passed as
-        vkAR.
+        vk_real.
 
-    vkAR : list of complex
+    vk_real : list of complex
         The frequencies (exponents) of the expansion terms for the real part of
         the correlation function. The corresponding ceofficients are passed as
-        ckAR.
+        ck_real.
 
-    ckAI : list of complex
+    ck_imag : list of complex
         The coefficients of the expansion terms in the imaginary part of the
         correlation function. The corresponding frequencies are passed as
-        vkAI.
+        vk_imag.
 
-    vkAI : list of complex
+    vk_imag : list of complex
         The frequencies (exponents) of the expansion terms for the imaginary
         part of the correlation function. The corresponding ceofficients are
-        passed as ckAI.
+        passed as ck_imag.
 
     combine : bool, default True
         Whether to combine exponents with the same frequency (and coupling
         operator). See :meth:`combine` for details.
     """
-    def _check_cks_and_vks(self, ckAR, vkAR, ckAI, vkAI):
-        if len(ckAI) != len(vkAI) or len(ckAR) != len(vkAR):
+    def _check_cks_and_vks(self, ck_real, vk_real, ck_imag, vk_imag):
+        if len(ck_real) != len(vk_real) or len(ck_imag) != len(vk_imag):
             raise ValueError(
-                "The bath exponent lists ckAI and vkAI, and ckAR and vkAR must"
-                " be the same length."
+                "The bath exponent lists ck_real and vk_real, and ck_imag and"
+                " vk_imag must be the same length."
             )
 
-    def __init__(self, Q, ckAR, vkAR, ckAI, vkAI, combine=True):
-        self._check_cks_and_vks(ckAR, vkAR, ckAI, vkAI)
-        Q = _convert_coup_op(Q, len(ckAR) + len(ckAI))
+    def _check_coup_op(self, Q):
+        if not isinstance(Q, Qobj):
+            raise ValueError("The coupling operator Q must be a Qobj.")
+
+    def __init__(self, Q, ck_real, vk_real, ck_imag, vk_imag, combine=True):
+        self._check_cks_and_vks(ck_real, vk_real, ck_imag, vk_imag)
+        self._check_coup_op(Q)
 
         exponents = []
         exponents.extend(
-            BathExponent("R", None, Qk, ck, vk)
-            for Qk, ck, vk in zip(Q[:len(ckAR)], ckAR, vkAR))
+            BathExponent("R", None, Q, ck, vk)
+            for ck, vk in zip(ck_real, vk_real)
+        )
         exponents.extend(
-            BathExponent("I", None, Qk, ck, vk)
[…]
