+++ qutip/solver/result.py
@@ -43,7 +43,8 @@ class Result:
 
     num_collapse : int
         Number of collapse operators in simualation.
-"""
+    """
+    def __init__(self, e_ops, options, _super, oper_state):
         self.times = []
 
         self._raw_e_ops = e_ops
@@ -52,11 +53,13 @@ class Result:
         self._last_state = None
 
         self.collapse = None
-        self.stats = {"num_expect": self._e_num}
 
         self._e_ops_dict = False
         self._e_num = 0
         self._e_ops = []
+        self.stats = {"num_expect": self._e_num}
+
+        self._read_options(options, _super, oper_state)
 
         if isinstance(self._raw_e_ops, (Qobj, QobjEvo)):
             e_ops = [self._raw_e_ops]
@@ -252,7 +255,7 @@ class MultiTrajResult:
         Standard derivation of expectation values over `ntraj` trajectories.
         Last state of each trajectories. (ket)
     """
-    def __init__(self):
+    def __init__(self, e_ops, target_tol=None):
         """
         Parameters:
         -----------
@@ -260,11 +263,9 @@ class MultiTrajResult:
             Number of collapses operator used in the McSolver
         """
         self.trajectories = []
-        self._to_dm = True # MCsolve
-        #self.num_c_ops = num_c_ops
-        #self.num_e_ops = num_e_ops
+        self.num_e_ops = len(e_ops)
         self.tlist = None
-        self._target_tols = None
+        self._set_expect_tol(target_tol)
 
     def add(self, one_traj):
         """
@@ -278,7 +279,7 @@ class MultiTrajResult:
         else:
             return np.inf
 
-    def _set_check_expect_tol(self, target_tol):
+    def _set_expect_tol(self, target_tol):
         """
         Set the capacity to stop the map when the estimated error on the
         expectation values is within given tolerance.
@@ -291,6 +292,7 @@ class MultiTrajResult:
             Lastly, target_tol can be a list of pairs of (atol, rtol) for each
             e_ops.
         """
+        self._target_tols = None
         if not target_tol:
             return
         if not self.num_e_ops:
@@ -299,9 +301,9 @@ class MultiTrajResult:
         targets = np.array(target_tol)
         if targets.ndim == 0:
             self._target_tols = np.array([(target_tol, 0.)] * self.num_e_ops)
-        elif targets.shape = (2,):
+        elif targets.shape == (2,):
             self._target_tols = np.ones((self.num_e_ops, 2)) * targets
-        elif targets.shape = (self.num_e_ops, 2):
+        elif targets.shape == (self.num_e_ops, 2):
             self._target_tols = targets
         else:
             raise ValueError("target_tol must be a number, a pair of (atol, "
@@ -328,8 +330,8 @@ class MultiTrajResult:
             axis=0
         ) for i in range(num_e)])
 
-        target = [atol + rtol * mean
-                  for mean, (atol, rtol) in zip(avg, self._target_tols)]
+        target = np.array([atol + rtol * mean
+                  for mean, (atol, rtol) in zip(avg, self._target_tols)])
 
         return np.max(std**2 / target**2 - num_traj + 1)
 
@@ -339,7 +341,7 @@ class MultiTrajResult:
 
     @property
     def average_states(self):
-        if self._to_dm:
+        if self.trajectories[0].states[0].isket:
             finals = [state.proj() for state in self.trajectories[0].states]
             for i in range(1, len(self.trajectories)):
                 finals = [state.proj() + final for final, state
@@ -357,7 +359,7 @@ class MultiTrajResult:
 
     @property
     def average_final_state(self):
-        if self._to_dm:
+        if self.trajectories[0].states[0].isket:
             final = sum(traj.final_state.proj() for traj in self.trajectories)
         else:
[…]
