+++ qutip/nonmarkov/bofin.py
@@ -488,8 +488,13 @@ class DrudeLorentzPadeBath(BosonicBath):
 class FermionicBath(Bath):
     """
     A helper class for constructing a fermionic bath from the expansion
-    coefficients and frequencies for the + and - modes of
+    coefficients and frequencies for the ``+`` and ``-`` modes of
     the bath correlation function.
+
+    There must be the same number of ``+`` and ``-`` modes and their
+    coefficients must be specified in the same order so that ``ck_plus[i],
+    vk_plus[i]`` are the plus coefficient and frequency corresponding
+    to the minus mode ``ck_minus[i], vk_minus[i]``.
 
     Parameters
     ----------
@@ -497,55 +502,58 @@ class FermionicBath(Bath):
         The coupling operator for the bath. ``Q.dag()`` is used as the coupling
         operator for ``+`` mode terms and ``Q`` for the ``-`` mode terms.
 
-    FIXME: Change to ck_plus, vk_plus, ck_minus, vk_minus.
-    FIXME: Assert that lists are the same length & remind user that the
-    plusses and minuses should be in the same order.
+    ck_plus : list of complex
+        The coefficients of the expansion terms for the ``+`` part of the
+        correlation function. The corresponding frequencies are passed as
+        vk_plus.
 
-    ck : list of complex
-        The coefficients of the expansion terms. The even elements of the
-        list are coefficients for ``+`` modes and the odd elements are
-        coefficients for the ``-`` modes. The corresponding frequencies
-        are passed as ``vk``.
+    vk_plus : list of complex
+        The frequencies (exponents) of the expansion terms for the ``+`` part
+        of the correlation function. The corresponding ceofficients are passed
+        as ck_plus.
 
-    vk : list of complex
-        The frequencies (exponents) of the expansion terms. The even elements
-        of the list are frequencies for ``+`` modes and the odd elements are
-        frequencies for the ``-`` modes. The corresponding coefficients
-        are passed as ``ck``.
+    ck_minus : list of complex
+        The coefficients of the expansion terms for the ``-`` part of the
+        correlation function. The corresponding frequencies are passed as
+        vk_minus.
+
+    vk_minus : list of complex
+        The frequencies (exponents) of the expansion terms for the ``-`` part
+        of the correlation function. The corresponding ceofficients are passed
+        as ck_minus.
     """
 
-    def _check_ck_and_vk(self, ck, vk):
-        if (len(ck) != len(vk)
-                or any(len(ck[i]) != len(vk[i]) for i in range(len(ck)))):
-            raise ValueError("Exponents ck and vk must be the same length.")
+    def _check_cks_and_vks(self, ck_plus, vk_plus, ck_minus, vk_minus):
+        if len(ck_plus) != len(vk_plus) or len(ck_minus) != len(vk_minus):
+            raise ValueError(
+                "The bath exponent lists ck_plus and vk_plus, and ck_minus and"
+                " vk_minus must be the same length."
+            )
+        if len(ck_plus) != len(ck_minus):
+            raise ValueError(
+                "The must be the same number of plus and minus exponents"
+                " in the bath, and elements of plus and minus arrays"
+                " should be arranged so that ck_plus[i] is the plus mode"
+                " corresponding to ck_minus[i]."
+            )
 
     def _check_coup_op(self, Q):
         if not isinstance(Q, Qobj):
             raise ValueError("The coupling operator Q must be a Qobj.")
 
-    def __init__(self, Q, ck, vk):
-        self._check_ck_and_vk(ck, vk)
[…]
