### 2021-10-28 45c1b57
Improve naming of ADO state related method parameters.
