### 2021-10-28 3e8d25a
Add a note explaining why Nk = N_exp - 1.
