### 2021-10-28 b247f24
Remove completed lmax removal TODO.
