### 2021-10-28 f856e3d
Reference Padé paper and describe assumption made by time-depedent solver.
