### 2021-10-28 9aee3eb
Add support for multiple baths of the same type to HEOMSolver.
