### 2021-10-28 d447592
Remove unused precalculated superoperators, move utility checking functions to the classes that use them, clean up system dimension handling.
