### 2021-10-28 671ca35
Improve parameter descriptions for HSolverDL and consistently refer to the coupling operator as Q everywhere else.
