+++ qutip/solver/ode/verner7efficient.py
@@ -1,5 +1,6 @@
 """
-Provide a cython implimentation verner 'most-efficient' order 7 runge-Kutta method.
+Provide a cython implimentation verner 'most-efficient'
+order 7 runge-Kutta method.
 See http://people.math.sfu.ca/~jverner/
 """
 # Verner 7 Efficient
@@ -18,7 +19,7 @@ c = np.zeros(rk_extra_step, dtype=np.float64)
 e = np.zeros(rk_step, dtype=np.float64)
 bi = np.zeros((rk_extra_step, denseout_order), dtype=np.float64)
 
-c[0] =  0.
+c[0] = 0.
 c[1] = .5000000000000000000000000000000000000000e-2
 c[2] = .1088888888888888888888888888888888888889
 c[3] = .1633333333333333333333333333333333333333
@@ -29,304 +30,304 @@ c[7] = .9250000000000000000000000000000000000000
 c[8] = 1.
 c[9] = 1.
 
-a[1,0] =  .5000000000000000000000000000000000000000e-2
+a[1, 0] = .5000000000000000000000000000000000000000e-2
 
-a[2,0] = -1.076790123456790123456790123456790123457
-a[2,1] =  1.185679012345679012345679012345679012346
+a[2, 0] = -1.076790123456790123456790123456790123457
+a[2, 1] = 1.185679012345679012345679012345679012346
 
-a[3,0] =  .4083333333333333333333333333333333333333e-1
-a[3,1] =  0.
-a[3,2] =  .1225000000000000000000000000000000000000
+a[3, 0] = .4083333333333333333333333333333333333333e-1
+a[3, 1] = 0.
+a[3, 2] = .1225000000000000000000000000000000000000
 
-a[4,0] =  .6389139236255726780508121615993336109954
-a[4,1] =  0.
-a[4,2] = -2.455672638223656809662640566430653894211
-a[4,3] =  2.272258714598084131611828404831320283215
+a[4, 0] = .6389139236255726780508121615993336109954
+a[4, 1] = 0.
+a[4, 2] = -2.455672638223656809662640566430653894211
+a[4, 3] = 2.272258714598084131611828404831320283215
 
-a[5,0] = -2.661577375018757131119259297861818119279
-a[5,1] =  0.
-a[5,2] =  10.80451388645613769565396655365532838482
-a[5,3] = -8.353914657396199411968048547819291691541
-a[5,4] =  .8204875949566569791420417341743839209619
+a[5, 0] = -2.661577375018757131119259297861818119279
+a[5, 1] = 0.
+a[5, 2] = 10.80451388645613769565396655365532838482
+a[5, 3] = -8.353914657396199411968048547819291691541
+a[5, 4] = .8204875949566569791420417341743839209619
 
-a[6,0] =  6.067741434696770992718360183877276714679
-a[6,1] =  0.
-a[6,2] = -24.71127363591108579734203485290746001803
-a[6,3] =  20.42751793078889394045773111748346612697
-a[6,4] = -1.906157978816647150624096784352757010879
-a[6,5] =  1.006172249242068014790040335899474187268
+a[6, 0] = 6.067741434696770992718360183877276714679
+a[6, 1] = 0.
+a[6, 2] = -24.71127363591108579734203485290746001803
+a[6, 3] = 20.42751793078889394045773111748346612697
+a[6, 4] = -1.906157978816647150624096784352757010879
+a[6, 5] = 1.006172249242068014790040335899474187268
 
-a[7,0] =  12.05467007625320299509109452892778311648
-a[7,1] =  0.
-a[7,2] = -49.75478495046898932807257615331444758322
[…]
