+++ qutip/solver/ode/verner9efficient.py
@@ -1,5 +1,6 @@
 """
-Provide a cython implimentation verner 'most-efficient' order 9 runge-Kutta method.
+Provide a cython implimentation verner 'most-efficient'
+order 9 runge-Kutta method.
 See http://people.math.sfu.ca/~jverner/
 """
 # Verner 9 Efficient
@@ -36,670 +37,670 @@ c[13] = .9012
 c[14] = 1.
 c[15] = 1.
 
-a[1,0] =  .3462e-1
+a[1, 0] = .3462e-1
 
-a[2,0] = -.389335438857287327017042687229284478532e-1
-a[2,1] =  .1359578945245091786499878854939346230295
+a[2, 0] = -.389335438857287327017042687229284478532e-1
+a[2, 1] = .1359578945245091786499878854939346230295
 
-a[3,0] =  .3638413148954266723060635628912731569111e-1
-a[3,1] =  0.
-a[3,2] =  .1091523944686280016918190688673819470733
+a[3, 0] = .3638413148954266723060635628912731569111e-1
+a[3, 1] = 0.
+a[3, 2] = .1091523944686280016918190688673819470733
 
-a[4,0] =  2.025763914393969636805657604282571047511
-a[4,1] =  0.
-a[4,2] = -7.638023836496292020387602153091964592952
-a[4,3] =  6.173259922102322383581944548809393545442
+a[4, 0] = 2.025763914393969636805657604282571047511
+a[4, 1] = 0.
+a[4, 2] = -7.638023836496292020387602153091964592952
+a[4, 3] = 6.173259922102322383581944548809393545442
 
-a[5,0] =  .5112275589406060872792270881648288397197e-1
-a[5,1] =  0.
-a[5,2] =  0.
-a[5,3] =  .1770823794555021537929910813839068684087
-a[5,4] =  .80277624092225014536138698108025283759e-3
+a[5, 0] = .5112275589406060872792270881648288397197e-1
+a[5, 1] = 0.
+a[5, 2] = 0.
+a[5, 3] = .1770823794555021537929910813839068684087
+a[5, 4] = .80277624092225014536138698108025283759e-3
 
-a[6,0] =  .1316006357975216279279871693164256985334
-a[6,1] =  0.
-a[6,2] =  0.
-a[6,3] = -.2957276252669636417685183174672273730699
-a[6,4] =  .878137803564295237421124704053886667082e-1
-a[6,5] =  .6213052975225274774321435005639430026100
+a[6, 0] = .1316006357975216279279871693164256985334
+a[6, 1] = 0.
+a[6, 2] = 0.
+a[6, 3] = -.2957276252669636417685183174672273730699
+a[6, 4] = .878137803564295237421124704053886667082e-1
+a[6, 5] = .6213052975225274774321435005639430026100
 
-a[7,0] =  .7166666666666666666666666666666666666667e-1
-a[7,1] =  0.
-a[7,2] =  0.
-a[7,3] =  0.
-a[7,4] =  0.
-a[7,5] =  .3305533578915319409260346730051472207728
-a[7,6] =  .2427799754418013924072986603281861125606
+a[7, 0] = .7166666666666666666666666666666666666667e-1
+a[7, 1] = 0.
+a[7, 2] = 0.
+a[7, 3] = 0.
+a[7, 4] = 0.
+a[7, 5] = .3305533578915319409260346730051472207728
+a[7, 6] = .2427799754418013924072986603281861125606
 
-a[8,0] =  .7180664062500000000000000000000000000000e-1
-a[8,1] =  0.
-a[8,2] =  0.
-a[8,3] =  0.
-a[8,4] =  0.
-a[8,5] =  .3294380283228177160744825466257672816401
-a[8,6] =  .1165190029271822839255174533742327183599
[…]
