+++ qutip/solver/ode/verner7efficient.py
@@ -129,7 +129,7 @@ a[11, 1] = 0.
 a[11, 2] = 0.
 a[11, 3] = .2249586182670571550244187743667190903405
 a[11, 4] = .1744370924877637539031751304611402542578e-1
-a[11, 5] = .7669379876829393188009028209348812321417e-2
+a[11, 5] = -.7669379876829393188009028209348812321417e-2
 a[11, 6] = .3435896044073284645684381456417912794447e-1
 a[11, 7] = -.4102097230093949839125144540100346681769e-1
 a[11, 8] = .2565113300520561655297104906598973655221e-1
