### 2021-10-28 84474a8
Update qutip/floquet.py
Co-authored-by: Jake Lishman <jake@binhbar.com>
