### 2021-10-28 4a3f662
Trajsolver doc
