### 2021-10-29 5c73300
Fix transposition errors and improve scaling of Floquet methods (#1248)
* Changes in floquet_master_equation_rates to make it much faster

* Correction of transposition errors in floquet_master_equation_tensor and changes to make it faster

* Correction in floquet_markov_mesolve. The basis change that returns the results in the computational had errors. NB: Results are still given in the interaction picture.

* Delete comments

* Delete comments

* Corrected code style

* Corrected code style

* Corrected code style

* Additional test functions.

* Added options for propagator and sesolve as d
[…]
