+++ qutip/floquet.py
@@ -10,6 +10,8 @@ __all__ = ['floquet_modes', 'floquet_modes_t', 'floquet_modes_table',
 import numpy as np
 import scipy.linalg as la
 import scipy
+import warnings
+from copy import copy
 from numpy import angle, pi, exp, sqrt
 from types import FunctionType
 from qutip.qobj import Qobj, isket
@@ -27,7 +29,8 @@ from qutip.cy.spmatfuncs import cy_ode_rhs
 from qutip.expect import expect
 from qutip.utilities import n_thermal
 
-def floquet_modes(H, T, args=None, sort=False, U=None):
+
+def floquet_modes(H, T, args=None, sort=False, U=None, options=None):
     """
     Calculate the initial Floquet modes Phi_alpha(0) for a driven system with
     period T.
@@ -55,6 +58,9 @@ def floquet_modes(H, T, args=None, sort=False, U=None):
         If U is `None` (default), it will be calculated from the Hamiltonian
         `H` using :func:`qutip.propagator.propagator`.
 
+    options : :class:`qutip.solver`
+        options for the ODE solver. For the propagator U.
+
     Returns
     -------
 
@@ -66,7 +72,7 @@ def floquet_modes(H, T, args=None, sort=False, U=None):
 
     if U is None:
         # get the unitary propagator
-        U = propagator(H, T, [], args)
+        U = propagator(H, T, [], args=args, options=copy(options))
 
     # find the eigenstates for the propagator
     evals, evecs = la.eig(U.full())
@@ -94,7 +100,9 @@ def floquet_modes(H, T, args=None, sort=False, U=None):
 
     return kets_order, e_quasi[order]
 
-def floquet_modes_t(f_modes_0, f_energies, t, H, T, args=None):
+
+def floquet_modes_t(f_modes_0, f_energies, t, H, T, args=None,
+                    options=None):
     """
     Calculate the Floquet modes at times tlist Phi_alpha(tlist) propagting the
     initial Floquet modes Phi_alpha(0)
@@ -120,6 +128,9 @@ def floquet_modes_t(f_modes_0, f_energies, t, H, T, args=None):
     T : float
         The period of the time-dependence of the hamiltonian.
 
+    options : :class:`qutip.solver`
+        options for the ODE solver. For the propagator.
+
     Returns
     -------
 
@@ -134,7 +145,7 @@ def floquet_modes_t(f_modes_0, f_energies, t, H, T, args=None):
 
     # get the unitary propagator from 0 to t
     if t > 0.0:
-        U = propagator(H, t, [], args)
+        U = propagator(H, t, [], args, options=copy(options))
 
         for n in np.arange(len(f_modes_0)):
             f_modes_t.append(U * f_modes_0[n] * exp(1j * f_energies[n] * t))
@@ -143,7 +154,9 @@ def floquet_modes_t(f_modes_0, f_energies, t, H, T, args=None):
 
     return f_modes_t
 
-def floquet_modes_table(f_modes_0, f_energies, tlist, H, T, args=None):
+
+def floquet_modes_table(f_modes_0, f_energies, tlist, H, T, args=None,
+                        options=None):
     """
     Pre-calculate the Floquet modes for a range of times spanning the floquet
     period. Can later be used as a table to look up the floquet modes for
@@ -170,6 +183,9 @@ def floquet_modes_table(f_modes_0, f_energies, tlist, H, T, args=None):
     args : dictionary
         dictionary with variables required to evaluate H
 
+    options : :class:`qutip.solver`
+        options for the ODE solver.
+
     Returns
     -------
 
@@ -178,12 +194,13 @@ def floquet_modes_table(f_modes_0, f_energies, tlist, H, T, args=None):
         A nested list of Floquet modes as kets for each time in `tlist`
 
     """
[…]
