### 2021-10-31 e9d5fab
Fixed typo in algorithm combining bath exponents
