### 2021-10-31 e04a3b8
Make bofin solver respect "store_states" option
