### 2021-10-31 6c161aa
Added e_ops parameter to qutip.nonmarkov.bofin.HEOMSolver#run
