+++ qutip/solver/multitraj.py
@@ -1,3 +1,13 @@
+from .. import Qobj, QobjEvo
+from .options import SolverOptions, SolverOdeOptions
+from .result import Result, MultiTrajResult, MultiTrajResultAveraged
+from ..ui.progressbar import progess_bars
+from ..core.data import to
+from .parallel import get_map
+from time import time
+from .solver_base import Solver, classproperty
+import numpy as np
+
 
 class MultiTrajSolver:
     """
@@ -36,14 +46,15 @@ class MultiTrajSolver:
         self.traj_solvers = []
 
         for seed in seeds:
-            traj_solver = self._traj_solver_class(self)
+            traj_solver = self._traj_solver_class(self, e_ops=self.e_ops,
+                                                  options=self.options)
             traj_solver.start(state0, t0, seed)
             self.traj_solvers.append(traj_solver)
 
     def step(self, t, args=None):
         if not self.traj_solvers:
             raise RuntimeError("The `start` method must called first.")
-        multi = len(self.traj_solvers) = 1
+        multi = len(self.traj_solvers) == 1
         out = [traj_solver.step(t, args, safe=multi)
                for traj_solver in self.traj_solvers]
         return out if len(out) > 1 else out[0]
@@ -52,9 +63,10 @@ class MultiTrajSolver:
         """
         Get a :cls:`Solver` for a single trajectory.
         """
-        return sefl._traj_solver_class(self)
+        return self._traj_solver_class(self, e_ops=self.e_ops,
+                                       options=self.options)
 
-    def run(self, state0, tlist, ntraj=1, * args=None, options=None,
+    def run(self, state0, tlist, ntraj=1, *, args=None, options=None,
             e_ops=None, timeout=0, target_tol=None, seed=None):
         """
         Do the evolution of the Quantum system.
@@ -111,9 +123,9 @@ class MultiTrajSolver:
         if options is not None:
             self.options = options
 
-        self.result = (MultiTrajResult()
+        self.result = (MultiTrajResult(e_ops or self.e_ops)
                        if self.options.mcsolve['keep_runs_results']
-                       else MultiTrajResultAveraged())
+                       else MultiTrajResultAveraged(e_ops or self.e_ops))
         self._run_args = state0, tlist
         self._run_kwargs = {'args': args, 'e_ops': e_ops or self.e_ops}
         self.result.stats['run time'] = 0
@@ -158,14 +170,14 @@ class MultiTrajSolver:
             raise RuntimeError("No previous computation, use `run` first.")
         start_time = time()
         seeds = self._read_seed(seed, ntraj)
-        map_func = get_map[self.options.mcsolve]
+        map_func = get_map[self.options.mcsolve['map']]
         map_kw = self.options.mcsolve['map_options']
         if timeout:
             map_kw['job_timeout'] = timeout
         if target_tol:
             self.result._set_check_expect_tol(target_tol)
         map_func(
-            self._traj_solver_class(self).run, seeds,
+            self._traj_solver_class(self)._run, seeds,
             self._run_args, self._run_kwargs,
             reduce_func=self.result.add, map_kw=map_kw,
             progress_bar=self.options["progress_bar"],
@@ -176,7 +188,7 @@ class MultiTrajSolver:
         return self.result
 
     @property
-    def option(self):
+    def options(self):
         return self._options
 
     @options.setter
@@ -206,17 +218,25 @@ class _TrajectorySolver(Solver):
     A new :method:`_argument` is used to update the args in `rhs` and other
     :class:`QobjEvo` and need to be overloaded if other operators are used.
     """
+    name = "Generic trajectory solver"
+    def __init__(self, rhs, *, e_ops=None, options=None):
[…]
