+++ qutip/solver/mcsolve.py
@@ -46,10 +46,10 @@ from ..core.data import to
 from ..core import data as _data
 from .options import SolverOptions
 from .result import Result, MultiTrajResult, MultiTrajResultAveraged
-from .solver_base import Solver, _to_qevo
+from .solver_base import Solver
+from .multitraj import MultiTrajSolver, _TrajectorySolver
 from .sesolve import sesolve
-from .mesolve import mesolve
-from .evolver import *
+from .integrator import Integrator
 from time import time
 
 
@@ -161,112 +161,58 @@ def mcsolve(H, psi0, tlist, c_ops=None, e_ops=None, ntraj=1,
     return mc.run(psi0, tlist=tlist, ntraj=ntraj)
 
 
-
-
-# -----------------------------------------------------------------------------
-# MONTE CARLO CLASS
-# -----------------------------------------------------------------------------
-class McSolver(MultiTrajSolver):
+class _McTrajectorySolver(_TrajectorySolver):
     """
     ... TODO
     """
-    _traj_solver_class = _OneTrajMcSolver
-    _super = False
-    name = "mcsolve"
-    optionsclass = SolverOptions
-
-    def __init__(self, H, c_ops, e_ops=None, options=None,
-                 times=None, args=None, seed=None):
-        _time_start = time()
-        self.stats = {}
-        if not isinstance(options, SolverOptions):
-            raise ValueError("options must be an instance of "
-                             "qutip.solver.SolverOptions")
-
-        self.seed_sequence = SeedSequence(seed)
-        self.traj_solvers = []
-        self.result = None
-
-        self.e_ops = e_ops or []
-        self.options = options
-
-        self._c_ops = [QobjEvo(op, args=args, tlist=times) for c_op in c_ops]
-        self._n_ops = [c_op.dag() * c_op for c_op in self._c_ops]
-        self._system = -1j* QobjEvo(H, args=args, tlist=times)
-        for n_evo in self._n_ops:
-            self._system -= 0.5 * n_evo
-
-        self.stats['solver'] = "MonteCarlo Evolution"
-        self.stats['num_collapse'] = len(c_ops)
-        self.stats["preparation time"] = time() - _time_start
-
-    def _check_state_dims(self, state):
-        if not state.isket:
-            raise TypeError("The unitary solver requires psi0 to be "
-                            "a ket as initial state "
-                            "or a unitary as initial operator.")
-
-        if self._system.dims[1] != state.dims[0]:
-            raise TypeError("".join([
-                            "incompatible dimensions ",
-                            repr(self._system.dims),
-                            " and ",
-                            repr(state.dims),])
-                           )
-
-
-
-class _OneTrajMcSolver(_OneTraj):
-    """
-    ... TODO
-    """
-    _super = False
     name = "mcsolve"
 
-    def __init__(self, rhs, *, e_ops=None, options=None):
-        if isinstance(rhs, (QobjEvo, Qobj)):
-            self.rhs = QobjEvo(rhs)
-        else:
-            TypeError("The rhs must be a QobjEvo")
-        self.e_ops = e_ops
-        self.options = options
-        self.stats = {"preparation time": 0}
-        self._state_metadata = {}
-
-        if options.mcsolve['BitGenerator']:
-            if hasattr(np_rng, options.mcsolve['BitGenerator']):
-                self.bit_gen = getattr(np_rng, options.mcsolve['BitGenerator'])
-            else:
-                raise ValueError("BitGenerator is not know to numpy.random")
-        else:
-            self.bit_gen = np_rng.PCG64
+    def __init__(self, parent, *, e_ops=None, options=None):
+        rhs = parent.rhs
+        self._c_ops = parent._c_ops
[…]
