### 2021-11-01 fffdcfe
Bofin: fixed typo related to handling multiple baths
