### 2021-11-01 17b609a
Bofin: added documentation for e_ops parameter
