### 2021-11-02 61d0064
Change tests/test_metrics.py so that only dnorm tests are skipped if cvxpy not installed
