### 2021-11-02 2d9d07d
Neaten up e_ops handling.
