+++ qutip/nonmarkov/bofin.py
@@ -86,6 +86,10 @@ class BathExponent:
         For exponents of type "-" it gives the offset of the corresponding
         "+" exponent.
 
+    label : optional, str
+        A label for the exponent (often the name of the bath). It
+        defaults to None.
+
     Attributes
     ----------
 
@@ -116,7 +120,9 @@ class BathExponent:
                 )
 
     def __init__(
-            self, type, dim, Q, ck, vk, ck2=None, sigma_bar_k_offset=None):
+            self, type, dim, Q, ck, vk, ck2=None, sigma_bar_k_offset=None,
+            label=None,
+    ):
         if not isinstance(type, self.types):
             type = self.types[type]
         self._check_ck2(type, ck2)
@@ -128,6 +134,7 @@ class BathExponent:
         self.vk = vk
         self.ck2 = ck2
         self.sigma_bar_k_offset = sigma_bar_k_offset
+        self.label = label
 
 
 class Bath:
@@ -182,6 +189,11 @@ class BosonicBath(Bath):
     combine : bool, default True
         Whether to combine exponents with the same frequency (and coupling
         operator). See :meth:`combine` for details.
+
+    label : optional, str
+        A label for the bath exponents (for example, the name of the
+        bath). It defaults to None but can be set to help identify which
+        bath an exponent is from.
     """
     def _check_cks_and_vks(self, ck_real, vk_real, ck_imag, vk_imag):
         if len(ck_real) != len(vk_real) or len(ck_imag) != len(vk_imag):
@@ -194,17 +206,20 @@ class BosonicBath(Bath):
         if not isinstance(Q, Qobj):
             raise ValueError("The coupling operator Q must be a Qobj.")
 
-    def __init__(self, Q, ck_real, vk_real, ck_imag, vk_imag, combine=True):
+    def __init__(
+            self, Q, ck_real, vk_real, ck_imag, vk_imag, combine=True,
+            label=None,
+    ):
         self._check_cks_and_vks(ck_real, vk_real, ck_imag, vk_imag)
         self._check_coup_op(Q)
 
         exponents = []
         exponents.extend(
-            BathExponent("R", None, Q, ck, vk)
+            BathExponent("R", None, Q, ck, vk, label=label)
             for ck, vk in zip(ck_real, vk_real)
         )
         exponents.extend(
-            BathExponent("I", None, Q, ck, vk)
+            BathExponent("I", None, Q, ck, vk, label=label)
             for ck, vk in zip(ck_imag, vk_imag)
         )
 
@@ -221,6 +236,10 @@ class BosonicBath(Bath):
 
         Exponents with the same frequency are only combined if they share the
         same coupling operator ``.Q``.
+
+        Note that combined exponents take their label from the first
+        exponent in the group being combined (i.e. the one that occurs first
+        in the given exponents list).
 
         Parameters
         ----------
@@ -263,9 +282,9 @@ class BosonicBath(Bath):
             ):
                 # the group is either type I or R
                 ck = sum(exp.ck for exp in combine)
-                new_exponents.append(
-                    BathExponent(exp1.type, None, exp1.Q, ck, exp1.vk)
-                )
+                new_exponents.append(BathExponent(
+                    exp1.type, None, exp1.Q, ck, exp1.vk, label=exp1.label,
+                ))
             else:
                 # the group includes both type I and R exponents
                 ck_R = (
@@ -276,9 +295,10 @@ class BosonicBath(Bath):
                     sum(exp.ck for exp in combine if exp.type == exp.types.I) +
                     sum(exp.ck2 for exp in combine if exp.type == exp.types.RI)
                 )
[…]
