### 2021-11-03 db2c9d4
working mcsolve with super
