+++ qutip/solver/mcsolve.py
@@ -31,7 +31,7 @@
 #    OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 ###############################################################################
 
-__all__ = ['mcsolve', "McSolver", "MeMcSolver"]
+__all__ = ['mcsolve', "McSolver"]
 
 import warnings
 
@@ -54,54 +54,19 @@ from time import time
 
 
 def mcsolve(H, psi0, tlist, c_ops=None, e_ops=None, ntraj=1,
-            args=None, options=None, seeds=None):
-    r"""Monte Carlo evolution of a state vector :math:`|\psi \rangle` for a
-    given Hamiltonian and sets of collapse operators, and possibly, operators
-    for calculating expectation values. Options for the underlying ODE solver
-    are given by the Options class.
-
-    mcsolve supports time-dependent Hamiltonians and collapse operators using
-    either Python functions of strings to represent time-dependent
-    coefficients. Note that, the system Hamiltonian MUST have at least one
-    constant term.
-
-    As an example of a time-dependent problem, consider a Hamiltonian with two
-    terms ``H0`` and ``H1``, where ``H1`` is time-dependent with coefficient
-    ``sin(w*t)``, and collapse operators ``C0`` and ``C1``, where ``C1`` is
-    time-dependent with coeffcient ``exp(-a*t)``.  Here, w and a are constant
-    arguments with values ``W`` and ``A``.
-
-    Using the Python function time-dependent format requires two Python
-    functions, one for each collapse coefficient. Therefore, this problem could
-    be expressed as::
-
-        def H1_coeff(t,args):
-            return sin(args['w']*t)
-
-        def C1_coeff(t,args):
-            return exp(-args['a']*t)
-
-        H = [H0, [H1, H1_coeff]]
-
-        c_ops = [C0, [C1, C1_coeff]]
-
-        args={'a': A, 'w': W}
-
-    or in String (Cython) format we could write::
-
-        H = [H0, [H1, 'sin(w*t)']]
-
-        c_ops = [C0, [C1, 'exp(-a*t)']]
-
-        args={'a': A, 'w': W}
-
-    Constant terms are preferably placed first in the Hamiltonian and collapse
-    operator lists.
+            args=None, options=None, seeds=None, target_tol=None):
+    r"""
+    Monte Carlo evolution of a state vector :math:`|\psi \rangle` for a
+    given Hamiltonian and sets of collapse operators. Options for the
+    underlying ODE solver are given by the Options class.
 
     Parameters
     ----------
-    H : :class:`qutip.Qobj`, ``list``
-        System Hamiltonian.
+    H : :class:`qutip.Qobj`, :class:`qutip.QobjEvo`, ``list``, callable.
+        System Hamiltonian as a Qobj, QobjEvo, can also be a function or list
+        that can be made into a Qobjevo. (See :class:`qutip.QobjEvo`'s
+        documentation). ``H`` can be a superoperator (liouvillian) if some
+        collapse operators are to be treated deterministicly.
 
     psi0 : :class:`qutip.Qobj`
         Initial state vector
@@ -110,55 +75,59 @@ def mcsolve(H, psi0, tlist, c_ops=None, e_ops=None, ntraj=1,
         Times at which results are recorded.
 
     ntraj : int
-        Number of trajectories to run.
+        Maximum number of trajectories to run. Can be cut short if a time limit
+        is passed in options (per default, mcsolve will stop after 1e8 sec)::
+            ``options.mcsolve['map_options']['timeout'] = max_sec``
+        Or if the target tolerance is reached, see ``target_tol``.
 
-    c_ops : :class:`qutip.Qobj`, ``list``
-        single collapse operator or a ``list`` of collapse operators.
+    c_ops : ``list``
[…]

[…]
