+++ qutip/solver/mcsolve.py
@@ -1,36 +1,3 @@
-# This file is part of QuTiP: Quantum Toolbox in Python.
-#
-#    Copyright (c) 2011 and later, Paul D. Nation and Robert J. Johansson.
-#    All rights reserved.
-#
-#    Redistribution and use in source and binary forms, with or without
-#    modification, are permitted provided that the following conditions are
-#    met:
-#
-#    1. Redistributions of source code must retain the above copyright notice,
-#       this list of conditions and the following disclaimer.
-#
-#    2. Redistributions in binary form must reproduce the above copyright
-#       notice, this list of conditions and the following disclaimer in the
-#       documentation and/or other materials provided with the distribution.
-#
-#    3. Neither the name of the QuTiP: Quantum Toolbox in Python nor the names
-#       of its contributors may be used to endorse or promote products derived
-#       from this software without specific prior written permission.
-#
-#    THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-#    "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-#    LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
-#    PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
-#    HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
-#    SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
-#    LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
-#    DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
-#    THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-#    (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-#    OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-###############################################################################
-
 __all__ = ['mcsolve', "McSolver"]
 
 import warnings
@@ -132,7 +99,7 @@ def mcsolve(H, psi0, tlist, c_ops=None, e_ops=None, ntraj=1,
 
 class _McTrajectorySolver(_TrajectorySolver):
     """
-    ... TODO
+    Solver for one mcsolve trajectory. Created by a :class:`McSolver`.
     """
     name = "mcsolve"
 
@@ -174,7 +141,6 @@ class _McTrajectorySolver(_TrajectorySolver):
         return _data.norm.l2(state)
 
     def _step(self, t, copy=True):
-        """Evolve to t, including jumps."""
         t_old, y_old = self._integrator.get_state(copy=False)
         norm_old = self.prob_func(y_old)
         while t_old < t:
@@ -191,6 +157,7 @@ class _McTrajectorySolver(_TrajectorySolver):
         return t_old, _data.mul(y_old, 1 / self.norm_func(y_old))
 
     def _do_collapse(self, norm_old, norm, t_prev):
+        """Find and apply a collapse."""
         t_final, _ = self._integrator.get_state()
         tries = 0
         while tries < self.norm_steps:
@@ -300,7 +267,7 @@ class McSolver(MultiTrajSolver):
 
         c_ops = [QobjEvo(c_op) for c_op in c_ops]
         if H.issuper:
[…]
