### 2021-11-04 5382027
Add __repr__ for BathExponents and combine option for all sub-classes of BosonicBath.
