### 2021-11-04 98277ca
Add tests for bath classes.
