+++ qutip/tests/test_bofin.py
@@ -8,11 +8,15 @@ from numpy.linalg import eigvalsh
 from scipy.integrate import quad
 
 from qutip import (
-    Qobj, sigmaz, basis, destroy, expect, Options
+    Qobj, spre, spost, sigmax, sigmaz, basis, destroy, expect, isequal, Options
 )  # FIXME: Add tests system specified using QobjEvo
 from qutip.nonmarkov.bofin import (
     BathExponent,
     Bath,
+    BosonicBath,
+    DrudeLorentzBath,
+    DrudeLorentzPadeBath,
+    FermionicBath,
     HierarchyADOs,
     BosonicHEOMSolver,
     FermionicHEOMSolver,
@@ -21,7 +25,7 @@ from qutip.nonmarkov.bofin import (
 
 
 def check_exponent(
-    exp, type, dim, Q, ck, vk, ck2=None, sigma_bar_k_offset=None,
+    exp, type, dim, Q, ck, vk, ck2=None, sigma_bar_k_offset=None, tag=None,
 ):
     """ Check the attributes of a BathExponent. """
     assert exp.type is BathExponent.types[type]
@@ -31,6 +35,7 @@ def check_exponent(
     assert exp.vk == vk
     assert exp.ck2 == ck2
     assert exp.sigma_bar_k_offset == sigma_bar_k_offset
+    assert exp.tag == tag
 
 
 class TestBathExponent:
@@ -41,8 +46,8 @@ class TestBathExponent:
         exp_i = BathExponent("I", None, Q=None, ck=1.0, vk=2.0)
         check_exponent(exp_i, "I", None, Q=None, ck=1.0, vk=2.0)
 
-        exp_i = BathExponent("RI", None, Q=None, ck=1.0, vk=2.0, ck2=3.0)
-        check_exponent(exp_i, "RI", None, Q=None, ck=1.0, vk=2.0, ck2=3.0)
+        exp_ri = BathExponent("RI", None, Q=None, ck=1.0, vk=2.0, ck2=3.0)
+        check_exponent(exp_ri, "RI", None, Q=None, ck=1.0, vk=2.0, ck2=3.0)
 
         exp_p = BathExponent(
             "+", 2, Q=None, ck=1.0, vk=2.0, sigma_bar_k_offset=-1,
@@ -57,6 +62,9 @@ class TestBathExponent:
         check_exponent(
             exp_m, "-", 2, Q=None, ck=1.0, vk=2.0, sigma_bar_k_offset=+1,
         )
+
+        exp_tag = BathExponent("R", None, Q=None, ck=1.0, vk=2.0, tag="tag1")
+        check_exponent(exp_tag, "R", None, Q=None, ck=1.0, vk=2.0, tag="tag1")
 
         for exp_type, kw in [
             ("R", {}),
@@ -84,6 +92,23 @@ class TestBathExponent:
                 " specified for + and - bath exponents"
             )
 
+    def test_repr(self):
+        exp1 = BathExponent("R", None, Q=sigmax(), ck=1.0, vk=2.0)
+        assert repr(exp1) == (
+            "<BathExponent type=R dim=None Q.dims=[[2], [2]]"
+            " ck=1.0 vk=2.0 ck2=None"
+            " sigma_bar_k_offset=None tag=None>"
+        )
+        exp2 = BathExponent(
+            "+", None, Q=None, ck=1.0, vk=2.0, sigma_bar_k_offset=-1,
+            tag="bath1",
+        )
+        assert repr(exp2) == (
+            "<BathExponent type=+ dim=None Q.dims=None"
+            " ck=1.0 vk=2.0 ck2=None"
+            " sigma_bar_k_offset=-1 tag='bath1'>"
+        )
+
 
 class TestBath:
[…]
