### 2021-11-04 2d5b175
Add test for filter_by_level.
