### 2021-11-05 bb13331
Add filter_by_exponents for extracting labels corresponding to 'excitations' in particular combinations of exponents.
