### 2021-11-05 e67b4b2
Added rtol option for tests.
