+++ qutip/tests/core/data/test_mathematics.py
@@ -302,6 +302,11 @@ class _GenericOpMixin:
         expected value.  If the output is a Data type, the tolerance is
         per-element of the output.
 
+    rtol: float
+        The relative tolerance to use when comparing the test value with the
+        expected value.  If the output is a Data type, the tolerance is
+        per-element of the output.
+
     shapes: list of (list of shapes)
         A list of the sets of shapes which should be used for the tests of
         mathematical correctness.  Each element of the list is a set of shapes,
@@ -323,6 +328,7 @@ class _GenericOpMixin:
     # produce slightly different results to sparse algebra, since the order of
     # multiplications and additions will be different.
     tol = 1e-10
+    rtol = 1e-7 # Same default as numpy
     shapes = []
     bad_shapes = []
     specialisations = []
@@ -400,14 +406,16 @@ class UnaryOpMixin(_GenericOpMixin):
         if issubclass(out_type, Data):
             assert test.shape == expected.shape
             np.testing.assert_allclose(test.to_array(), expected,
-                                       atol=self.tol)
+                                       atol=self.tol,
+                                       rtol=self.rtol)
         elif out_type is list:
             for test_, expected_ in zip(test, expected):
                 assert test_.shape == expected_.shape
                 np.testing.assert_allclose(test_.to_array(),
-                                           expected_, atol=self.tol)
+                                           expected_, atol=self.tol,
+                                           rtol=self.rtol)
         else:
-            assert abs(test - expected) < self.tol
+            assert abs(test - expected) < self.tol + self.rtol*abs(expected)
 
     def test_incorrect_shape_raises(self, op, data_m):
         """
@@ -441,9 +449,10 @@ class UnaryScalarOpMixin(_GenericOpMixin):
         if issubclass(out_type, Data):
             assert test.shape == expected.shape
             np.testing.assert_allclose(test.to_array(), expected,
-                                       atol=self.tol)
+                                       atol=self.tol,
+                                       rtol=self.rtol)
         else:
-            assert abs(test - expected) < self.tol
+            assert abs(test - expected) < self.tol + self.rtol*abs(expected)
 
 
 class BinaryOpMixin(_GenericOpMixin):
@@ -463,9 +472,10 @@ class BinaryOpMixin(_GenericOpMixin):
         if issubclass(out_type, Data):
             assert test.shape == expected.shape
             np.testing.assert_allclose(test.to_array(), expected,
-                                       atol=self.tol)
+                                       atol=self.tol,
+                                       rtol=self.rtol)
         else:
-            assert abs(test - expected) < self.tol
+            assert abs(test - expected) < self.tol + self.rtol*abs(expected)
 
     def test_incorrect_shape_raises(self, op, data_l, data_r):
         """
@@ -497,9 +507,10 @@ class TernaryOpMixin(_GenericOpMixin):
         if issubclass(out_type, Data):
             assert test.shape == expected.shape
             np.testing.assert_allclose(test.to_array(), expected,
-                                       atol=self.tol)
+                                       atol=self.tol,
+                                       rtol=self.rtol)
         else:
-            assert abs(test - expected) < self.tol
+            assert abs(test - expected) < self.tol + self.rtol*abs(expected)
 
     def test_incorrect_shape_raises(self, op, data_l, data_m, data_r):
         """
@@ -544,9 +555,10 @@ class TestAdd(BinaryOpMixin):
         if issubclass(out_type, Data):
             assert test.shape == expected.shape
             np.testing.assert_allclose(test.to_array(), expected,
-                                       atol=self.tol)
+                                       atol=self.tol,
+                                       rtol=self.rtol)
         else:
[…]
