+++ qutip/tests/core/data/test_reshape.py
@@ -7,7 +7,7 @@ from qutip.core.data import CSR, Dense
 
 class TestSplitColumns(UnaryOpMixin):
     def op_numpy(self, matrix):
-       return [column[:, np.newaxis] for column in matrix.T]
+        return [column[:, np.newaxis] for column in matrix.T]
 
     specialisations = [
         pytest.param(data.split_columns_csr, CSR, list),
@@ -112,7 +112,7 @@ class TestReshape(UnaryOpMixin):
                                    rtol=self.rtol)
 
     @pytest.mark.parametrize('rows, columns', [(-2, -50), (-50, -2), (3, 10)],
-                            ids=["negative1","negative2","invalid"])
+                             ids=["negative1", "negative2", "invalid"])
     def test_incorrect_rows_raises(self, op, data_m, out_type, rows, columns):
         with pytest.raises(ValueError):
             op(data_m(), rows, columns)
