### 2021-11-05 f9b7e80
Combine filter_by_level and filter_by_exponents.
