### 2021-11-05 607dce2
Add tests for valid combinations of level and excitations.
