### 2021-11-05 9cba85b
remove e_ops in __init__
