### 2021-11-05 5778eb9
remove default e_ops in tests
