### 2021-11-05 cf38a7c
more test + e_ops in run
