### 2021-11-06 9d59e56
[fix] adding title to hinton plot
