+++ qutip/visualization.py
@@ -267,6 +267,7 @@ def hinton(rho, xlabels=None, ylabels=None, title=None, ax=None, cmap=None,
 
     if not (xlabels or ylabels):
         ax.axis('off')
+    ax.set_title(title)
 
     ax.axis('equal')
     ax.set_frame_on(False)
