### 2021-11-06 cf85146
[fix] title is set for matrix_histogram even if "there's no figure"
