### 2021-11-06 0585ccb
[feature] modifications to title in other functions
