### 2021-11-07 0898592
Remove unnecessary whitespace
