### 2021-11-07 b84929d
Fix the color style phase
