### 2021-11-08 f256cce
Fix typos in equations and text
