+++ doc/guide/dynamics/dynamics-stochastic.rst
@@ -15,10 +15,10 @@ phase of the system. With this method, the resulting detection rate depends is
 .. math::
    :label: jump_rate
 
-   \tau = \tr \left((\gamma^2 + \gamma (C+C^\dag) + C^\dag C)\rho \right)
+   \tau = \tr \left((\gamma^2 + \gamma (C+C^\dagger) + C^\dagger C)\rho \right)
 
 With :math:`\gamma`, the strength of the external beam and :math:`C` the collapse
-operator. When the beam is very strong :math:`(\gamma >> C^\dag C)`,
+operator. When the beam is very strong :math:`(\gamma >> C^\dagger C)`,
 the rate becomes a constant term plus a term proportional to the quadrature of
 the system.
 
@@ -32,20 +32,20 @@ In closed systems, the resulting stochastic differential equation is
 	:label: jump_ssesolve
 
 	\delta \psi(t) = - i H \psi(t) \delta t
-	                 - \sum_n \left( \frac{C_n^{+} C_n}{2} -\frac{e_n}{2} C_n
-					 + \frac{e_n^2}{8} \right) \psi \delta t
-	                 + \sum_n \left( C_n - \frac{e_n}{2} \right) \psi \delta \omega
+	                 - \sum_n \left( \frac{C_n^\dagger C_n}{2} -\frac{e_n}{2} C_n
+					 + \frac{e_n^2}{8} \right) \psi(t) \delta t
+	                 + \sum_n \left( C_n - \frac{e_n}{2} \right) \psi(t) \delta \omega
 
 with
 
 .. math::
    :label: jump_matrix_element
 
-   e_n = \left<\psi(t)|C_n + C_n^{+}|\psi(t)\right>
+   e_n = \left<\psi(t)|C_n + C_n^\dagger|\psi(t)\right>
 
 Here :math:`\delta \omega` is a Wiener increment.
 
-In QuTiP, this is available with the function :func:`ssesolve`.
+In QuTiP, this is available with the function :func:`qutip.ssesolve`.
 
 .. plot::
     :context:
@@ -56,7 +56,8 @@ In QuTiP, this is available with the function :func:`ssesolve`.
     sm = tensor(destroy(2), qeye(10))
 
     H = 2*np.pi*a.dag()*a + 2*np.pi*sm.dag()*sm + 2*np.pi*0.25*(sm*a.dag() + sm.dag()*a)
-    data = ssesolve(H, psi0, times, sc_ops=[np.sqrt(0.1) * a], e_ops=[a.dag()*a, sm.dag()*sm], method="homodyne")
+    data = ssesolve(H, psi0, times, sc_ops=[np.sqrt(0.1) * a],
+                    e_ops=[a.dag()*a, sm.dag()*sm], method="homodyne")
 
     plt.figure()
     plt.plot(times, data.expect[0], times, data.expect[1])
@@ -86,18 +87,18 @@ collapses
 with
 
 .. math::
- 	:label: disipator
+ 	:label: dissipator
 
-	D(C, \rho) = \frac{1}{2} \left[2 C \rho(t) C^{+}
-			   - \rho(t) C^{+} C - C^{+} C \rho(t) \right].
+	D(C, \rho) = \frac{1}{2} \left[2 C \rho(t) C^\dagger
+			   - \rho(t) C^\dagger C - C^ C \rho(t) \right].
 
 The stochastic part is given by
 
 .. math::
 	:label: stochastic_smesolve
 
-	d_2 = \left(C \rho(t) + \rho(t) C^{+} - \rm{tr}\left(C \times \rho
-					 + \rho \times C^{+} \right)\rho(t) \right),
+	d_2 = \left(C \rho(t) + \rho(t) C^\dagger - \rm{tr}\left(C \times \rho
[…]
