### 2021-11-08 7f3af70
Apply suggestions from code review
Co-authored-by: Simon Cross <hodgestar+github@gmail.com>
