### 2021-11-08 8943072
Cleaned up formulas in DrudeLorentzBath
