### 2021-11-08 f5cbe13
Added option to calculate terminator for DrudeLorentzPadeBath, made DrudeLorentzBath and DrudeLorentzPadeBath provide delta parameter
