### 2021-11-09 875e697
Document MKL parameters.
