### 2021-11-09 888bc1e
Fixed bug in calculating terminator when exponents are not combined
