### 2021-11-09 be7fa34
Add .exps(label) for returning the exponents used within an ADO label.
