### 2021-11-09 7476cbb
Set use_mkl=True by default in steady_state.
