+++ qutip/nonmarkov/bofin.py
@@ -1283,7 +1283,7 @@ class HEOMSolver:
 
     def steady_state(
         self,
-        use_mkl=False, mkl_max_iter_refine=100, mkl_weighted_matching=False
+        use_mkl=True, mkl_max_iter_refine=100, mkl_weighted_matching=False
     ):
         """
         Compute the steady state of the system.
@@ -1294,14 +1294,14 @@ class HEOMSolver:
             Whether to use mkl or not. If mkl is not installed or if
             this is false, use the scipy splu solver instead.
 
-        mkl_max_iter_refine : Int
+        mkl_max_iter_refine : int
             Specifies the the maximum number of iterative refinement steps that
             the MKL PARDISO solver performs.
 
             For a complete description, see iparm(8) in
             http://cali2.unilim.fr/intel-xe/mkl/mklman/GUID-264E311E-ACED-4D56-AC31-E9D3B11D1CBF.htm.
 
-        mkl_weighted_matching : Boolean
+        mkl_weighted_matching : bool
             MKL PARDISO can use a maximum weighted matching algorithm to
             permute large elements close the diagonal. This strategy adds an
             additional level of reliability to the factorization methods.
