### 2021-11-10 8f5a83b
Add tests for HierarchyADOsState.
