### 2021-11-10 702fcf0
Remove nice use of sum(..., start=...) in order to support Python <=3.6.
