### 2021-11-10 e71698b
Add combined option to high-level Bosonic solvers.
