### 2021-11-10 0d4a8f1
Add steady state tests for high-level bosonic solvers and factor out common pure-dephasing model.
