+++ qutip/tests/test_bofin.py
@@ -8,8 +8,9 @@ from numpy.linalg import eigvalsh
 from scipy.integrate import quad
 
 from qutip import (
-    Qobj, spre, spost, sigmax, sigmaz, basis, destroy, expect, isequal, Options
-)  # FIXME: Add tests system specified using QobjEvo
+    basis, destroy, expect, isequal, liouvillian, spre, spost, sigmax, sigmaz,
+    Qobj, QobjEvo, Options,
+)
 from qutip.nonmarkov.bofin import (
     BathExponent,
     Bath,
@@ -477,32 +478,45 @@ class TestHierarchyADOsState:
         ado_state.extract(1) == Qobj(ado_soln[1, :], dims=rho.dims)
 
 
-class TestBosonicHEOMSolver:
-    @pytest.mark.filterwarnings("ignore::scipy.integrate.IntegrationWarning")
-    @pytest.mark.parametrize(['fake_timedep'], [
-        pytest.param(False, id="static"),
-        pytest.param(True, id="timedep"),
-    ])
-    def test_run_pure_dephasing_model(self, fake_timedep):
-        """ Compare with pure-dephasing analytical result. """
-        tol = 1e-3
-        gamma = 0.05
-        lam = 0.025
+class DrudeLorentzPureDephasingModel:
+    """ Analytic model for testing the HEOM solver. """
+    def __init__(self, lam, gamma, T, Nk):
+        self.lam = lam
+        self.gamma = gamma
+        self.T = T
+        self.Nk = Nk
+        self.H = Qobj(np.zeros((2, 2)))
+        self.Q = sigmaz()
+
+    def rho(self):
+        """ Initial state. """
+        return 0.5 * Qobj(np.ones((2, 2)))
+
+    def state_results(self, states):
+        projector = basis(2, 0) * basis(2, 1).dag()
+        return expect(states, projector)
+
+    def analytic_results(self, tlist):
+        lam, gamma, T = self.lam, self.gamma, self.T
         lam_c = lam / np.pi
-        T = 1 / 0.95
-        times = np.linspace(0, 10, 21)
 
         def _integrand(omega, t):
-            J = 2*lam_c * omega * gamma / (omega**2 + gamma**2)
+            J = 2 * lam_c * omega * gamma / (omega**2 + gamma**2)
             return (-4 * J * (1 - np.cos(omega*t))
                     / (np.tanh(0.5*omega / T) * omega**2))
 
         # Calculate the analytical results by numerical integration
-        expected = [0.5*np.exp(quad(_integrand, 0, np.inf, args=(t,))[0])
-                    for t in times]
+        return [
+            0.5 * np.exp(quad(_integrand, 0, np.inf, args=(t,))[0])
+            for t in tlist
+        ]
 
-        Nk = 2
-        ck_real = [lam * gamma * (1/np.tan(gamma / (2 * T)))]
+    def bath_coefficients(self):
+        """ Correlation function expansion coefficients for the Drude-Lorentz bath.
+        """
+        lam, gamma, T = self.lam, self.gamma, self.T
+        Nk = self.Nk
+        ck_real = [lam * gamma * (1 / np.tan(gamma / (2 * T)))]
         ck_real.extend([
             (4 * lam * gamma * T * 2 * np.pi * k * T /
                 ((2 * np.pi * k * T)**2 - gamma**2))
@@ -512,66 +526,107 @@ class TestBosonicHEOMSolver:
         vk_real.extend([2 * np.pi * k * T for k in range(1, Nk + 1)])
         ck_imag = [lam * gamma * (-1.0)]
         vk_imag = [gamma]
+        return ck_real, vk_real, ck_imag, vk_imag
 
-        H_sys = Qobj(np.zeros((2, 2)))
[…]
