### 2021-11-10 039802c
Update fermionic solver tests to the new-style H_sys selection.
