### 2021-11-10 95ca469
Refactor Fermionic solver tests.
