+++ qutip/tests/test_bofin.py
@@ -479,7 +479,8 @@ class TestHierarchyADOsState:
 
 
 class DrudeLorentzPureDephasingModel:
-    """ Analytic model for testing the HEOM solver. """
+    """ Analytic Drude-Lorentz pure-dephasing model for testing the HEOM solver.
+    """
     def __init__(self, lam, gamma, T, Nk):
         self.lam = lam
         self.gamma = gamma
@@ -629,46 +630,95 @@ class TestHSolverDL:
         assert rho_final == ado_state.extract(0)
 
 
-class TestFermionicHEOMSolver:
-    @pytest.mark.filterwarnings("ignore::scipy.integrate.IntegrationWarning")
-    @pytest.mark.parametrize(['evo'], [
-        pytest.param("qobj"),
-        pytest.param("qobjevo"),
-        pytest.param("listevo"),
-        pytest.param("qobj"),
-    ])
-    @pytest.mark.parametrize(['liouvillianize'], [
-        pytest.param(False, id="hamiltonian"),
-        pytest.param(True, id="liouvillian"),
-    ])
-    def test_discrete_level_model(self, evo, liouvillianize):
-        """ Compare to discrete-level current analytics. """
-        atol = 1e-3
-        Gamma = 0.01  # coupling strength
-        W = 1.  # cut-off
-        T = 0.025851991  # temperature
-        beta = 1. / T
+class DiscreteLevelCurrentModel:
+    """ Analytic discrete level current model for testing the HEOM solver.
+    """
+    def __init__(self, gamma, W, T, lmax):
+        self.gamma = gamma
+        self.W = W
+        self.T = T
+        self.lmax = lmax  # Pade cut-off
+        self.beta = 1. / T
 
-        theta = 2.  # Bias
-        mu_l = theta/2.
-        mu_r = -theta/2.
+        # single fermion
+        self.e1 = 1.
+        d1 = destroy(2)
+        self.H = self.e1 * d1.dag() * d1
+        self.Q = d1
 
-        lmax = 10  # Pade cut-off
+        # bias
+        self.theta = 2.
+
+    def rho(self):
+        """ Initial state. """
+        return 0.5 * Qobj(np.ones((2, 2)))
+
+    def state_current(self, ado_state):
+        level_1_aux = [
+            (ado_state.extract(label), ado_state.exps(label)[0])
+            for label in ado_state.filter(level=1)
+        ]
+
+        def exp_sign(exp):
+            return 1 if exp.type == exp.types["+"] else -1
+
+        # right hand modes are the first k modes in ck/vk_plus and ck/vk_minus
+        # and thus the first 2 * k exponents
+        k = self.lmax + 1
+        return 1.0j * sum(
+            exp_sign(exp) * (exp.Q * aux).tr()
+            for aux, exp in level_1_aux[:2 * k]
+        )
+
+    def analytic_current(self):
+        Gamma, W, beta, e1 = self.gamma, self.W, self.beta, self.e1
+        mu_l = self.theta / 2.
+        mu_r = - self.theta / 2.
+
+        def f(x):
+            return 1 / (np.exp(x) + 1.)
+
+        def Gamma_w(w, mu):
+            return Gamma * W**2 / ((w-mu)**2 + W**2)
+
+        def lamshift(w, mu):
+            return (w-mu)*Gamma_w(w, mu)/(2*W)
+
+        def integrand(w):
+            return (
+                ((2 / (np.pi)) * Gamma_w(w, mu_l) * Gamma_w(w, mu_r) *
+                    (f(beta * (w - mu_l)) - f(beta * (w - mu_r)))) /
[…]
