### 2021-11-10 6327157
Use scipy's spsolve directly.
