### 2021-11-10 be772d6
mcsolve support dict e_ops
