### 2021-11-11 795692d
mcsolve more responsive to options changes
