### 2021-11-11 984754f
Add a very small entries to the pure-dephasing system hamiltonian to make the system non-singular.
