### 2021-11-11 bb24fd2
Tweak some error messages and docstrings. Use max_depth consistently as the name of the maximum hierarchy depth.
