### 2021-11-11 eb7faa6
Add tests for the HEOMSolver.
