+++ qutip/tests/test_bofin.py
@@ -9,7 +9,7 @@ from scipy.integrate import quad
 
 from qutip import (
     basis, destroy, expect, isequal, liouvillian, spre, spost, sigmax, sigmaz,
-    Qobj, QobjEvo, Options,
+    tensor, Qobj, QobjEvo, Options,
 )
 from qutip.nonmarkov.bofin import (
     BathExponent,
@@ -20,6 +20,7 @@ from qutip.nonmarkov.bofin import (
     FermionicBath,
     HierarchyADOs,
     HierarchyADOsState,
+    HEOMSolver,
     BosonicHEOMSolver,
     FermionicHEOMSolver,
     HSolverDL,
@@ -317,10 +318,10 @@ class TestHierarchyADOs:
 
     def test_create(self):
         exponents = self.mk_exponents([2, 3])
-        ados = HierarchyADOs(exponents, cutoff=2)
+        ados = HierarchyADOs(exponents, max_depth=2)
 
         assert ados.exponents == exponents
-        assert ados.cutoff == 2
+        assert ados.max_depth == 2
 
         assert ados.dims == [2, 3]
         assert ados.vk == [2.0, 2.0]
@@ -333,7 +334,7 @@ class TestHierarchyADOs:
         ]
 
     def test_state_idx(self):
-        ados = HierarchyADOs(self.mk_exponents([2, 3]), cutoff=2)
+        ados = HierarchyADOs(self.mk_exponents([2, 3]), max_depth=2)
         assert ados.idx((0, 0)) == 0
         assert ados.idx((0, 1)) == 1
         assert ados.idx((0, 2)) == 2
@@ -341,7 +342,7 @@ class TestHierarchyADOs:
         assert ados.idx((1, 1)) == 4
 
     def test_next(self):
-        ados = HierarchyADOs(self.mk_exponents([2, 3]), cutoff=2)
+        ados = HierarchyADOs(self.mk_exponents([2, 3]), max_depth=2)
         assert ados.next((0, 0), 0) == (1, 0)
         assert ados.next((0, 0), 1) == (0, 1)
         assert ados.next((1, 0), 0) is None
@@ -349,7 +350,7 @@ class TestHierarchyADOs:
         assert ados.next((1, 1), 1) is None
 
     def test_prev(self):
-        ados = HierarchyADOs(self.mk_exponents([2, 3]), cutoff=2)
+        ados = HierarchyADOs(self.mk_exponents([2, 3]), max_depth=2)
         assert ados.prev((0, 0), 0) is None
         assert ados.prev((0, 0), 1) is None
         assert ados.prev((1, 0), 0) == (0, 0)
@@ -358,7 +359,7 @@ class TestHierarchyADOs:
         assert ados.prev((0, 2), 1) == (0, 1)
 
     def test_exps(self):
-        ados = HierarchyADOs(self.mk_exponents([3, 3, 2]), cutoff=4)
+        ados = HierarchyADOs(self.mk_exponents([3, 3, 2]), max_depth=4)
         assert ados.exps((0, 0, 0)) == ()
         assert ados.exps((1, 0, 0)) == (ados.exponents[0],)
         assert ados.exps((2, 0, 0)) == (
@@ -371,13 +372,13 @@ class TestHierarchyADOs:
         )
 
     def test_filter_by_nothing(self):
-        ados = HierarchyADOs(self.mk_exponents([2, 3]), cutoff=2)
+        ados = HierarchyADOs(self.mk_exponents([2, 3]), max_depth=2)
         assert ados.filter() == [
             (0, 0), (0, 1), (0, 2), (1, 0), (1, 1),
         ]
 
     def test_filter_by_level(self):
[…]
