+++ qutip/solver/result.py
@@ -2,7 +2,7 @@
 import numpy as np
 from ..core import Qobj, QobjEvo, spre, issuper, expect
 
-__all__ = ["Result", "MultiTrajResult", "MultiTrajResultAveraged"]
+__all__ = ["Result", "MultiTrajResult"]
 
 
 class _Expect_Caller:
@@ -46,25 +46,14 @@ class Result:
     """
     def __init__(self, e_ops, options, _super, oper_state):
         self.times = []
-
-        self._raw_e_ops = e_ops
         self._states = []
         self._expects = []
         self._last_state = None
-
         self.collapse = None
 
+        self._raw_e_ops = e_ops
         self._e_ops_dict = False
-        self._e_num = len(self._raw_e_ops) if self._raw_e_ops else 0
         self._e_ops = []
-        self.stats = {
-            "num_expect": self._e_num,
-            "solver": "",
-            "method": "",
-        }
-
-        self._read_options(options, _super, oper_state)
-
         if isinstance(self._raw_e_ops, (Qobj, QobjEvo)):
             e_ops = [self._raw_e_ops]
         elif isinstance(self._raw_e_ops, dict):
@@ -85,6 +74,13 @@ class Result:
             elif callable(e):
                 self._e_ops.append(e)
             self._expects.append([])
+        self._e_num = len(self._e_ops) if self._e_ops else 0
+        self._read_options(options, _super, oper_state)
+        self.stats = {
+            "num_expect": self._e_num,
+            "solver": "",
+            "method": "",
+        }
 
     def _read_options(self, options, _super, oper_state):
         if options['store_states'] is not None:
@@ -108,10 +104,13 @@ class Result:
     def _normalize(self, state):
         return state * (1/state.norm())
 
-    def add(self, t, state):
+    def add(self, t, state, last=True):
         """
         Add a state to the results for the time t of the evolution.
         The state is expected to be a Qobj with the right dims.
+
+        last : bool {True}
+            Whether the given state *can* be the last state.
         """
         self.times.append(t)
         # this is so we don't make a copy if normalize is
@@ -122,14 +121,11 @@ class Result:
 
         if self._store_states:
             self._states.append(state_norm or state.copy())
-        elif self._store_final_state:
+        elif self._store_final_state and last:
             self._last_state = state_norm or state.copy()
 
         for i, e_call in enumerate(self._e_ops):
             self._expects[i].append(e_call(t, state_norm or state))
-
-    def copy(self):
-        return Result(self._raw_e_ops, self.options, self.super)
 
     @property
     def states(self):
@@ -184,8 +180,11 @@ class Result:
 
 class MultiTrajResult:
     """
-    Contain result of simulations with multiple trajectories. Keeps all
-    trajectories' data.
+    Contain result of simulations with multiple trajectories.
+
+    Parameters
+    ----------
+
 
     Property
     --------
@@ -260,18 +259,30 @@ class MultiTrajResult:
         Standard derivation of expectation values over `ntraj` trajectories.
         Last state of each trajectories. (ket)
     """
-    def __init__(self, ntraj, e_ops=(), c_ops=(), target_tol=None):
+    def __init__(self, ntraj, e_ops=None, c_ops=None, target_tol=None, options=None):
         """
         Parameters:
         -----------
         num_c_ops: int
             Number of collapses operator used in the McSolver
         """
+        self.options = options or SolverResultsOptions()
+        self._save_traj = self.options['keep_runs_results']
         self.trajectories = []
-        self.num_e_ops = len(e_ops)
-        self.num_c_ops = len(c_ops)
-        self.tlist = None
+        self._sum_states = None
+        self._sum_last_states = None
[…]
