+++ qutip/solver/options.py
@@ -177,6 +177,10 @@ class SolverResultsOptions:
         "all" will normalize both ket and dm.
         On "ket", only 'ket' output are normalized.
         Leave empty for no normalization.
+
+
+    keep_runs_results: bool
+        Keep all trajectories results or save only the average.
     """
     options = {
         # store final state?
@@ -186,6 +190,8 @@ class SolverResultsOptions:
         # Normalize output of solvers
         # (turned off for batch unitary propagator mode)
         "normalize_output": "ket",
+
+        "keep_runs_results": False,
     }
 
 
@@ -222,9 +228,6 @@ class McOptions:
         Max. number of steps used to find wavefunction norm to within norm_tol
         in mcsolve.
 
-    keep_runs_results: bool
-        Keep all trajectories results or save only the average.
-
     map : str  {'parallel', 'serial', 'loky'}
         How to run the trajectories.
         'parallel' use python's multiprocessing.
@@ -252,11 +255,9 @@ class McOptions:
 
         "map": "parallel_map",
 
-        "keep_runs_results": False,
-
         "mc_corr_eps": 1e-10,
 
-        "BitGenerator": False,
+        "bitgenerator": "",
 
         "map_options": {
             'num_cpus': multiprocessing.cpu_count(),

[…]
