### 2021-11-12 d5f6af1
Rewrite alpha using np.diag.
