### 2021-11-12 2ba051a
Add tests for using the Drude-Lorentz baths with HEOMSolver.
