### 2021-11-12 4e1d023
Rewrite alpha_p using np.diag.
