### 2021-11-12 36dc6a4
Replace range(Nk - 1, Nk) with [Nk - 1] to make it clearer that it has a single element.
