### 2021-11-12 10a9f11
Add a progress bar option and test and document it.
