+++ qutip/tests/test_bofin.py
@@ -25,6 +25,7 @@ from qutip.nonmarkov.bofin import (
     FermionicHEOMSolver,
     HSolverDL,
 )
+from qutip.ui.progressbar import BaseProgressBar, TextProgressBar
 
 
 def check_exponent(
@@ -583,6 +584,19 @@ class TestHEOMSolver:
         hsolver = HEOMSolver(H, [bath] * 3, 2)
         assert hsolver.ados.exponents == exponents * 3
         assert hsolver.ados.max_depth == 2
+
+    def test_create_progress_bar(self):
+        Q = sigmaz()
+        H = sigmax()
+        bath = Bath([
+            BathExponent("R", None, Q=Q, ck=1.1, vk=2.1),
+        ])
+
+        hsolver = HEOMSolver(H, bath, 2)
+        assert isinstance(hsolver.progress_bar, BaseProgressBar)
+
+        hsolver = HEOMSolver(H, bath, 2, progress_bar=True)
+        assert isinstance(hsolver.progress_bar, TextProgressBar)
 
     def test_create_bath_errors(self):
         Q = sigmaz()
