### 2021-11-14 db7d2df
Split HEOM bath descriptions out into their own module.
