+++ qutip/tests/test_bofin.py
@@ -1,5 +1,5 @@
 """
-Tests for the Bosonic HEOM solvers.
+Tests for qutip.nonmarkov.bofin.
 """
 
 import numpy as np
@@ -8,16 +8,18 @@ from numpy.linalg import eigvalsh
 from scipy.integrate import quad
 
 from qutip import (
-    basis, destroy, expect, isequal, liouvillian, spre, spost, sigmax, sigmaz,
+    basis, destroy, expect, liouvillian, sigmax, sigmaz,
     tensor, Qobj, QobjEvo, Options,
 )
-from qutip.nonmarkov.bofin import (
+from qutip.nonmarkov.heom_baths import (
     BathExponent,
     Bath,
     BosonicBath,
     DrudeLorentzBath,
     DrudeLorentzPadeBath,
     FermionicBath,
+)
+from qutip.nonmarkov.bofin import (
     HierarchyADOs,
     HierarchyADOsState,
     HEOMSolver,
@@ -26,289 +28,6 @@ from qutip.nonmarkov.bofin import (
     HSolverDL,
 )
 from qutip.ui.progressbar import BaseProgressBar, TextProgressBar
-
-
-def check_exponent(
-    exp, type, dim, Q, ck, vk, ck2=None, sigma_bar_k_offset=None, tag=None,
-):
-    """ Check the attributes of a BathExponent. """
-    assert exp.type is BathExponent.types[type]
-    assert exp.dim == dim
-    assert exp.Q == Q
-    assert exp.ck == ck
-    assert exp.vk == vk
-    assert exp.ck2 == ck2
-    assert exp.sigma_bar_k_offset == sigma_bar_k_offset
-    assert exp.tag == tag
-
-
-class TestBathExponent:
-    def test_create(self):
-        exp_r = BathExponent("R", None, Q=None, ck=1.0, vk=2.0)
-        check_exponent(exp_r, "R", None, Q=None, ck=1.0, vk=2.0)
-
-        exp_i = BathExponent("I", None, Q=None, ck=1.0, vk=2.0)
-        check_exponent(exp_i, "I", None, Q=None, ck=1.0, vk=2.0)
-
-        exp_ri = BathExponent("RI", None, Q=None, ck=1.0, vk=2.0, ck2=3.0)
-        check_exponent(exp_ri, "RI", None, Q=None, ck=1.0, vk=2.0, ck2=3.0)
-
-        exp_p = BathExponent(
-            "+", 2, Q=None, ck=1.0, vk=2.0, sigma_bar_k_offset=-1,
-        )
-        check_exponent(
-            exp_p, "+", 2, Q=None, ck=1.0, vk=2.0, sigma_bar_k_offset=-1,
-        )
-
-        exp_m = BathExponent(
-            "-", 2, Q=None, ck=1.0, vk=2.0, sigma_bar_k_offset=+1,
-        )
-        check_exponent(
-            exp_m, "-", 2, Q=None, ck=1.0, vk=2.0, sigma_bar_k_offset=+1,
-        )
-
-        exp_tag = BathExponent("R", None, Q=None, ck=1.0, vk=2.0, tag="tag1")
-        check_exponent(exp_tag, "R", None, Q=None, ck=1.0, vk=2.0, tag="tag1")
-
-        for exp_type, kw in [
-            ("R", {}),
-            ("I", {}),
-            ("+", {"sigma_bar_k_offset": 1}),
-            ("-", {"sigma_bar_k_offset": -1}),
-        ]:
-            with pytest.raises(ValueError) as err:
-                BathExponent(
-                    exp_type, None, Q=None, ck=1.0, vk=2.0, ck2=3.0, **kw,
-                )
-            assert str(err.value) == (
[…]
