### 2021-11-14 893de87
Rename bofin module to HEOM solvers.
