### 2021-11-14 b47de22
Rename BoFiN derived HEOM solver modules to bofin_baths and bofin_solvers.
