### 2021-11-14 bb71ee9
Update non-markovian API docs.
