### 2021-11-15 bb0fec8
Fix syntax for Pade approximant link.
