### 2021-11-16 c42e637
Add test for _GatherHEOMRHS.
