### 2021-11-16 529f4c2
Add implementation of underdamped bosonic bath.
