### 2021-11-16 aae51af
Fix name of Returns section in BosonicBath.combine docstring.
