### 2021-11-16 6f7c031
Add UnderDampedBath to qutip.heom API.
