### 2021-11-16 5fe5774
Add solver test for UnderDampedBath.
