### 2021-11-16 8f308e2
Use pytest.approx for checking bath coefficients.
