### 2021-11-18 1343163
Add 'order' to interCoeff to replace stepCoeff
