+++ qutip/core/cy/coefficient.pyx
@@ -8,6 +8,7 @@ from ..interpolate import Cubic_Spline
 import inspect
 import pickle
 import scipy
+from scipy.interpolate import interp1d
 import numpy as np
 cimport numpy as cnp
 cimport cython
@@ -87,9 +88,11 @@ cdef class Coefficient:
         """
         Replace the arguments (``args``) of a coefficient.
 
-        Returns a new :obj:`Coefficient` if the coefficient has arguments, or the original coefficient if it does not.
-        Arguments to replace may be supplied either in a dictionary as the first position argument, or passed as
-        keywords, or as a combination of the two. Arguments not replaced retain their previous values.
+        Returns a new :obj:`Coefficient` if the coefficient has arguments, or
+        the original coefficient if it does not. Arguments to replace may be
+        supplied either in a dictionary as the first position argument, or
+        passed as keywords, or as a combination of the two. Arguments not
+        replaced retain their previous values.
 
         Parameters
         ----------
@@ -346,9 +349,11 @@ def coeff(t, args):
         """
         Replace the arguments (``args``) of a coefficient.
 
-        Returns a new :obj:`Coefficient` if the coefficient has arguments, or the original coefficient if it does not.
-        Arguments to replace may be supplied either in a dictionary as the first position argument, or passed as
-        keywords, or as a combination of the two. Arguments not replaced retain their previous values.
+        Returns a new :obj:`Coefficient` if the coefficient has arguments, or
+        the original coefficient if it does not. Arguments to replace may be
+        supplied either in a dictionary as the first position argument, or
+        passed as keywords, or as a combination of the two. Arguments not
+        replaced retain their previous values.
 
         Parameters
         ----------
@@ -403,7 +408,8 @@ cdef class InterpolateCoefficient(Coefficient):
 
 cdef class InterCoefficient(Coefficient):
     """
-    A :obj:`Coefficient` built from a cubic spline interpolation of a numpy array.
+    A :obj:`Coefficient` built from a cubic spline interpolation of a numpy
+    array.
 
     Parameters
     ----------
@@ -411,113 +417,131 @@ cdef class InterCoefficient(Coefficient):
         The array of coefficient values to interpolate.
 
     tlist : np.ndarray
-        An array of times corresponding to each coefficient value. The times must be
-        increasing, but do not need to be uniformly spaced.
+        An array of times corresponding to each coefficient value. The times
+        must be increasing, but do not need to be uniformly spaced.
     """
-    cdef int n_t, constant
+    cdef int order
     cdef double dt
     cdef double[::1] tlist
-    cdef complex[::1] coeff_arr, second_derr
-    cdef object tlist_np, coeff_np, second_np
+    cdef complex[:, :] poly
+    cdef object np_arrays
 
-    def __init__(self, coeff_arr, tlist, _second=None, _constant=None):
-        self.tlist_np = tlist
-        self.tlist = tlist
-        self.coeff_np = coeff_arr
-        self.coeff_arr = coeff_arr
-        if _second is None:
-            self.second_np, self.constant = _prep_cubic_spline(coeff_arr,
-                                                                tlist)
-        else:
-            self.second_np = _second
-            self.constant = _constant
-        self.second_derr = self.second_np
-        self.dt = tlist[1] - tlist[0]
+    def __init__(self, coeff_arr, tlist, int order):
+        if coeff_arr.ndim != 1:
+            raise ValueError("The array to interpolate must be a 1D array")
+        if coeff_arr.shape != tlist.shape:
+            raise ValueError("tlist must be the same len "
+                             "as the array to interpolate")
[…]
