### 2021-11-18 768fe14
Correctly use .dag'ed exponent operators in _grad_next_fermionic.
