### 2021-11-18 59955f7
Use a non-self-adjoint operator for the FermionicBath test so that differences between Q and Q.dag() are detectable by the test.
