### 2021-11-18 2220832
Use the same coupling operator for + and - FermionicBath exponents (they are the same bath, after all) and instead handle the adjoint taking in the HEOMSolver.
