### 2021-11-18 ff1078f
Remove extra return op from _grad_next_fermionic.
