### 2021-11-18 7f86e9d
Added packaging to the installation.rst requirements.
