+++ qutip/tests/core/test_coefficient.py
@@ -32,8 +32,9 @@
 ###############################################################################
 import pytest
 import pickle
-import qutip as qt
+import qutip
 import numpy as np
+import scipy.interpolate as interp
 from functools import partial
 from qutip.core.coefficient import (coefficient, norm, conj, shift,
                                     CompilationOptions,
@@ -116,7 +117,7 @@ def coeff_generator(style, func):
     if style == "arraylog":
         return coefficient(base(tlistlog, **args), tlist=tlistlog)
     if style == "spline":
-        return coefficient(qt.Cubic_Spline(0, 1, base(tlist, **args)))
+        return coefficient(qutip.Cubic_Spline(0, 1, base(tlist, **args)))
     if style == "string" and func == "f":
         return coefficient("exp(w * t * pi)", args=args)
     if style == "string" and func == "g":
@@ -144,7 +145,7 @@ def coeff_generator(style, func):
                  1e-6, id="nonlinear_array"),
     pytest.param(f_asarraylog, {'tlist': tlistlog, 'order': 0},
                  1e-1, id="nonlinear_step_array"),
-    pytest.param(qt.Cubic_Spline(0, 1, f_asarray), {},
+    pytest.param(qutip.Cubic_Spline(0, 1, f_asarray), {},
                  1e-6, id="Cubic_Spline"),
     pytest.param("exp(w * t * pi)", {'args': args},
                  1e-10, id="string")
@@ -154,7 +155,6 @@ def test_CoeffCreationCall(base, kwargs, tol):
     expected = lambda t: np.exp(1j * t * np.pi)
     coeff = coefficient(base, **kwargs, compile_opt=opt)
     _assert_eq_over_interval(coeff, expected, rtol=tol, inside=True)
-
 
 
 @pytest.mark.parametrize(['base', 'kwargs', 'tol'], [
@@ -327,7 +327,7 @@ from qutip import basis
 from qutip.core.data cimport CSR
 from qutip.core.data.expect cimport expect_csr
 """)
-    csr = qt.num(3).data
+    csr = qutip.num(3).data
     coeff = coefficient("expect_csr(op, op)",
                         args={"op": csr},
                         args_ctypes={"op": "CSR"},
@@ -397,3 +397,36 @@ def test_Coeffcopy(style, transform):
     coeff = transform(coeff)
     coeff_cp = coeff.copy()
     _assert_eq_over_interval(coeff, coeff_cp)
+
+
+@pytest.mark.parametrize('order', [0, 1, 2, 3])
+def test_CoeffArray(order):
+    tlist = np.linspace(0, 1, 101)
+    y = np.exp((-1 + 1j) * tlist)
+    coeff = coefficient(y, tlist=tlist, order=order)
+    expected = coefficient(lambda t: np.exp((-1 + 1j) * t))
+    _assert_eq_over_interval(coeff, expected, rtol=0.01**(order+1))
+    dt = 1e-6
+    t = 0.025
+    derr = (coeff(t+dt) - coeff(t-dt)) / (2*dt)
+    deff2 = (coeff(t+dt) + coeff(t-dt) - 2 * coeff(t)) / (dt**2)
+    deff3 = (coeff(t + 2*dt) - 2*coeff(t + dt)
+             + 2*coeff(t + dt) -coeff(t - 2*dt)) / (12 * dt**3)
+    derrs = [derr, derr2, derr3]
+    for i in range(order):
+        assert derrs[i] != 0
+    for i in range(order, 4):
+        assert derrs[i] == pytest.approx(0.0)
+
+
+def test_CoeffFromScipy():
+    tlist = np.linspace(0, 1, 101)
[…]
