### 2021-11-18 adc2743
add new tests
