### 2021-11-19 c93c02c
Lowered the version to be the same as the runtime version in setup.cfg
I previously changed the version to match the build requirements. It
should instead use the runtime requirements.
