### 2021-11-19 7a73eca
Changed tol for atol.
tol was referring to absolut tolerance but with the inclusion of relative
tolerance, `rtol` it makes more sense to name it atol instead of tol to
be explicit and avoid confusion.
