+++ qutip/tests/core/data/test_mathematics.py
@@ -297,7 +297,7 @@ class _GenericOpMixin:
         but with all data arguments numpy arrays, and returns the expected
         result.
 
-    tol: float
+    atol: float
         The absolute tolerance to use when comparing the test value with the
         expected value.  If the output is a Data type, the tolerance is
         per-element of the output.
@@ -327,7 +327,7 @@ class _GenericOpMixin:
     # addition not being associative; the maths on full numpy arrays will often
     # produce slightly different results to sparse algebra, since the order of
     # multiplications and additions will be different.
-    tol = 1e-10
+    atol = 1e-10
     rtol = 1e-7  # Same default as numpy
     shapes = []
     bad_shapes = []
@@ -406,16 +406,16 @@ class UnaryOpMixin(_GenericOpMixin):
         if issubclass(out_type, Data):
             assert test.shape == expected.shape
             np.testing.assert_allclose(test.to_array(), expected,
-                                       atol=self.tol,
+                                       atol=self.atol,
                                        rtol=self.rtol)
         elif out_type is list:
             for test_, expected_ in zip(test, expected):
                 assert test_.shape == expected_.shape
                 np.testing.assert_allclose(test_.to_array(),
-                                           expected_, atol=self.tol,
+                                           expected_, atol=self.atol,
                                            rtol=self.rtol)
         else:
-            assert abs(test - expected) < self.tol + self.rtol*abs(expected)
+            assert abs(test - expected) < self.atol + self.rtol*abs(expected)
 
     def test_incorrect_shape_raises(self, op, data_m):
         """
@@ -449,10 +449,10 @@ class UnaryScalarOpMixin(_GenericOpMixin):
         if issubclass(out_type, Data):
             assert test.shape == expected.shape
             np.testing.assert_allclose(test.to_array(), expected,
-                                       atol=self.tol,
+                                       atol=self.atol,
                                        rtol=self.rtol)
         else:
-            assert abs(test - expected) < self.tol + self.rtol*abs(expected)
+            assert abs(test - expected) < self.atol + self.rtol*abs(expected)
 
 
 class BinaryOpMixin(_GenericOpMixin):
@@ -472,10 +472,10 @@ class BinaryOpMixin(_GenericOpMixin):
         if issubclass(out_type, Data):
             assert test.shape == expected.shape
             np.testing.assert_allclose(test.to_array(), expected,
-                                       atol=self.tol,
+                                       atol=self.atol,
                                        rtol=self.rtol)
         else:
-            assert abs(test - expected) < self.tol + self.rtol*abs(expected)
+            assert abs(test - expected) < self.atol + self.rtol*abs(expected)
 
     def test_incorrect_shape_raises(self, op, data_l, data_r):
         """
@@ -507,10 +507,10 @@ class TernaryOpMixin(_GenericOpMixin):
         if issubclass(out_type, Data):
             assert test.shape == expected.shape
             np.testing.assert_allclose(test.to_array(), expected,
-                                       atol=self.tol,
+                                       atol=self.atol,
                                        rtol=self.rtol)
         else:
-            assert abs(test - expected) < self.tol + self.rtol*abs(expected)
+            assert abs(test - expected) < self.atol + self.rtol*abs(expected)
 
     def test_incorrect_shape_raises(self, op, data_l, data_m, data_r):
         """
@@ -555,10 +555,10 @@ class TestAdd(BinaryOpMixin):
         if issubclass(out_type, Data):
             assert test.shape == expected.shape
             np.testing.assert_allclose(test.to_array(), expected,
-                                       atol=self.tol,
+                                       atol=self.atol,
                                        rtol=self.rtol)
[…]
