### 2021-11-19 ad2fdc7
Changed assertions of scalar check to use same function as assertions of matrix checks.
Instead of checking scalars using our own expression we use
np.testing.assert_allclose for consistency in the code.
