### 2021-11-19 ee19b3f
Merge pull request #1715 from AGaliciaMartinez/doc_contribution_update
Added packaging to the installation.rst requirements.
