### 2021-11-21 8e7cfc5
- Add warning about behavior change of process_fidelity in 5.0 - Refactor tests of process_fidelity using pytest.mark.parametrize - In one branch of process_fidelity, divide fidelity by dimensional factor instead of dividing density matrices
