### 2021-11-22 848a65d
Improve description of delta (the approximation discrepancy).
