### 2021-11-22 5159416
Remove BosonicHEOMSolver and FermionicHEOMSolver -- these are redundant now that we have the BosonicBath and FermionicBath.
