+++ qutip/tests/test_bofin_solvers.py
@@ -26,8 +26,6 @@ from qutip.nonmarkov.bofin_solvers import (
     HierarchyADOs,
     HierarchyADOsState,
     HEOMSolver,
-    BosonicHEOMSolver,
-    FermionicHEOMSolver,
     HSolverDL,
     _GatherHEOMRHS,
 )
@@ -299,376 +297,9 @@ class UnderdampedPureDephasingModel:
         ]
 
 
-_HAMILTONIAN_EVO_KINDS = {
-    "qobj": lambda H: H,
-    "qobjevo": lambda H: QobjEvo([H]),
-    "listevo": lambda H: [H],
-}
-
-
-def hamiltonian_to_sys(H, evo, liouvillianize):
-    if liouvillianize:
-        H = liouvillian(H)
-    H = _HAMILTONIAN_EVO_KINDS[evo](H)
-    return H
-
-
-class TestHEOMSolver:
-    def test_create_bosonic(self):
-        Q = sigmaz()
-        H = sigmax()
-        exponents = [
-            BathExponent("R", None, Q=Q, ck=1.1, vk=2.1),
-            BathExponent("I", None, Q=Q, ck=1.2, vk=2.2),
-            BathExponent("RI", None, Q=Q, ck=1.3, vk=2.3, ck2=3.3),
-        ]
-        bath = Bath(exponents)
-
-        hsolver = HEOMSolver(H, bath, 2)
-        assert hsolver.ados.exponents == exponents
-        assert hsolver.ados.max_depth == 2
-
-        hsolver = HEOMSolver(H, [bath] * 3, 2)
-        assert hsolver.ados.exponents == exponents * 3
-        assert hsolver.ados.max_depth == 2
-
-    def test_create_fermionic(self):
-        Q = sigmaz()
-        H = sigmax()
-        exponents = [
-            BathExponent("+", 2, Q=Q, ck=1.1, vk=2.1, sigma_bar_k_offset=1),
-            BathExponent("-", 2, Q=Q, ck=1.2, vk=2.2, sigma_bar_k_offset=-1),
-        ]
-        bath = Bath(exponents)
-
-        hsolver = HEOMSolver(H, bath, 2)
-        assert hsolver.ados.exponents == exponents
-        assert hsolver.ados.max_depth == 2
-
-        hsolver = HEOMSolver(H, [bath] * 3, 2)
-        assert hsolver.ados.exponents == exponents * 3
-        assert hsolver.ados.max_depth == 2
-
-    def test_create_progress_bar(self):
-        Q = sigmaz()
-        H = sigmax()
-        bath = Bath([
-            BathExponent("R", None, Q=Q, ck=1.1, vk=2.1),
-        ])
-
-        hsolver = HEOMSolver(H, bath, 2)
-        assert isinstance(hsolver.progress_bar, BaseProgressBar)
-
-        hsolver = HEOMSolver(H, bath, 2, progress_bar=True)
-        assert isinstance(hsolver.progress_bar, TextProgressBar)
-
-    def test_create_bath_errors(self):
-        Q = sigmaz()
-        H = sigmax()
-        mixed_types = [
-            BathExponent("+", 2, Q=Q, ck=1.1, vk=2.1, sigma_bar_k_offset=1),
-            BathExponent("-", 2, Q=Q, ck=1.2, vk=2.2, sigma_bar_k_offset=-1),
-            BathExponent("R", 2, Q=Q, ck=1.2, vk=2.2),
-        ]
-        mixed_q_dims = [
-            BathExponent("I", 2, Q=tensor(Q, Q), ck=1.2, vk=2.2),
-            BathExponent("R", 2, Q=Q, ck=1.2, vk=2.2),
-        ]
-
-        with pytest.raises(ValueError) as err:
[…]
