### 2021-11-22 2e7ee5d
Add better descriptions of bath coefficients.
