### 2021-11-22 51f8ae7
Add checks for successful integration.
