### 2021-11-22 d573f67
optimize state copies
