+++ qutip/tests/solver/test_mesolve.py
@@ -29,8 +29,6 @@ class TestMESolveDecay:
     @pytest.fixture(params=[
         pytest.param([ada, lambda t, args: 1], id='Hlist_func'),
         pytest.param([ada, '1'], id='Hlist_str'),
-        pytest.param([ada, qutip.Cubic_Spline(0, 10, np.ones_like(tlist))],
-                      id='Hlist_cubic_spline'),
         pytest.param([ada, np.ones_like(tlist)], id='Hlist_array'),
         pytest.param(qutip.QobjEvo([ada, '1']), id='HQobjEvo'),
         pytest.param(lambda t, args: qutip.create(10) * qutip.destroy(10),
@@ -49,10 +47,6 @@ class TestMESolveDecay:
                      id='list_func'),
         pytest.param([a, 'sqrt(kappa)'],
                      id='list_str'),
-        pytest.param([a, qutip.Cubic_Spline(0, 10,
-                             np.sqrt(kappa) * np.ones_like(tlist)
-                         )],
-                     id='list_cubic_spline'),
         pytest.param([a, np.sqrt(kappa) * np.ones_like(tlist)],
                      id='list_array'),
         pytest.param(qutip.QobjEvo([a, 'sqrt(kappa)'], args={'kappa': kappa}),
@@ -66,9 +60,6 @@ class TestMESolveDecay:
                   id='list_func'),
         pytest.param([a, 'sqrt(kappa * exp(-t))'],
                   id='list_str'),
-        pytest.param([a, qutip.Cubic_Spline(0, 10,
-                                            np.sqrt(kappa * np.exp(-tlist)))],
-                  id='list_cubic_spline'),
         pytest.param([a, np.sqrt(kappa * np.exp(-tlist))],
                   id='list_array'),
         pytest.param(qutip.QobjEvo([a, 'sqrt(kappa * exp(-t))'],
