+++ qutip/tests/core/data/test_norm.py
@@ -6,6 +6,8 @@ from qutip import data
 from qutip.core.data import CSR, Dense
 import numbers
 
+__all__ = ['TestOneNorm', 'TestFrobeniusNorm', 'TestMaxNorm', 'TestL2Norm',
+          'TestTraceNorm']
 
 class TestOneNorm(testing.UnaryOpMixin):
     def op_numpy(self, matrix):
