+++ qutip/tests/core/data/__init__.py
@@ -1,0 +1,4 @@
+from .test_expect import *
+from .test_mathematics import *
+from .test_norm import *
+from .test_reshape import *
