### 2021-11-23 45692ca
Add link to HEOM from guide index.
