+++ doc/guide/guide-heom.rst
@@ -1,0 +1,16 @@
+.. _heom:
+
+********************************
+Hierarchical Equations of Motion
+********************************
+
+.. toctree::
+   :maxdepth: 2
+
+   heom/index.rst
+   heom/intro.rst
+   heom/bosonic.rst
+   heom/fermionic.rst
+   heom/dynamics.rst
+   heom/steadystate.rst
+   heom/zzreferences.rst
