### 2021-11-23 6935321
Remove bibliography into main heom guide folder.
