+++ qutip/tests/core/data/test_reshape.py
@@ -5,7 +5,8 @@ from qutip import data
 from qutip.core.data import CSR, Dense
 
 __all__ = ['TestSplitColumns', 'TestColumnStack', 'TestColumnUnstack',
-           'TestReshape']
+           'TestReshape', ]
+
 
 class TestSplitColumns(UnaryOpMixin):
     def op_numpy(self, matrix):
