### 2021-11-23 7366a72
Add sphinxcontrib-bibtex for heom guide.
