+++ doc/guide/heom/heom.bib
@@ -1,5 +1,5 @@
 
-  
+
 @article{lambert2019virtual,
   title={Virtual excitations in the ultra-strongly-coupled spin-boson model: physical results from unphysical modes},
   author={Lambert, Neill and Ahmed, Shahnawaz and Cirio, Mauro and Nori, Franco},
