### 2021-11-23 ced5b69
Clean up overall HEOM guide structure and start updating bosonic example.
