+++ doc/guide/guide-heom.rst
@@ -7,10 +7,7 @@ Hierarchical Equations of Motion
 .. toctree::
    :maxdepth: 2
 
-   heom/index.rst
    heom/intro.rst
    heom/bosonic.rst
    heom/fermionic.rst
-   heom/dynamics.rst
-   heom/steadystate.rst
-   heom/zzreferences.rst
+   heom/references.rst
