### 2021-11-23 c441472
simpler coefficient function
