### 2021-11-24 e024f59
Don't show the plot download links in the source because it's not very readable.
