### 2021-11-24 0184318
Flesh out the bosonic guide.
