+++ doc/guide/heom/bosonic.rst
@@ -3,14 +3,14 @@ Bosonic Environments
 ####################
 
 In this section we consider a simple two-level system coupled to a
-Drude-Lorentz bosonic bath. The system Hamiltonian, $H_{sys}$, and the bath
-spectral density, $J_D$, are
+Drude-Lorentz bosonic bath. The system Hamiltonian, :math:`H_{sys}`, and the bath
+spectral density, :math:`J_D`, are
 
 .. math::
 
-    H_{sys} = \frac{\epsilon \sigma_z}{2} + \frac{\Delta \sigma_x}{2}
+    H_{sys} &= \frac{\epsilon \sigma_z}{2} + \frac{\Delta \sigma_x}{2}
 
-    J_D = \frac{2\lambda \gamma \omega}{(\gamma^2 + \omega^2)},
+    J_D &= \frac{2\lambda \gamma \omega}{(\gamma^2 + \omega^2)},
 
 We will demonstrate how to describe the bath using two different expansions
 of the spectral density correlation function (Matsubara's expansion and
@@ -25,21 +25,20 @@ Afterwards, we will show how to calculate the bath expansion coefficients and to
 use those coefficients to construct your own bath description so that you can
 implement your own bosonic baths.
 
-A notebook containing a complete example similar to this one can be found in
-`example notebook 1a <FIXME>`__).
-
-.. todo::
-
-    Fix notebook link above.
+A notebook containing a complete example similar to this one implemented in
+BoFiN can be found in
+`example notebook 1a <https://github.com/tehruhn/bofin/blob/main/examples/example-1a-Spin-bath-model-basic.ipynb>`__).
 
 
 Describing the system and bath
 ------------------------------
 
-First, let us construct the system Hamiltonian, $H_{sys}$, and the initial
-system state, $rho0$:
+First, let us construct the system Hamiltonian, :math:`H_{sys}`, and the initial
+system state, ``rho0``:
 
-.. code-block:: python
+.. plot::
+    :context: reset
+    :nofigs:
 
     from qutip import basis, sigmax, sigmaz
 
@@ -53,7 +52,9 @@ system state, $rho0$:
 
 Now let us describe the bath properties:
 
-.. code-block:: python
+.. plot::
+    :context:
+    :nofigs:
 
     # Bath properties:
     gamma = 0.5  # cut off frequency
@@ -63,14 +64,16 @@ Now let us describe the bath properties:
     # System-bath coupling operator:
     Q = sigmaz()
 
-where $\gamma$ (``gamma``), $\lambda$ (``lam``) and $T$ are the parameters
-of a Drude-Lorentz bath, and ``Q`` is the coupling operator between the system
-and the bath.
+where :math:`\gamma` (``gamma``), :math:`\lambda` (``lam``) and :math:`T` are
+the parameters of a Drude-Lorentz bath, and ``Q`` is the coupling operator
+between the system and the bath.
 
 We may the pass these parameters to either ``DrudeLorentzBath`` or
 ``DrudeLorentzPadeBath`` to construct an expansion of the bath correlations:
 
-.. code-block:: python
+.. plot::
+    :context:
+    :nofigs:
 
     from qutip.nonmarkov.heom import DrudeLorentzBath
     from qutip.nonmarkov.heom import DrudeLorentzPadeBath
@@ -93,7 +96,9 @@ System and bath dynamics
 
 Now we are ready to construct a solver:
 
-.. code-block:: python
+.. plot::
+    :context:
+    :nofigs:
 
     from qutip.nonmarkov.heom import HEOMSolver
     from qutip import Options
@@ -125,7 +130,8 @@ The ``result`` is a standard QuTiP results object with the attributes:
[…]
