### 2021-11-25 35fdbc4
Add an example of using multiple baths.
