+++ doc/guide/heom/bosonic.rst
@@ -45,7 +45,7 @@ system state, ``rho0``:
     # The system Hamiltonian:
     eps = 0.5  # energy of the 2-level system
     Del = 1.0  # tunnelling term
-    H_sys = 0.5 * eps * sigmaz() + 0.5 * Del* sigmax()
+    H_sys = 0.5 * eps * sigmaz() + 0.5 * Del * sigmax()
 
     # Initial state of the system:
     rho0 = basis(2,0) * basis(2,0).dag()
@@ -307,6 +307,8 @@ After all that, constructing the bath is very straight forward:
     :context:
     :nofigs:
 
+    from qutip.nonmarkov.heom import BosonicBath
+
     bath = BosonicBath(Q, ck_real, vk_real, ck_imag, vk_imag)
 
 And we're done!
@@ -319,60 +321,69 @@ Drude-Lorentz bath expansions.
 Multiple baths
 --------------
 
-.. todo::
+The :class:`HEOMSolver` supports having a system interact with multiple
+environments. All that is needed is to supply a list of baths instead
+of a singe bath.
 
-    Clean up this section to use the new multiple baths feature.
+In the example below we calculate the evolution of a small system where
+each energy eigenstate of the system interacts with a separate bath.
 
+For each bath expansion, we also include the terminator in the system
+Liouvillian.
 
-The above example describes a single environment parameterized by the lists of
-coefficients and frequencies in the correlation functions.
+At the end, we plot the populations of the system states as a function of
+time:
 
-For multiple environments, the list of coupling operators and bath properties
-must all be extended in a particular way.  Note this functionality differs in
-the case of the Fermionic solver.
+.. plot::
+    :context: close-figs
 
-For the Bosonic solver, for ``N`` baths, each ``ckAR``, ``vkAR``, ``ckAI``, and
-``vkAI`` are extended ``N`` times with the appropriate number of terms of that
-bath.
+    # The size of the system:
+    N_sys = 3
 
-On the other hand, the list of coupling operators is defined in such a way that
-the terms corresponding to the real cooefficients are **given first**, and the
-imaginary terms after. Thus if each bath has :math:`N_k` coefficients, the list
-of coupling operators is of length :math:`N_k \times (N_R + N_I)`.
+    def proj(i, j):
+        """ A helper function for creating an interaction operator. """
+        return basis(N_sys, i) * basis(N_sys, j).dag()
 
-This is best illustrated by the example in `example notebook 2
-<https://github.com/tehruhn/bofin/blob/main/examples/example-2-FMO-example.ipynb>`_.
-In that case each bath is identical, and there are seven baths, each with a
-unique coupling operator defined by a projector onto a single state:
+    # Construct one bath for each system state:
+    baths = []
+    for i in range(N_sys):
+        Q = proj(i, i)
+        baths.append(DrudeLorentzBath(Q, lam, T, Nk, gamma, terminator=True))
 
-.. code-block:: python
+    # Construct the system Liouvillian from the system Hamiltonian and
+    # bath expansion terminators:
+    H_sys = sum((i + 0.5) * eps * proj(i, i) for i in range(N_sys))
+    H_sys += sum(
+      (i + j + 0.5) * Del * proj(i, j)
+      for i in range(N_sys) for j in range(N_sys)
+      if i != j
+    )
[…]
