### 2021-11-25 72887fe
Improve the description of the system used for multiple baths.
