### 2021-11-25 cbdb52e
Merge pull request #1704 from fhopfmueller/master
Only skip the dnorm tests if cvxpy is not installed (previously all metrics tests were skipped).
