### 2021-11-25 d2225c3
Add first draft of lorentzian bath.
