### 2021-11-25 4a08d47
Small improvements to the bosonic guide.
