### 2021-11-25 39b3a5c
Made a start on the fermionic guide.
