+++ doc/guide/heom/fermionic.rst
@@ -2,152 +2,386 @@
 Fermionic Environments
 ######################
 
-The basic class object used to construct the problem is imported in the following way (alongside QuTiP, with which we define system Hamiltonian and coupling operators), for the pure Python BoFiN version
+Here we model a single fermion coupled to two electronic leads or reservoirs (e.g.,  this can describe a single quantum dot, a molecular transistor, etc).  Note that in this implementation we primarily follow the definitions used by Christian Schinabeck in his Dissertation https://opus4.kobv.de/opus4-fau/files/10984/DissertationChristianSchinabeck.pdf and related publications.
 
-.. code-block:: python
-
-    from qutip import *
-    from bofin.heom import FermionicHEOMSolver
-    
-If one is using the C++ BoFiN-fast package, the import is instead
-
-.. code-block:: python
-
-    from qutip import *
-    from bofinfast.heom import FermionicHEOMSolver
-    
-    
-Apart from this difference in import, and some additional features in the solvers in the C++ variant, the functionality that follows applies to both libraries.
-
-One defines a particular problem instance in the following way:
-
-.. code-block:: python
-
-    Solver = FermionicHEOMSolver(Hsys, Q,  eta_list, gamma_list, NC,o ptions=options)
+A notebook containing a complete example similar to this one implemented in
+BoFiN can be found in `example notebook 4b <https://github.com/tehruhn/bofin/blob/main/examples/example-4b-fermions-single-impurity-model.ipynb>`__.
 
 
-The parameters accepted by the solver are :
+Describing the system and bath
+------------------------------
 
-- ``Hsys`` : system Hamiltonian in quantum object form
-- ``Q`` : a list of coupling operators (minimum two) that couple the system to the environment
-- ``eta_list`` and ``gamma_list`` : the coefficients and frequencies of the correlation functions of the environment
-- ``NC`` : the truncation parameter of the hierarchy
-- ``options`` : standard QuTiP ``ODEoptions`` object, which is used by the ODE solver.
+First, let us construct the system Hamiltonian, :math:`H_{sys}`, and the initial
+system state, ``rho0``:
 
-Note that for the Fermionic case, unlike the Bosonic case, we **don't** explicitly split the real and imaginary parts of the correlation functions. Also, for the Fermionic case, a single environment has a minimum of two coupling operators (due to the fundemental difference in how the environment interacts with the system for fermionic environments).  
+.. plot::
+    :context: reset
+    :nofigs:
 
-A simple example of how this code works, taken from `example notebook 4b <https://github.com/tehruhn/bofin/blob/main/examples/example-4b-fermions-single-impurity-model.ipynb>`_, is:
+    from qutip import basis, destroy
 
-.. code-block:: python
+    # The system Hamiltonian:
+    e1 = 1.  # site energy
+    H_sys = e1 * destroy(2).dag() * destroy(2)
 
-    %pylab inline
-    from qutip import *
-    from bofin.heom import FermionicHEOMSolver
+    # Initial state of the system:
[…]
