### 2021-11-26 81e082a
Add a compile time option for 64 bit indices.
