+++ setup.py
@@ -72,12 +72,12 @@ def _determine_user_arguments(options):
     )
     if "--with-openmp" in sys.argv:
         sys.argv.remove("--with-openmp")
-    options['intidx_64'] = (
-        '--with-intidx-64' in sys.argv
-        or _get_environment_bool('CI_QUTIP_WITH_INTIDX_64')
+    options['idxint_64'] = (
+        '--with-idxint-64' in sys.argv
+        or _get_environment_bool('CI_QUTIP_WITH_IDXINT_64')
     )
-    if "--with-intidx-64" in sys.argv:
-        sys.argv.remove("--with-intidx-64")
+    if "--with-idxint-64" in sys.argv:
+        sys.argv.remove("--with-idxint-64")
     return options
 
 
@@ -234,7 +234,7 @@ def create_extension_modules(options):
                              extra_link_args=options['ldflags'],
                              language='c++'))
     compile_time_env = {
-        "QUTIP_INTIDX_64": options["intidx_64"],
+        "QUTIP_IDXINT_64": options["idxint_64"],
     }
     return cythonize(out, compile_time_env=compile_time_env)
 
