### 2021-11-26 2c707c7
Slightly clean-up option parsing.
