### 2021-11-26 ff02059
Slightly clean-up option parsing.
