### 2021-11-26 3082497
Pass QUTIP_IDXINT_64 when compiling coefficients.
