### 2021-11-27 8dafaca
Add whitespace in qutip/core/metrics.py
Co-authored-by: Simon Cross <hodgestar+github@gmail.com>
