### 2021-11-27 d48e85d
Add whitespace in qutip/core/metrics.py
Co-authored-by: Simon Cross <hodgestar+github@gmail.com>
