### 2021-11-27 d6aab1e
Add whitespace in qutip/core/metrics.py
Co-authored-by: Simon Cross <hodgestar+github@gmail.com>
