+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,20 +1,24 @@
 ---
-name: Feature request
-about: Suggest an idea for this project
+name: Feature Request
+about: Suggest an idea for this project!
 title: ''
 labels: 'ENH'
 assignees: ''
 
 ---
 
-**Is your feature request related to a problem? Please describe.**
-A clear and concise description of what the problem is. Ex. I'm always frustrated when [...]
+**Problem Description**
+<!-- A clear and concise description of what the problem is. Ex. It would be better if [...] -->
 
-**Describe the solution you'd like**
-A clear and concise description of what you want to happen.
 
-**Describe alternatives you've considered**
-A clear and concise description of any alternative solutions or features you've considered.
+**Proposed Solution**
+<!-- A clear and concise description of what you want to happen -->
+
+
+**Alternate Solutions**
+<!-- A clear and concise description of any alternative solutions or features you've considered -->
+
 
 **Additional context**
-Add any other context or screenshots about the feature request here.
+<!-- Add any other context or screenshots about the feature request here -->
+
