+++ .github/ISSUE_TEMPLATE/others.md
@@ -8,16 +8,18 @@ assignees: ''
 ---
 
 **Describe the issue**
-Please describe the issue here.
+<!-- Please describe the issue here -->
 
 **To Reproduce**
-If applicable, Please provide a minimal working example. For instance:
+<!-- If applicable, Please provide a minimal working example. For instance: -->
 
+Code -  
 ```python
 from qutip import identity
 print(identity(2))
 ```
-The terminal out put is
+
+Output -  
 ```
 Quantum object: dims = [[2], [2]], shape = (2, 2), type = oper, isherm = True
 Qobj data =
