+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,23 +1,26 @@
 ---
-name: Bug report
-about: Create a report to help us improve
+name: Bug Report
+about: Spotted a bug? Report it to us!
 title: ''
 labels: 'BUG'
 assignees: ''
 
 ---
 
-**Describe the bug**
-A clear and concise description of what the bug is.
+**Bug Description**
+<!-- A clear and concise description of what the bug is -->
+
 
 **To Reproduce**
-Please provide a minimal working example. For instance:
+<!-- Please provide a minimal working example. For instance: -->
 
+Code -  
 ```python
 from qutip import identity
 print(identity(2))
 ```
-The terminal out put is
+
+Output -  
 ```
 Quantum object: dims = [[2], [2]], shape = (2, 2), type = oper, isherm = True
 Qobj data =
@@ -25,11 +28,16 @@ Qobj data =
  [0. 1.]]
 ```
 
-**Expected behavior**
-A clear and concise description of what you expected to happen.
+
+**Expected Behavior**
+<!-- A clear and concise description of what you expected to happen -->
+
 
 **Your Environment**
-Please use `qutip.about()` to get the information about your environment and paste it here.
+<!-- Please use `qutip.about()` to get the information about your environment and paste it here. -->
+Output of `qutip.about()` -  
+
 
 **Additional context**
-Add any other context about the problem here.
+<!-- Add any other context about the problem here -->
+
