### 2021-11-29 58efa02
updated 'your environment' section
added the expected output of qutip.about() in bug_report.md
