### 2021-11-29 1cc0307
Update the epsilon and kappa LorentzianPadeBath calculations to be more similar to the existing DrudeLorentzPadeBath versions.
