### 2021-11-29 f316ef9
Update LorentzianBath and LorentzianPadeBath to only calculate the coefficients for a single bath.
