+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,57 +1,0 @@
----
-name: Bug Report
-about: Spotted a bug? Report it to us!
-title: ''
-labels: 'BUG'
-assignees: ''
-
----
-
-**Bug Description**
-<!-- A clear and concise description of what the bug is -->
-
-
-**To Reproduce**
-<!-- Please provide a minimal working example. For instance: -->
-
-Code -  
-```python
-from qutip import identity
-print(identity(2))
-```
-
-Output -  
-```
-Quantum object: dims = [[2], [2]], shape = (2, 2), type = oper, isherm = True
-Qobj data =
-[[1. 0.]
- [0. 1.]]
-```
-
-
-**Expected Behavior**
-<!-- A clear and concise description of what you expected to happen -->
-
-
-**Your Environment**
-<!-- Please use `qutip.about()` to get the information about your environment and paste it here. -->
-Output of `qutip.about()` -  
-
-```
-QuTiP Version:      ***
-Numpy Version:      ***
-Scipy Version:      ***
-Cython Version:     ***
-Matplotlib Version: ***
-Python Version:     ***
-Number of CPUs:     ***
-BLAS Info:          ***
-OPENMP Installed:   ***
-INTEL MKL Ext:      ***
-Platform Info:      ***
-```
-
-
-**Additional context**
-<!-- Add any other context about the problem here -->
-
