+++ .github/ISSUE_TEMPLATE/others.md
@@ -1,28 +1,0 @@
----
-name: Others
-about: Other issues
-title: ''
-labels: ''
-assignees: ''
-
----
-
-**Describe the issue**
-<!-- Please describe the issue here -->
-
-**To Reproduce**
-<!-- If applicable, Please provide a minimal working example. For instance: -->
-
-Code -  
-```python
-from qutip import identity
-print(identity(2))
-```
-
-Output -  
-```
-Quantum object: dims = [[2], [2]], shape = (2, 2), type = oper, isherm = True
-Qobj data =
-[[1. 0.]
- [0. 1.]]
-```
