+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,24 +1,0 @@
----
-name: Feature Request
-about: Suggest an idea for this project!
-title: ''
-labels: 'ENH'
-assignees: ''
-
----
-
-**Problem Description**
-<!-- A clear and concise description of what the problem is. Ex. It would be better if [...] -->
-
-
-**Proposed Solution**
-<!-- A clear and concise description of what you want to happen -->
-
-
-**Alternate Solutions**
-<!-- A clear and concise description of any alternative solutions or features you've considered -->
-
-
-**Additional context**
-<!-- Add any other context or screenshots about the feature request here -->
-
