+++ qutip/tests/solver/test_brmesolve.py
@@ -1,0 +1,294 @@
+import pytest
+import numpy as np
+import qutip
+from qutip.solver.brmesolve import brmesolve
+
+
+def pauli_spin_operators():
+    return [qutip.sigmax(), qutip.sigmay(), qutip.sigmaz()]
+
+
+_simple_qubit_gamma = 0.25
+coeff = qutip.coefficient(lambda t, w: _simple_qubit_gamma * (w >= 0),
+                          args={'w':0})
+_m_c_op = np.sqrt(_simple_qubit_gamma) * qutip.sigmam()
+_z_c_op = np.sqrt(_simple_qubit_gamma) * qutip.sigmaz()
+_x_a_op = [qutip.sigmax(), coeff]
+
+
+@pytest.mark.parametrize("me_c_ops, brme_c_ops, brme_a_ops", [
+    pytest.param([_m_c_op], [], [_x_a_op], id="me collapse-br coupling"),
+    pytest.param([_m_c_op], [_m_c_op], [], id="me collapse-br collapse"),
+    pytest.param([_m_c_op, _z_c_op], [_z_c_op], [_x_a_op],
+                 id="me collapse-br collapse-br coupling"),
+])
+def test_simple_qubit_system(me_c_ops, brme_c_ops, brme_a_ops):
+    """
+    Test that the BR solver handles collapse and coupling operators correctly
+    relative to the standard ME solver.
+    """
+    delta = 0.0
+    epsilon = 0.5 * 2 * np.pi
+    e_ops = pauli_spin_operators()
+    H = delta * 0.5 * qutip.sigmax() + epsilon * 0.5 * qutip.sigmaz()
+    psi0 = (2 * qutip.basis(2, 0) + qutip.basis(2, 1)).unit()
+    times = np.linspace(0, 10, 100)
+    me = qutip.mesolve(H, psi0, times, c_ops=me_c_ops, e_ops=e_ops).expect
+    brme = brmesolve(H, psi0, times, brme_a_ops,
+                     e_ops=e_ops, c_ops=brme_c_ops).expect
+    for me_expectation, brme_expectation in zip(me, brme):
+        np.testing.assert_allclose(me_expectation, brme_expectation, atol=1e-2)
+
+
+def _harmonic_oscillator_spectrum_frequency(n_th, w0, kappa):
+    if n_th == 0:
+        return lambda w: kappa * (w >= 0)
+
+    w_th = w0 / np.log(1 + 1/n_th)
+
+    def f(t, w):
+        scale = np.exp(w / w_th) if w < 0 else 1
+        return (n_th + 1) * kappa * scale
+    return f
+
+
+def _harmonic_oscillator_c_ops(n_th, kappa, dimension):
+    a = qutip.destroy(dimension)
+    if n_th == 0:
+        return [np.sqrt(kappa) * a]
+    return [np.sqrt(kappa * (n_th+1)) * a, np.sqrt(kappa * n_th) * a.dag()]
+
+
+@pytest.mark.parametrize("n_th", [0, 0.15])
+def test_harmonic_oscillator(n_th):
+    N = 10
+    w0 = 1.0 * 2*np.pi
+    g = 0.05 * w0
+    kappa = 0.15
+    S_w = _harmonic_oscillator_spectrum_frequency(n_th, w0, kappa)
+
+    a = qutip.destroy(N)
+    H = w0*a.dag()*a + g*(a+a.dag())
+    psi0 = (qutip.basis(N, 4) + qutip.basis(N, 2) + qutip.basis(N, 0)).unit()
+    psi0 = qutip.ket2dm(psi0)
[…]
