### 2021-11-30 02c0fb2
Fix typo in docstring of chi_to_choi.
