### 2021-12-01 024dbff
Remove reference to Pauli basis in chi_to_choi docstring.
