### 2021-12-01 e4e9402
Merge pull request #1731 from hodgestar/fix/chi-to-choi-docstring
Fix typo in docstring of chi_to_choi.
