### 2021-12-01 dc0c62a
Apply suggestions from code review
Co-authored-by: Asier Galicia <57414022+AGaliciaMartinez@users.noreply.github.com>
