+++ qutip/solver/multitraj.py
@@ -32,14 +32,11 @@ class MultiTrajSolver:
     optionsclass = SolverOptions
     odeoptionsclass = SolverOdeOptions
 
-    def __init__(self, rhs, c_ops=(), *, options=None):
+    def __init__(self, rhs, *, options=None):
         self.rhs = rhs
-        self._c_ops = c_ops
         self.stats = {}
         self.options = options
         self.seed_sequence = np.random.SeedSequence()
-        self.traj_solver = False
-        self.result = None
 
     def _read_seed(self, seed, ntraj):
         """
@@ -62,14 +59,14 @@ class MultiTrajSolver:
             raise ValueError("A seed list must be longer than ntraj")
         return seeds
 
-    def start(self, state0, t0, ntraj=1, seed=None, *, safe_ODE=None):
+    def start(self, state, t0, ntraj=1, seed=None, *, safe_ODE=None):
         """
         Set the initial state and time for a step evolution.
         ``options`` for the evolutions are read at this step.
 
         Parameters
         ----------
-        state0 : :class:`Qobj`
+        state : :class:`Qobj`
             Initial state of the evolution.
 
         t0 : double
@@ -89,6 +86,11 @@ class MultiTrajSolver:
             Whether to safe the states in the ODE solver or in the solver.
             Many ODE solver are not re-entrant, thus must be used with
             `safe_ODE` if multiple trajectory are to be ran in parallel.
+
+        .. note ::
+            Using the solver with ``start``, ``step`` is independent to use
+            with ``run``. Calling ``run`` between ``step`` will not affect
+            ``step``'s result.
         """
         seeds = self._read_seed(seed, ntraj)
         self.traj_solvers = []
@@ -97,8 +99,9 @@ class MultiTrajSolver:
             safe_ODE = ntraj != 1
 
         for seed in seeds:
-            traj_solver = self._traj_solver_class(self, options=self.options)
-            traj_solver.start(state0, t0, seed=seed, safe_ODE=safe_ODE)
+            traj_solver = self._traj_solver_class(*self.traj_args,
+                                                  options=self.options)
+            traj_solver.start(state, t0, seed=seed, safe_ODE=safe_ODE)
             self.traj_solvers.append(traj_solver)
 
     def step(self, t, *, args=None, options=None, copy=True):
@@ -122,6 +125,11 @@ class MultiTrajSolver:
 
         copy : bool, optional {True}
             Whether to return a copy of the data or the data in the ODE solver.
+
+        .. note ::
+            Using the solver with ``start``, ``step`` is independent to use
+            with ``run``. Calling ``run`` between ``step`` will not affect
+            ``step``'s result.
         """
         if not self.traj_solvers:
             raise RuntimeError("The `start` method must called first.")
@@ -134,21 +142,21 @@ class MultiTrajSolver:
         """
         Get a :cls:`Solver` for a single trajectory.
         """
-        return self._traj_solver_class(self, options=self.options)
+        return self._traj_solver_class(*self.traj_args, options=self.options)
 
-    def run(self, state0, tlist, ntraj=1, *, args=None, options=None,
-            e_ops=None, timeout=0, target_tol=None, seed=None):
+    def run(self, state, tlist, ntraj=1, *, args=None, options=None,
+            e_ops=(), timeout=0, target_tol=None, seed=None):
         """
         Do the evolution of the Quantum system.
 
-        For a ``state0`` at time ``tlist[0]`` do the evolution as directed by
+        For a ``state`` at time ``tlist[0]`` do the evolution as directed by
         ``rhs`` and for each time in ``tlist`` store the state and/or
         expectation values in a :cls:`Result`. The evolution method and stored
         results are determined by ``options``.
 
         Parameters
         ----------
-        state0 : :class:`Qobj`
+        state : :class:`Qobj`
[…]
