### 2021-12-02 c587f01
Add docstrings for Lorentzian baths.
