### 2021-12-02 1e6ab92
Consistently use Nk for the cut-off on the number of exponentials retained in the series.
