+++ qutip/nonmarkov/bofin_baths.py
@@ -325,15 +325,15 @@ class DrudeLorentzBath(BosonicBath):
     lam : float
         Coupling strength.
 
+    gamma : float
+        Bath spectral density cutoff frequency.
+
     T : float
         Bath temperature.
 
     Nk : int
         Number of exponential terms used to approximate the bath correlation
         functions.
-
-    gamma : float
-        Bath spectral density cutoff frequency.
 
     terminator : bool, default False
         Whether to calculate the Matsubara terminator for the bath. If
@@ -367,13 +367,13 @@ class DrudeLorentzBath(BosonicBath):
         where ``dirac(t)`` denotes the Dirac delta function.
     """
     def __init__(
-        self, Q, lam, T, Nk, gamma, terminator=False, combine=True, tag=None,
+        self, Q, lam, gamma, T, Nk, terminator=False, combine=True, tag=None,
     ):
         ck_real, vk_real, ck_imag, vk_imag = self._matsubara_params(
             lam=lam,
             gamma=gamma,
-            Nk=Nk,
             T=T,
+            Nk=Nk,
         )
 
         super().__init__(
@@ -407,7 +407,7 @@ class DrudeLorentzBath(BosonicBath):
         L_bnd = -delta * op
         return delta, L_bnd
 
-    def _matsubara_params(self, lam, gamma, Nk, T):
+    def _matsubara_params(self, lam, gamma, T, Nk):
         """ Calculate the Matsubara coefficents and frequencies. """
         ck_real = [lam * gamma / np.tan(gamma / (2 * T))]
         ck_real.extend([
@@ -452,15 +452,15 @@ class DrudeLorentzPadeBath(BosonicBath):
     lam : float
         Coupling strength.
 
+    gamma : float
+        Bath spectral density cutoff frequency.
+
     T : float
         Bath temperature.
 
     Nk : int
         Number of Padé exponentials terms used to approximate the bath
         correlation functions.
-
-    gamma : float
-        Bath spectral density cutoff frequency.
 
     terminator : bool, default False
         Whether to calculate the Padé terminator for the bath. If true, the
@@ -494,7 +494,7 @@ class DrudeLorentzPadeBath(BosonicBath):
         where ``dirac(t)`` denotes the Dirac delta function.
     """
     def __init__(
-        self, Q, lam, T, Nk, gamma, terminator=False, combine=True, tag=None
+        self, Q, lam, gamma, T, Nk, terminator=False, combine=True, tag=None
     ):
         eta_p, gamma_p = self._corr(lam=lam, gamma=gamma, T=T, Nk=Nk)
 
@@ -594,18 +594,18 @@ class UnderDampedBath(BosonicBath):
     lam : float
         Coupling strength.
 
+    gamma : float
+        Bath spectral density cutoff frequency.
+
+    w0 : float
+        Bath spectral density resonance frequency.
+
     T : float
         Bath temperature.
 
     Nk : int
         Number of exponential terms used to approximate the bath correlation
         functions.
-
-    gamma : float
-        Bath spectral density cutoff frequency.
-
-    w0 : float
-        Bath spectral density resonance frequency.
 
     combine : bool, default True
         Whether to combine exponents with the same frequency (and coupling
@@ -617,21 +617,21 @@ class UnderDampedBath(BosonicBath):
         bath an exponent is from.
     """
     def __init__(
-        self, Q, lam, T, Nk, gamma, w0, combine=True, tag=None,
[…]
