### 2021-12-02 f8d0a17
Add bath tests for Lorentzian Pade and Matsubara baths.
