### 2021-12-02 628c5f7
Apply suggestions from code review
Co-authored-by: Simon Cross <hodgestar+github@gmail.com>
