### 2021-12-02 bd1ff09
Replace terminator=True/False argument with a terminator method on the baths that support it.
