### 2021-12-02 c95ab97
Update bosonic HEOM guide to new bath parameter ordering and terminator support.
