### 2021-12-02 ce3f774
Add an introduction to the fermionic guide.
