### 2021-12-02 3d4713a
Add Lorentzian baths to the qutip.heom API.
