### 2021-12-02 b4e2389
Use capital gamma consistently in maths blocks and small gamma consistently in code.
