### 2021-12-06 028cb3c
Add the HierarchyADOs and HierarchyADOsState to the heom API since users will need to know how these classes work even though they aren't likely to construct instances themselves.
