### 2021-12-06 134c5b2
Add a section showing how to determine the fermionic bath currents from the ADOs.
