+++ doc/guide/heom/fermionic.rst
@@ -15,14 +15,13 @@ are
 
 We will demonstrate how to describe the bath using two different expansions
 of the spectral density correlation function (Matsubara's expansion and
-a Padé expansion), how to evolve the system in time and how to calculate
+a Padé expansion), how to evolve the system in time, and how to calculate
 the steady state.
 
-Since our fermion is coupled to two reservoirs, we will construct two baths
--- one for each reservoir or lead -- in each case and call them the
-left (:math:`L`) and right (:math:`R`) baths for convenience. Each bath will
-have a different chemical potential :math:`\mu` which we will label
-:math:`\mu_L` and :math:`\mu_R`.
+Since our fermion is coupled to two reservoirs, we will construct two baths --
+one for each reservoir or lead -- and call them the left (:math:`L`) and right
+(:math:`R`) baths for convenience. Each bath will have a different chemical
+potential :math:`\mu` which we will label :math:`\mu_L` and :math:`\mu_R`.
 
 First we will do this using the built-in implementations of
 the bath expansions, :class:`LorentzianBath` and
@@ -33,9 +32,9 @@ use those coefficients to construct your own bath description so that you can
 implement your own fermionic baths.
 
 Our implementation of fermionic baths primarily follows the definitions used by
-Christian Schinabeck in his Dissertation
+Christian Schinabeck in his dissertation (
 https://opus4.kobv.de/opus4-fau/files/10984/DissertationChristianSchinabeck.pdf
-and related publications.
+) and related publications.
 
 A notebook containing a complete example similar to this one implemented in
 BoFiN can be found in `example notebook 4b
@@ -83,7 +82,7 @@ Now let us describe the bath properties:
 where :math:`\Gamma` (``gamma``), :math:`W` and :math:`T` are the parameters of
 an Lorentzian bath, :math:`\mu_L` (``mu_L``) and :math:`\mu_R` (``mu_R``) are
 the chemical potentials of the left and right baths, and ``Q`` is the coupling
-operator between the system and the bath.
+operator between the system and the baths.
 
 We may the pass these parameters to either ``LorentzianBath`` or
 ``LorentzianPadeBath`` to construct an expansion of the bath correlations:
@@ -99,15 +98,19 @@ We may the pass these parameters to either ``LorentzianBath`` or
     Nk = 2
 
     # Matsubara expansion:
-    bath_L = LorentzianBath(Q, gamma, W, mu_L, T, Nk)
-    bath_R = LorentzianBath(Q, gamma, W, mu_R, T, Nk)
+    bath_L = LorentzianBath(Q, gamma, W, mu_L, T, Nk, tag="L")
+    bath_R = LorentzianBath(Q, gamma, W, mu_R, T, Nk, tag="R")
 
     # Padé expansion:
-    bath_L = LorentzianPadeBath(Q, gamma, W, mu_L, T, Nk)
-    bath_R = LorentzianPadeBath(Q, gamma, W, mu_R, T, Nk)
+    bath_L = LorentzianPadeBath(Q, gamma, W, mu_L, T, Nk, tag="L")
+    bath_R = LorentzianPadeBath(Q, gamma, W, mu_R, T, Nk, tag="R")
 
 Where ``Nk`` is the number of terms to retain within the expansion of the
 bath.
+
+Note that we haved labelled each bath with a tag (either "L" or "R") so that
+we can identify the exponents from individual baths later when calculating
+the currents between the system and the bath.
 
 
 System and bath dynamics
@@ -137,6 +140,177 @@ and to calculate the system evolution as a function of time:
 
[…]
