### 2021-12-06 342e334
increase binary search limit
