### 2021-12-06 dee01f9
Add remaining sections to the fermionic bath HEOM guide.
