+++ doc/guide/heom/fermionic.rst
@@ -185,12 +185,12 @@ The current for each exponent is given by:
 
 .. math::
 
-    Current &= \pm i Tr(Q^\pm \cdot A)
+    \mathrm{Current} = \pm i \mathrm{Tr}(Q^\pm \cdot A)
 
-where the :math:`pm` sign is the sign of the exponent (see the
-description later in :ref:`Calculating the bath expansion coefficients`) and
-:math:`Q^\pm` is :math:`Q` for :math:`+` exponents and :math:`Q.dag()` for
-:math:`-` exponents.
+where the :math:`\pm` sign is the sign of the exponent (see the
+description later in :ref:`Padé expansion coefficients`) and
+:math:`Q^\pm` is :math:`Q` for ``+`` exponents and :math:`Q^{\dagger}` for
+``-`` exponents.
 
 The first-level exponents for the left bath are retrieved by calling
 ``.filter(tags=["L"])`` on ``ado_state`` which is an instance of
@@ -313,60 +313,70 @@ As you can see, there is still some way to go beyond ``t = 100`` before the
 steady state is reached!
 
 
-Calculating the bath expansion coefficients
--------------------------------------------
+Padé expansion coefficients
+---------------------------
 
-We choose a Lorentzian spectral density for the leads, with a peak at the
-chemical potential. The latter simplifies a little the notation required for the
-correlation functions, but can be relaxed if neccessary.
+We now look at how to calculate the correlation expansion coefficients
+for the Lorentzian spectral density ourselves. Once we have calculated
+the coefficients we can construct a :class:`FermionicBath` directly from them.
+A similar procedure can be used to apply :class:`HEOMSolver` to any fermionic
+bath for which we can calculate the expansion coefficients.
+
+In the fermionic case we must descriminate between the order in which
+excitations are created within the bath, so we define two different correlation
+functions, :math:`C_{+}(t)`, and :math:`C_{-}(t)`:
 
 .. math::
 
-    J(\omega) = \frac{\Gamma W^2}{((\omega - \mu_K)^2 + W^2)}
+    C^{\sigma}(t) = \frac{1}{2\pi} \int_{-\infty}^{\infty} d\omega e^{\sigma i \omega t} J(\omega) f_F[\sigma\beta(\omega - \mu)]
 
-Fermi distribution is:
+where :math:`\sigma` is either ``+`` or ``-`` and, :math:`f_F` is the Fermi
+distribution function, and :math:`J(\omega)` is the Lorentzian spectral density
+we defined at the start.
+
+The Fermi distribution function is:
 
 .. math::
 
     f_F (x) = (\exp(x) + 1)^{-1}
 
-gives correlation functions:
+As in the bosonic case we can approximate this integral with a Matsubara or
+Padé expansion. For the Lorentzian bath the Padé expansion converges much
+more quickly, so we will calculate the Padé expansion coefficients here.
+
+The Padé decomposition approximates the Fermi distubition as:
 
 .. math::
 
-    C^{\sigma}_K(t) = \frac{1}{2\pi} \int_{-\infty}^{\infty} d\omega e^{\sigma i \omega t} \Gamma_K(\omega) f_F[\sigma\beta(\omega - \mu)]
[…]
