### 2021-12-06 a9f69e4
Add a page on the history of the HEOM implementation.
