### 2021-12-06 7b4e242
Add Lorentzian baths to class documentation.
