+++ doc/apidoc/classes.rst
@@ -83,6 +83,12 @@ Non-Markovian Solvers
 .. autoclass:: qutip.nonmarkov.heom.FermionicBath
     :members:
 
+.. autoclass:: qutip.nonmarkov.heom.LorentzianBath
+    :members:
+
+.. autoclass:: qutip.nonmarkov.heom.LorentzianPadeBath
+    :members:
+
 .. autoclass:: qutip.nonmarkov.heom.HierarchyADOs
     :members:
 
