### 2021-12-06 cd831bc
Update docstring of process_fidelity
