### 2021-12-07 9ac5243
Fix typos and apply simpler phrasing suggestions from review.
