### 2021-12-07 0617650
Add a link to the full BoFiN FMO example notebook.
