### 2021-12-07 8557f8d
Add support for passing HierarchyADOsState as rho0 to .run(rho0, ..., ado_init=True).
