### 2021-12-07 68b196e
Add test for passing a raw numpy array as the ADO state.
