### 2021-12-07 ec33515
Remove unused self._ext_axes.
