### 2021-12-07 47c8c86
better handle scipy.ode
