### 2021-12-08 ed9a6e8
Add test for concurrent integrator usage
