### 2021-12-08 a3fecaa
Add special handling for the inline notebook backend.
