### 2021-12-09 5b955e8
Remove upper bound on numpy version to allow building from source against recent versions of numpy.
