+++ setup.cfg
@@ -31,7 +31,7 @@ install_requires =
     scipy>=1.0
     packaging
 setup_requires =
-    numpy>=1.16.6,<1.20
+    numpy>=1.16.6
     scipy>=1.0
     cython>=0.29.20
     packaging
