### 2021-12-09 e014e65
Update cibuildwheel to support cibuildwheel 2.3.0.
