### 2021-12-09 58648db
Explicitly specify the cibuildwheel frontend since it might change and note that we should switch to the manylinux_2 tags eventually.
