### 2021-12-09 310de7d
Change our default test Python version to 3.10.
