### 2021-12-09 cf28ae0
Remove numpy version upperbound from pyproject.toml.
