+++ pyproject.toml
@@ -4,7 +4,7 @@ requires = [
     "packaging",
     "wheel",
     "cython>=0.29.20",
-    "numpy>=1.16.6,<1.20",
+    "numpy>=1.16.6",
     "scipy>=1.0",
 ]
 build-backend = "setuptools.build_meta"
