### 2021-12-09 f0deeef
Build documentation on Python 3.10.
