+++ .github/workflows/build_documentation.yml
@@ -14,7 +14,7 @@ jobs:
       - uses: actions/setup-python@v2
         name: Install Python
         with:
-          python-version: '3.8'
+          python-version: "3.10"
 
       - name: Install documentation dependencies
         run: |
