### 2021-12-09 db53f73
Switch default Python version back to 3.9 -- miniconda doesn't appear to support scipy on 3.10 yet.
