### 2021-12-09 99f949b
Fix comment about cibuildwheel version to say 'minor' new release.
