### 2021-12-09 ead740b
Check whether specifying the documentation requirements as a constraint builds qutip with the installed version of numpy.
