+++ .github/workflows/build_documentation.yml
@@ -22,7 +22,7 @@ jobs:
 
       - name: Install QuTiP from GitHub
         run: |
-          python -mpip install -e .[full]
+          python -mpip install -e .[full] -c doc/requirements.txt
           # Install in editable mode so it doesn't matter if we import from
           # inside the installation directory, otherwise we can get some errors
           # because we're importing from the wrong location.
