### 2021-12-09 c059c3f
Add a custom pyproject.toml for the documentation build.
