### 2021-12-09 9aa6d24
Add a custom pyproject.toml for the wheel builds.
