+++ .github/workflows/tests.yml
@@ -91,7 +91,7 @@ jobs:
           else
             conda install nomkl "numpy${{ matrix.numpy-requirement }}" "scipy${{ matrix.scipy-requirement }}"
           fi
-          python -m pip install -e .[$QUTIP_TARGET]
+          python -m pip install -e .[$QUTIP_TARGET] --no-build-isolation
           python -m pip install pytest-cov coveralls
 
       - name: Package information
