+++ setup.cfg
@@ -30,11 +30,6 @@ install_requires =
     numpy>=1.16.6
     scipy>=1.0
     packaging
-setup_requires =
-    numpy>=1.16.6,<1.20
-    scipy>=1.0
-    cython>=0.29.20
-    packaging
 
 [options.packages.find]
 include = qutip*
