+++ pyproject.toml
@@ -10,7 +10,7 @@ requires = [
 build-backend = "setuptools.build_meta"
 
 [tool.cibuildwheel]
-manylinux-x86_64-image: "manylinux2014"
+manylinux-x86_64-image = "manylinux2014"
 manylinux-i686-image = "manylinux2014"
 # Change in future version to "build"
 build-frontend = "pip"
