+++ pyproject.toml
@@ -18,5 +18,5 @@ build-frontend = "pip"
 [tool.cibuildwheel.override]
 # NumPy support manylinux2010 on CPython 3.6-3.9
 select = "cp3?-*"
-manylinux-x86_64-image: "manylinux2010"
+manylinux-x86_64-image = "manylinux2010"
 manylinux-i686-image = "manylinux2010"
