### 2021-12-09 e8e40dc
Remove documentation pyproject.toml file (replaced with non-isolated build of qutip when building documentation).
