+++ doc/doc-pyproject.toml
@@ -1,13 +1,0 @@
-# A custom version of pyproject.toml used to ensure that
-# the versions of numpy, scipy and Cython installed by doc/requirements.txt
-# are used to build QuTiP itself during the documentation build.
-[build-system]
-requires = [
-    "setuptools",
-    "packaging",
-    "wheel",
-    "Cython==0.29.23",
-    "numpy==1.19.5",
-    "scipy==1.6.2",
-]
-build-backend = "setuptools.build_meta"
