### 2021-12-09 d3a5c6e
add comment about lsoda step forcing
