### 2021-12-09 0437a09
Simply build and PyPI pyproject.toml using ideas from @henryiii and PR #1738.
