### 2021-12-09 351a561
fix: restore setup_requires
