+++ setup.cfg
@@ -30,6 +30,11 @@ install_requires =
     numpy>=1.16.6
     scipy>=1.0
     packaging
+setup_requires =
+     numpy
+     scipy>=1.0
+     cython>=0.29.20
+     packaging
 
 [options.packages.find]
 include = qutip*
