### 2021-12-09 78c566e
Make cibuildwheel version requirement any 2.3.* version (and not any version before 2.4).
