### 2021-12-09 c898b89
Add a build case for Python 3.10.
