### 2021-12-09 9a03c0f
Add a note on the double-show case when creating ones own figure and how to avoid it.
