### 2021-12-09 c138c52
Loosen build requirement in setup_requires.
