### 2021-12-09 c76a766
Fix grammer if .show() docstring.
