+++ qutip/bloch.py
@@ -644,7 +644,7 @@ class Bloch:
 
         When using inline plotting in Jupyter notebooks, any figure created
         in a notebook cell is displayed after the cell executes. Thus if you
-        create a figure yourself and pass it to with
+        create a figure yourself and use it create a Bloch sphere with
         ``b = Bloch(..., fig=fig)`` and then call ``b.show()`` in the same
         cell, then the figure will be displayed twice. If you do create your
         own figure, the simplest solution to this is to not call ``.show()``
