### 2021-12-09 288e69f
Merge pull request #1619 from hodgestar/feature/clean-up-bloch-sphere-rendering
More carefully track the state of the Bloch sphere figure and axes
