+++ qutip/bloch.py
@@ -73,12 +73,12 @@ try:
 
             self.set_positions((xs[0], ys[0]), (xs[1], ys[1]))
             FancyArrowPatch.draw(self, renderer)
-except:
+except ImportError:
     pass
 
 try:
     from IPython.display import display
-except:
+except ImportError:
     pass
 
 
@@ -145,6 +145,7 @@ class Bloch:
                  background=False):
         # Figure and axes
         self.fig = fig
+        self._ext_fig = fig is not None
         self.axes = axes
         # Background axes, default = False
         self.background = background
@@ -210,14 +211,6 @@ class Bloch:
         self.savenum = 0
         # Style of points, 'm' for multiple colors, 's' for single color
         self.point_style = []
-
-        # status of rendering
-        self._rendered = False
-        # status of showing
-        if fig is None:
-            self._shown = False
-        else:
-            self._shown = True
 
     def set_label_convention(self, convention):
         """Set x, y and z labels according to one of conventions.
@@ -438,7 +431,7 @@ class Bloch:
         """
         Plots Bloch sphere and data sets.
         """
-        self.render(self.fig, self.axes)
+        self.render()
 
     def run_from_ipython(self):
         try:
@@ -447,30 +440,50 @@ class Bloch:
         except NameError:
             return False
 
-    def render(self, fig=None, axes=None):
+    def _is_inline_backend(self):
+        backend = matplotlib.get_backend()
+        return backend == "module://matplotlib_inline.backend_inline"
+
+    def render(self):
         """
         Render the Bloch sphere and its data sets in on given figure and axes.
         """
-        if self._rendered:
-            self.axes.clear()
+        if not self._ext_fig and not self._is_inline_backend():
+            # If no external figure was supplied, we check to see if the
+            # figure we created in a previous call to .render() has been
+            # closed, and re-create if has been. This has the unfortunate
+            # side effect of losing any modifications made to the axes or
+            # figure, but the alternative is to crash the matplotlib backend.
+            #
+            # The inline backend used by, e.g. jupyter notebooks, is happy to
+            # use closed figures so we leave those figures intact.
+            if (
+                self.fig is not None and
+                not plt.fignum_exists(self.fig.number)
+            ):
+                self.fig = None
+                self.axes = None
 
-        self._rendered = True
-
-        # Figure instance for Bloch sphere plot
-        if not fig:
+        if self.fig is None:
             self.fig = plt.figure(figsize=self.figsize)
+            if self._is_inline_backend():
+                # We immediately close the inline figure do avoid displaying
+                # the figure twice when .show() calls display.
+                plt.close(self.fig)
 
-        if not axes:
-            self.axes = _axes3D(self.fig, azim=self.view[0],
-                                elev=self.view[1])
+        if self.axes is None:
+            self.axes = _axes3D(self.fig, azim=self.view[0], elev=self.view[1])
 
+        # Clearing the axes is horrifically slow and loses a lot of the
+        # axes state, but matplotlib doesn't seem to provide a better way
+        # to redraw Axes3D. :/
+        self.axes.clear()
+        self.axes.grid(False)
         if self.background:
-            self.axes.clear()
             self.axes.set_xlim3d(-1.3, 1.3)
             self.axes.set_ylim3d(-1.3, 1.3)
             self.axes.set_zlim3d(-1.3, 1.3)
         else:
-            self.plot_axes()
             self.axes.set_axis_off()
             self.axes.set_xlim3d(-0.7, 0.7)
             self.axes.set_ylim3d(-0.7, 0.7)
@@ -479,14 +492,17 @@ class Bloch:
         # Matplotlib did this stretching for < 3.3.0, but not above.
[…]
