### 2021-12-09 8c9c3cf
Update qutip/visualization.py
Co-authored-by: Simon Cross <hodgestar+github@gmail.com>
