### 2021-12-09 d0b53c8
Update qutip/visualization.py
