### 2021-12-09 989b7fe
Update qutip/visualization.py
Co-authored-by: Simon Cross <hodgestar+github@gmail.com>
