### 2021-12-09 ea70ed9
Update qutip/visualization.py
