+++ qutip/visualization.py
@@ -967,7 +967,8 @@ def plot_fock_distribution(rho, offset=0, fig=None, ax=None,
     ax.set_xlabel('Fock number', fontsize=12)
     ax.set_ylabel('Occupation probability', fontsize=12)
 
-    ax.set_title(title)
+    if title:
+        ax.set_title(title)
 
     return fig, ax
 
