### 2021-12-09 4ef1223
Update qutip/visualization.py
