### 2021-12-09 a1e71f1
Try use conda-forge for Python 3.10 scipy.
