### 2021-12-09 10fc340
Apply small error message and method passing fixes.
