### 2021-12-10 9ec8be9
PR krylovsolve merge qutip
