+++ qutip/krylovsolve.py
@@ -1,0 +1,603 @@
+"""
+This module provides approximations of the time evolution operator
+using small dimensional Krylov subspaces.
+"""
+
+from functools import reduce
+from math import ceil
+import operator
+
+import numpy as np
+from qutip import mcsolve
+from qutip.expect import expect
+from qutip.qobj import Qobj
+from qutip.solver import Result
+from qutip.ui.progressbar import BaseProgressBar, TextProgressBar
+from scipy.linalg import eigh, expm
+
+from qutip import mcsolve
+
+
+def krylovsolve(
+    H: Qobj,
+    psi0: Qobj,
+    tlist: np.array,
+    krylov_dim: int = 30,
+    e_ops=None,
+    tolerance: float = 1e-7,
+    store_states: bool = False,
+    store_final_state: bool = False,
+    progress_bar: bool = None,
+    sparse: bool = False,
+):
+    """
+     Time evolution of state vectors for time independent Hamiltonians.
+     Evolve the state vector ("psi0") finding an approximation for the time
+     evolution operator of Hamiltonian ("H") by obtaining the projection of
+     the time evolution operator on a set of small dimensional Krylov
+     subspaces (m<<dim(H)).
+     The output is either the state vector or the expectation values of
+     supplied operators ("e_ops") at arbitrary points in a time range built
+     from inputs "t0", "tf" and "dt". Optionally, a custom ("tlist") without an
+     even time stepping between times can be provided, but the algorithm
+     will become slower.
+     **Additional options**
+     Additional options to krylovsolve can be set with the following:
+     "store_states": stores states even though expectation values are
+     requested via the "e_ops" argument.
+     "store_final_state": store final state even though expectation values are
+     requested via the "e_ops" argument.
+     "krylov_algorithm": default behavior uses lanczos algorithm to
+     calculate the different Krylov subspaces, and it is only valid for
+     self-adjoint operators. If by any chance you decide to use this evolution
+     on a non self-adjoint Hamiltonian, Arnoldi iteration (slower than
+     lanczos but does not require self-adjoint) can be enabled.
+     Another alternative is to use Krylov subspaces obtained from Taylor
+     expansion of the Hamiltonian.
+    Parameters
+    -------------
+     H : :class:`qutip.Qobj`
+        System Hamiltonian.
+     psi0 : :class: `qutip.Qobj`
+         initial state vector (ket).
+     tlist : None / *list* / *array*
+        List of times on which to evolve the initial state.
+     krylov_dim: int
+         Dimension of Krylov approximation subspaces used for the time
+         evolution approximation.
+     e_ops : None / list of :class:`qutip.Qobj`
+         Single operator or list of operators for which to evaluate
+         expectation values.
+     if store_states : bool (default False)
+         If e_ops is provided, store each state vector corresponding to each
+         time in tlist.
+     store_final_state : bool (default False)
+         If e_ops is provided, store the final state vector of the evolution.
+     progress_bar : None / BaseProgressBar
+         Optional instance of BaseProgressBar, or a subclass thereof, for
+         showing the progress of the simulation.
+     sparse : bool (default False)
+         Use np.array to represent system Hamiltonians. If True, scipy sparse
[…]
