### 2021-12-10 49baaac
PR krylovsolve merge qutip
