### 2021-12-10 b6c6add
PR krylovsolve merge qutip
