### 2021-12-10 dcaa63a
Feat: Reduce cognitive complexity
