+++ qutip/krylovsolve.py
@@ -31,57 +31,57 @@ def krylovsolve(
     sparse: bool = False,
 ):
     """
-     Time evolution of state vectors for time independent Hamiltonians.
-     Evolve the state vector ("psi0") finding an approximation for the time
-     evolution operator of Hamiltonian ("H") by obtaining the projection of
-     the time evolution operator on a set of small dimensional Krylov
-     subspaces (m<<dim(H)).
-     The output is either the state vector or the expectation values of
-     supplied operators ("e_ops") at arbitrary points at ("tlist").
+    Time evolution of state vectors for time independent Hamiltonians.
+    Evolve the state vector ("psi0") finding an approximation for the time
+    evolution operator of Hamiltonian ("H") by obtaining the projection of
+    the time evolution operator on a set of small dimensional Krylov
+    subspaces (m<<dim(H)).
+    The output is either the state vector or the expectation values of
+    supplied operators ("e_ops") at arbitrary points at ("tlist").
 
-     **Additional options**
-     Additional options to krylovsolve can be set with the following:
-     "store_states": stores states even though expectation values are
-     requested via the "e_ops" argument.
-     "store_final_state": store final state even though expectation values are
-     requested via the "e_ops" argument.
+    **Additional options**
+    Additional options to krylovsolve can be set with the following:
+    "store_states": stores states even though expectation values are
+    requested via the "e_ops" argument.
+    "store_final_state": store final state even though expectation values are
+    requested via the "e_ops" argument.
 
     Parameters
     -------------
-     H : :class:`qutip.Qobj`
-        System Hamiltonian.
-     psi0 : :class: `qutip.Qobj`
-         initial state vector (ket).
-     tlist : None / *list* / *array*
-        List of times on which to evolve the initial state. If None, nothing
-        happens but the code won't break.
-     krylov_dim: int
-         Dimension of Krylov approximation subspaces used for the time
-         evolution approximation.
-     e_ops : None / list of :class:`qutip.Qobj`
-         Single operator or list of operators for which to evaluate
-         expectation values.
-     if store_states : bool (default False)
-         If e_ops is provided, store each state vector corresponding to each
-         time in tlist.
-     store_final_state : bool (default False)
-         If e_ops is provided, store the final state vector of the evolution.
-     progress_bar : None / BaseProgressBar
-         Optional instance of BaseProgressBar, or a subclass thereof, for
-         showing the progress of the simulation.
-     sparse : bool (default False)
-         Use np.array to represent system Hamiltonians. If True, scipy sparse
-         arrays are used instead.
-     tolerance : :float: (default 1e-7)
-         Minimum bound value for the final state infidelity with respect to
-         the exact state.
-     Returns
-     ---------
-      result: :class:`qutip.Result`
-         An instance of the class :class:`qutip.Result`, which contains
-         either an *array* `result.expect` of expectation values for the times
-         `tlist`, or an *array* `result.states` of state vectors corresponding
-         to the times `tlist` [if `e_ops` is an empty list].
+    H : :class:`qutip.Qobj`
+            System Hamiltonian.
+    psi0 : :class: `qutip.Qobj`
+            initial state vector (ket).
[…]
