### 2021-12-12 6cc84d8
Default value and doc for hinton
Change the default value back to scaled to be consistent with the old behavior.
Mention in the docstring that the sign is chosen according to the real part of the matrix.
