### 2021-12-12 00109a0
Updated add_arc and add_line to include Qobj input, matplotlib format, and steps. Removed dependency of add_arc on add_points.
