### 2021-12-13 3c729cb
Add example for smesolve, add comment on dW, remove non-related image
