+++ doc/guide/dynamics/dynamics-stochastic.rst
@@ -29,12 +29,12 @@ Closed system
 In closed systems, the resulting stochastic differential equation is
 
 .. math::
-	:label: jump_ssesolve
+    :label: jump_ssesolve
 
-	\delta \psi(t) = - i H \psi(t) \delta t
-	                 - \sum_n \left( \frac{C_n^\dagger C_n}{2} -\frac{e_n}{2} C_n
-					 + \frac{e_n^2}{8} \right) \psi(t) \delta t
-	                 + \sum_n \left( C_n - \frac{e_n}{2} \right) \psi(t) \delta \omega
+    d \psi(t) = - i H \psi(t) dt
+                     - \sum_n \left( \frac{C_n^\dagger C_n}{2} -\frac{e_n}{2} C_n
+                     + \frac{e_n^2}{8} \right) \psi(t) dt
+                     + \sum_n \left( C_n - \frac{e_n}{2} \right) \psi(t) dW
 
 with
 
@@ -43,30 +43,35 @@ with
 
    e_n = \left<\psi(t)|C_n + C_n^\dagger|\psi(t)\right>
 
-Here :math:`\delta \omega` is a Wiener increment.
+Here :math:`dW` is a Wiener increment with :math:`E[dW] = dt`.
 
-In QuTiP, this is available with the function :func:`qutip.ssesolve`.
+In QuTiP, you can solve this equation using :func:`qutip.ssesolve` passing the argument `method="homodyne"`. See the example below.
 
 .. plot::
     :context:
 
+    import qutip
+
+    kappa = 0.1
+    g = 0.25
     times = np.linspace(0.0, 10.0, 201)
     psi0 = tensor(fock(2, 0), fock(10, 5))
     a  = tensor(qeye(2), destroy(10))
     sm = tensor(destroy(2), qeye(10))
 
-    H = 2*np.pi*a.dag()*a + 2*np.pi*sm.dag()*sm + 2*np.pi*0.25*(sm*a.dag() + sm.dag()*a)
-    data = ssesolve(H, psi0, times, sc_ops=[np.sqrt(0.1) * a],
+    H = (2*np.pi*a.dag()*a + 2*np.pi*sm.dag()*sm
+         + 2*np.pi*g*(sm*a.dag() + sm.dag()*a))
+    data = ssesolve(H, psi0, times, sc_ops=[np.sqrt(kappa) * a],
                     e_ops=[a.dag()*a, sm.dag()*sm], method="homodyne")
 
-    plt.figure()
-    plt.plot(times, data.expect[0], times, data.expect[1])
-    plt.title('Homodyne time evolution')
-    plt.xlabel('Time')
-    plt.ylabel('Expectation values')
-    plt.legend(("cavity photon number", "atom excitation probability"))
-    plt.show()
+    fig, ax = plt.subplots()
+    ax.plot(times, data.expect[0], times, data.expect[1])
+    ax.set_title('Homodyne Detection - Closed System')
+    ax.set_xlabel('Time')
+    ax.set_ylabel('Expectation values')
+    ax.legend(("cavity photon number", "atom excitation probability"))
 
+For more examples, see this `notebook <https://nbviewer.ipython.org/github/qutip/qutip-notebooks/blob/master/development/development-ssesolve-tests.ipynb>`_.
 
 Open system
 --------------
@@ -74,42 +79,92 @@ Open system
 
 In open systems, 2 types of collapse operators are considered, :math:`S_i`
 represent the dissipation in the environment, :math:`C_i` are monitored operators.
-The deterministic part of the evolution is the liouvillian with both types of
-collapses
[…]
