+++ qutip/krylovsolve.py
@@ -15,8 +15,6 @@ from qutip.solver import Result
 from qutip.ui.progressbar import BaseProgressBar, TextProgressBar
 from scipy.linalg import eigh, expm
 
-from qutip import mcsolve
-
 
 def krylovsolve(
     H: Qobj,
@@ -100,18 +98,16 @@ def krylovsolve(
 
     # verify that the hamiltonian meets the requirements
     assert isinstance(H, Qobj) | isinstance(
-        H, np.ndarray
-    ), "the Hamiltonian must be either a Qobj or a np.ndarray."
+        H,
+        np.ndarray), "the Hamiltonian must be either a Qobj or a np.ndarray."
 
     assert len(H.shape) == 2, "the Hamiltonian must be 2-dimensional."
 
-    assert (
-        H.shape[0] == H.shape[1]
-    ), "the Hamiltonian must be a square 2-dimensional."
+    assert (H.shape[0] == H.shape[1]
+            ), "the Hamiltonian must be a square 2-dimensional."
 
-    assert (
-        H.shape[0] >= krylov_dim
-    ), "the Hamiltonian dimension must be greater or equal to the \
+    assert (H.shape[0] >= krylov_dim
+            ), "the Hamiltonian dimension must be greater or equal to the \
             maximum allowed krylov dimension."
 
     if isinstance(H, Qobj):
@@ -124,12 +120,11 @@ def krylovsolve(
 
     # verify that the state vector meets the requirements
     assert isinstance(psi0, Qobj) | isinstance(
-        psi0, np.ndarray
-    ), "The state vector must be either a Qobj or a np.ndarray."
+        psi0,
+        np.ndarray), "The state vector must be either a Qobj or a np.ndarray."
 
-    assert (
-        psi0.shape[0] == _H.shape[0]
-    ), "The state vector and the Hamiltonian must share the same \
+    assert (psi0.shape[0] == _H.shape[0]
+            ), "The state vector and the Hamiltonian must share the same \
         dimension."
 
     # transform state type from Qobj to np.ndarray for faster operations
@@ -140,161 +135,157 @@ def krylovsolve(
         _psi = psi0.copy()
         _psi = _psi / np.linalg.norm(_psi)
 
-    krylov_results = _krylovsolve(
-        H,
-        _H,
-        psi0,
-        _psi,
-        tlist,
-        krylov_dim,
-        e_ops,
-        tolerance,
-        store_states,
-        store_final_state,
-        progress_bar,
-        pbar,
-        sparse,
-    )
+    k_solve = KSolve(tlist, krylov_dim, sparse, progress_bar, pbar, tolerance)
 
+    krylov_results = k_solve.solve(_H, _psi, e_ops, store_states,
+                                   store_final_state)
     return krylov_results
 
 
-def _krylovsolve(
-    H,
-    _H,
-    psi0,
-    _psi,
-    tlist,
-    krylov_dim,
-    e_ops,
-    tolerance,
-    store_states,
-    store_final_state,
-    progress_bar,
-    pbar,
-    sparse,
-):
-    # optimization step
-    dim_m = krylov_dim
+class KSolve:
+    def __init__(self, tlist, krylov_dim, sparse, progress_bar, pbar,
+                 tolerance):
+        self.tlist = tlist
+        self.krylov_dim = krylov_dim
+        self.sparse = sparse
+        self.tolerance = tolerance
+        self.progress_bar = progress_bar
+        self.pbar = pbar
 
-    tf = tlist[-1]
-    t0 = tlist[0]
+        self.t0, self.tf = self.tlist[0], self.tlist[-1]
+        self.n_timesteps = None
+        self.partitions = None
[…]
