### 2021-12-14 e732239
Raise a ValueError if method 'operator' is used with offset != 0 in coherent and coherent_dm.
