### 2021-12-16 #46 @lwasser
hi there @hodgestar i'm really sorry for the super slow response here. For the time being this package is out of scope for us. You might look at another journal such as JOSS as a potential landing spot for it. Given our current reviewer pool and scope we will not be able to review. I am going to close this issue for the time being but please feel free to keep in touch with us and to ask any additional questions.
